Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2579.779
#> 
#> $assign
#>  [1] 1 2 2 2 1 1 2 2 2 1 2 1 1 1 2 1 2 2 2 1 2 2 2 2 2 1 2 1 1 1 2 2 2 1 2 1 2 1
#> [39] 2 1 1 2 2 1 2 1 1 2 1 1
#> 
#> $a
#> $a[[1]]
#>              [,1]          [,2]          [,3] [,4] [,5]        [,6] [,7]
#>  [1,]  0.85410142 -1.504506e-16 -1.325406e-17    0    0 -0.34576986    0
#>  [2,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#>  [3,] -0.11421420 -3.304202e-17  1.321592e-17    0    0 -0.06751064    0
#>  [4,] -0.08367953  3.367310e-17  4.169145e-19    0    0  0.08300883    0
#>  [5,] -0.79752336 -1.720687e-16 -4.340021e-17    0    0 -0.67809007    0
#>  [6,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#>  [7,]  0.21468896  2.160021e-18  2.730873e-17    0    0  0.10456881    0
#>  [8,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#>  [9,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [10,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [11,]  0.52985866 -3.044844e-17 -3.454908e-19    0    0 -0.02214086    0
#> [12,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [13,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [14,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [15,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [16,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [17,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [18,] -0.02485686 -4.375286e-18 -1.206763e-19    0    0 -0.01512144    0
#> [19,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [20,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [21,] -1.16306389 -9.904387e-18  3.316580e-17    0    0 -0.07271143    0
#> [22,]  0.42230267  4.173820e-17  1.046166e-18    0    0  0.16550898    0
#> [23,]  0.38727515  2.050890e-16 -3.074412e-17    0    0  0.52006143    0
#> [24,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [25,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [26,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [27,]  0.08060675 -8.818529e-18  1.639833e-18    0    0 -0.01019018    0
#> [28,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [29,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [30,]  0.00000000  0.000000e+00  0.000000e+00    0    0  0.00000000    0
#> [31,] -0.88108192 -7.719980e-16  5.881889e-16    0    0 -0.61515345    0
#>               [,8] [,9] [,10]        [,11]        [,12] [,13]       [,14]
#>  [1,] -0.264657716    0     0  1.197891522 -0.316010826     0 -1.09669772
#>  [2,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#>  [3,] -0.063071908    0     0 -0.063298017 -0.054275555     0  0.03998237
#>  [4,]  0.069874410    0     0 -0.164850911  0.077202330     0  0.15071564
#>  [5,] -0.673832949    0     0 -0.048689394 -0.729498802     0  0.19355322
#>  [6,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#>  [7,]  0.116611118    0     0  0.068716080  0.138915591     0 -0.13060341
#>  [8,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#>  [9,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [10,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [11,]  0.011814812    0     0  0.541477095 -0.002320310     0 -0.51878745
#> [12,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [13,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [14,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [15,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [16,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [17,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [18,] -0.015156850    0     0 -0.009189686 -0.015578109     0  0.01040138
#> [19,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [20,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [21,] -0.123905868    0     0 -1.113285315 -0.075251047     0  1.01102720
#> [22,]  0.174839181    0     0  0.248135891  0.175686192     0 -0.25537089
#> [23,]  0.480991275    0     0 -0.093565025  0.479642715     0  0.10748536
#> [24,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [25,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [26,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [27,] -0.003757338    0     0  0.086788211 -0.005016627     0 -0.08601572
#> [28,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [29,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [30,]  0.000000000    0     0  0.000000000  0.000000000     0  0.00000000
#> [31,] -0.402149959    0     0 -1.069840537  0.034727029     0 -0.10969630
#>             [,15]        [,16] [,17]        [,18]        [,19] [,20] [,21]
#>  [1,]  1.11598447 -1.498932368     0 -0.299404944  0.322749081     0     0
#>  [2,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#>  [3,] -0.04719053 -0.024668531     0 -0.061906877  0.070686785     0     0
#>  [4,] -0.15311999  0.242976184     0  0.074786856 -0.081013873     0     0
#>  [5,] -0.15542659 -0.538337852     0 -0.672190477  0.708855679     0     0
#>  [6,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#>  [7,]  0.11234955 -0.011522396     0  0.113790835 -0.113202009     0     0
#>  [8,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#>  [9,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [10,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [11,]  0.52172369 -0.556670341     0 -0.003824599  0.006230744     0     0
#> [12,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [13,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [14,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [15,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [16,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [17,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [18,] -0.01010556 -0.005382989     0 -0.014919663  0.015976439     0     0
#> [19,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [20,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [21,] -1.03491632  0.976291446     0 -0.093268149  0.106653210     0     0
#> [22,]  0.25293946 -0.086620941     0  0.168001310 -0.179408184     0     0
#> [23,] -0.09803790  0.647936287     0  0.485998286 -0.533411475     0     0
#> [24,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [25,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [26,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [27,]  0.08556194 -0.098462541     0 -0.006344271  0.007721580     0     0
#> [28,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [29,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [30,]  0.00000000  0.000000000     0  0.000000000  0.000000000     0     0
#> [31,] -0.24488895 -0.507891072     0 -0.395543028  0.614112819     0     0
#>             [,22] [,23] [,24]        [,25] [,26]       [,27] [,28] [,29] [,30]
#>  [1,] -0.79762730     0     0 -1.073886239     0  1.14345251     0     0     0
#>  [2,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#>  [3,]  0.10345106     0     0  0.042490622     0 -0.05659897     0     0     0
#>  [4,]  0.07579726     0     0  0.148138682     0 -0.15505835     0     0     0
#>  [5,]  0.86048862     0     0  0.141837295     0 -0.16510051     0     0     0
#>  [6,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#>  [7,] -0.24099895     0     0 -0.108926388     0  0.10771817     0     0     0
#>  [8,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#>  [9,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [10,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [11,] -0.51444404     0     0 -0.500231566     0  0.53690089     0     0     0
#> [12,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [13,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [14,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [15,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [16,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [17,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [18,]  0.02529913     0     0  0.009426966     0 -0.01088390     0     0     0
#> [19,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [20,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [21,]  1.10721694     0     0  0.986467044     0 -1.08292630     0     0     0
#> [22,] -0.42301406     0     0 -0.239520465     0  0.26588536     0     0     0
#> [23,] -0.38244988     0     0  0.106734603     0 -0.06986719     0     0     0
#> [24,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [25,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [26,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [27,] -0.07937990     0     0 -0.082349392     0  0.08709966     0     0     0
#> [28,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [29,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [30,]  0.00000000     0     0  0.000000000     0  0.00000000     0     0     0
#> [31,]  0.35678989     0     0  0.155965112     0 -0.52466340     0     0     0
#>             [,31]        [,32]       [,33] [,34] [,35]
#>  [1,] -0.83460086  0.270057877 -1.10848148     0     0
#>  [2,]  0.00000000  0.000000000  0.00000000     0     0
#>  [3,]  0.11263054  0.078566088  0.04511161     0     0
#>  [4,]  0.07965623 -0.073314106  0.15224667     0     0
#>  [5,]  0.85268552  0.653758650  0.16218153     0     0
#>  [6,]  0.00000000  0.000000000  0.00000000     0     0
#>  [7,] -0.23203701 -0.092247679 -0.11606022     0     0
#>  [8,]  0.00000000  0.000000000  0.00000000     0     0
#>  [9,]  0.00000000  0.000000000  0.00000000     0     0
#> [10,]  0.00000000  0.000000000  0.00000000     0     0
#> [11,] -0.53069257 -0.008440871 -0.51929177     0     0
#> [12,]  0.00000000  0.000000000  0.00000000     0     0
#> [13,]  0.00000000  0.000000000  0.00000000     0     0
#> [14,]  0.00000000  0.000000000  0.00000000     0     0
#> [15,]  0.00000000  0.000000000  0.00000000     0     0
#> [16,]  0.00000000  0.000000000  0.00000000     0     0
#> [17,]  0.00000000  0.000000000  0.00000000     0     0
#> [18,]  0.02576646  0.015525617  0.01010029     0     0
#> [19,]  0.00000000  0.000000000  0.00000000     0     0
#> [20,]  0.00000000  0.000000000  0.00000000     0     0
#> [21,]  1.15767686  0.154159469  1.02543233     0     0
#> [22,] -0.43273854 -0.178355275 -0.25220046     0     0
#> [23,] -0.40236016 -0.531329067  0.10140435     0     0
#> [24,]  0.00000000  0.000000000  0.00000000     0     0
#> [25,]  0.00000000  0.000000000  0.00000000     0     0
#> [26,]  0.00000000  0.000000000  0.00000000     0     0
#> [27,] -0.08108600  0.006257312 -0.08540582     0     0
#> [28,]  0.00000000  0.000000000  0.00000000     0     0
#> [29,]  0.00000000  0.000000000  0.00000000     0     0
#> [30,]  0.00000000  0.000000000  0.00000000     0     0
#> [31,]  0.67362640  1.015695930  0.15719623     0     0
#> 
#> $a[[2]]
#>              [,1]          [,2] [,3] [,4] [,5]        [,6] [,7]        [,8]
#>  [1,]  0.51720994 -5.986584e-16    0    0    0 -0.52198904    0 -0.52814945
#>  [2,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#>  [3,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#>  [4,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#>  [5,] -0.20836616 -1.083058e-17    0    0    0 -0.14271089    0 -0.14339764
#>  [6,] -2.16772123  1.027030e-15    0    0    0 -0.02084656    0 -0.01843194
#>  [7,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#>  [8,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#>  [9,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [10,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [11,]  2.09352307 -8.188946e-16    0    0    0  0.15017767    0  0.14598205
#> [12,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [13,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [14,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [15,]  0.98931795 -2.149275e-16    0    0    0 -0.05489806    0 -0.06791572
#> [16,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [17,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [18,] -0.41338595  2.249685e-16    0    0    0  0.03325704    0  0.03395592
#> [19,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [20,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [21,] -1.02568491  1.614346e-16    0    0    0 -0.57583926    0 -0.58278492
#> [22,]  1.37449732 -8.436940e-17    0    0    0  0.35078594    0  0.33996129
#> [23,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [24,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [25,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [26,] -0.41690661 -7.438619e-17    0    0    0 -0.27848661    0 -0.27761320
#> [27,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [28,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [29,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [30,]  0.00000000  0.000000e+00    0    0    0  0.00000000    0  0.00000000
#> [31,]  0.04980503 -7.722814e-16    0    0    0  0.97849847    0  1.04194267
#>       [,9] [,10]       [,11]       [,12] [,13]       [,14]       [,15]
#>  [1,]    0     0  1.09193410 -0.54711522     0 -1.10385454  1.05761227
#>  [2,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#>  [3,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#>  [4,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#>  [5,]    0     0 -0.05029435 -0.13841194     0  0.07115412 -0.06444747
#>  [6,]    0     0 -2.10771790  0.07022940     0  2.34630176 -2.16648929
#>  [7,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#>  [8,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#>  [9,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [10,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [11,]    0     0  1.91841301  0.08316063     0 -2.08011516  1.96215011
#> [12,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [13,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [14,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [15,]    0     0  1.13166276 -0.02721788     0 -0.95845629  1.06611023
#> [16,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [17,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [18,]    0     0 -0.44220337  0.05183354     0  0.48836644 -0.45119726
#> [19,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [20,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [21,]    0     0 -0.34645736 -0.52342897     0  0.55915597 -0.44228038
#> [22,]    0     0  1.07109497  0.37502995     0 -0.93428573  1.04025704
#> [23,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [24,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [25,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [26,]    0     0 -0.12741939 -0.28331364     0  0.11505491 -0.13836590
#> [27,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [28,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [29,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [30,]    0     0  0.00000000  0.00000000     0  0.00000000  0.00000000
#> [31,]    0     0 -1.50709314  0.65319281     0  0.10973108 -1.00701159
#>             [,16] [,17]        [,18]       [,19] [,20] [,21]        [,22] [,23]
#>  [1,] -1.74471040     0 -0.636052305  0.56637553     0     0 -0.559863527     0
#>  [2,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#>  [3,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#>  [4,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#>  [5,] -0.09255732     0 -0.146919500  0.14848040     0     0  0.196707793     0
#>  [6,]  2.40479980     0  0.198829539 -0.03023710     0     0  2.167908928     0
#>  [7,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#>  [8,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#>  [9,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [10,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [11,] -1.97089674     0  0.003443593 -0.12058515     0     0 -2.081494235     0
#> [12,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [13,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [14,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [15,] -1.02363998     0  0.011289157  0.04127053     0     0 -0.988738247     0
#> [16,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [17,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [18,]  0.54434567     0  0.081923663 -0.04584189     0     0  0.416513908     0
#> [19,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [20,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [21,] -0.07937634     0 -0.505732958  0.57864310     0     0  0.981103532     0
#> [22,] -0.52608984     0  0.450333118 -0.38585785     0     0 -1.340591596     0
#> [23,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [24,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [25,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [26,] -0.21148994     0 -0.319733620  0.29750595     0     0  0.393014457     0
#> [27,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [28,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [29,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [30,]  0.00000000     0  0.000000000  0.00000000     0     0  0.000000000     0
#> [31,]  1.09146260     0  0.264897888 -0.84372353     0     0  0.001487152     0
#>       [,24]       [,25] [,26]       [,27] [,28] [,29] [,30]      [,31]
#>  [1,]     0 -1.14811921     0  1.08939482     0     0     0 -0.5096954
#>  [2,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#>  [3,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#>  [4,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#>  [5,]     0  0.06054764     0 -0.06666699     0     0     0  0.1981714
#>  [6,]     0  2.35115563     0 -2.28559534     0     0     0  2.0369877
#>  [7,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#>  [8,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#>  [9,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [10,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [11,]     0 -2.07664091     0  2.02915473     0     0     0 -2.0106724
#> [12,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [13,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [14,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [15,]     0 -0.97451990     0  0.95734566     0     0     0 -1.1032088
#> [16,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [17,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [18,]     0  0.49221851     0 -0.47642178     0     0     0  0.3877557
#> [19,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [20,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [21,]     0  0.51355820     0 -0.52545684     0     0     0  0.9066361
#> [22,]     0 -0.91957998     0  0.92624660     0     0     0 -1.4719787
#> [23,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [24,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [25,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [26,]     0  0.09606958     0 -0.11122068     0     0     0  0.4298353
#> [27,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [28,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [29,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [30,]     0  0.00000000     0  0.00000000     0     0     0  0.0000000
#> [31,]     0  0.23020593     0 -0.23695852     0     0     0  0.8044191
#>              [,32]       [,33] [,34] [,35]
#>  [1,]  0.637163998 -1.14577339     0     0
#>  [2,]  0.000000000  0.00000000     0     0
#>  [3,]  0.000000000  0.00000000     0     0
#>  [4,]  0.000000000  0.00000000     0     0
#>  [5,]  0.150526685  0.05273307     0     0
#>  [6,] -0.195997896  2.26523838     0     0
#>  [7,]  0.000000000  0.00000000     0     0
#>  [8,]  0.000000000  0.00000000     0     0
#>  [9,]  0.000000000  0.00000000     0     0
#> [10,]  0.000000000  0.00000000     0     0
#> [11,] -0.022808529 -2.01915956     0     0
#> [12,]  0.000000000  0.00000000     0     0
#> [13,]  0.000000000  0.00000000     0     0
#> [14,]  0.000000000  0.00000000     0     0
#> [15,] -0.068460370 -1.04139475     0     0
#> [16,]  0.000000000  0.00000000     0     0
#> [17,]  0.000000000  0.00000000     0     0
#> [18,] -0.082289731  0.47567287     0     0
#> [19,]  0.000000000  0.00000000     0     0
#> [20,]  0.000000000  0.00000000     0     0
#> [21,]  0.498552839  0.43407823     0     0
#> [22,] -0.517549884 -0.97248410     0     0
#> [23,]  0.000000000  0.00000000     0     0
#> [24,]  0.000000000  0.00000000     0     0
#> [25,]  0.000000000  0.00000000     0     0
#> [26,]  0.337777555  0.10076718     0     0
#> [27,]  0.000000000  0.00000000     0     0
#> [28,]  0.000000000  0.00000000     0     0
#> [29,]  0.000000000  0.00000000     0     0
#> [30,]  0.000000000  0.00000000     0     0
#> [31,] -0.002924506  0.75719279     0     0
#> 
#> 
#> $BIC
#> [1] 6669.599
#>