Simulate Heterogeneous, Low-Rank, and Sparse Data

Simulate Heterogeneous, Low-Rank, and Sparse Data

Usage

simulate_lsr(
  N = 100,
  k = 2,
  p = 30,
  m = 35,
  b = 1,
  d = 20,
  h = 0.2,
  case = "independent"
)

Arguments

N

The sample size, default = 100.

k

The number of groups, default = 2.

p

The number of predictor features, default = 30.

m

The number of response features, default = 35.

b

The signal-to-noise ratio, default = 1.

d

The singular value, default = 20.

h

The lower bound for the singular matrix simulation, default = 0.2.

case

The covariance case, "independent" or "dependent", default = "independent".

Value

A list of simulation values, including x matrix, y matrix, coefficients and true clustering assignments.

Examples

simulate_lsr()
#> $x
#>                [,1]         [,2]        [,3]        [,4]        [,5]
#>   [1,]  0.852890075  0.585517124  0.54232737  1.55901248 -0.90384487
#>   [2,] -0.478549119  0.198504479  0.19111429  0.03175168  1.51065146
#>   [3,] -0.365153283  1.333696121 -1.00603369 -2.36528604 -0.50638268
#>   [4,]  0.259828585  1.013596154  0.18842679  1.78926479 -1.22197484
#>   [5,]  0.545548911 -1.434560159 -0.77550183 -0.65612744 -1.00881751
#>   [6,]  0.971135147 -0.265275962 -0.15604521 -0.66454736 -0.30464590
#>   [7,]  0.086953577 -0.855338015  0.77786405 -0.53683055  1.74187928
#>   [8,]  2.408469300  0.530779715  0.31453117 -0.19920403 -0.48707339
#>   [9,]  1.377751977 -0.390955109 -0.82632905  1.49108498  2.38045704
#>  [10,] -0.605293649 -0.839648251  0.80848461 -1.58525382 -0.39829676
#>  [11,]  0.459786474 -0.434328617  1.19970696 -0.45158114 -0.96182963
#>  [12,] -0.321581433 -0.309178825  0.81252719 -0.04872838  0.99296967
#>  [13,] -0.617628512  0.687263538 -0.48882556 -1.00879828 -0.55646206
#>  [14,] -0.398483653 -1.017975519 -0.12169684 -0.98209145 -1.02871896
#>  [15,]  0.075767782 -1.187042063  0.97826134  0.19683937 -0.72080130
#>  [16,]  0.011223349  1.870688156 -2.40396112  0.69119355  0.21528059
#>  [17,] -1.543464702 -0.173732623 -0.18133298  0.77873820  0.04412200
#>  [18,] -0.215552710  0.237940544  1.41664289  0.36540528 -0.49359768
#>  [19,]  0.810279446 -0.927568318 -2.48019888 -0.91065054 -1.19663691
#>  [20,] -0.563147435 -0.399489775  0.65348973 -0.09437556 -1.10194523
#>  [21,] -0.768939001  2.317713728  0.76949345  0.49650611 -1.41750853
#>  [22,] -0.294523704  0.503255053 -1.07991173 -0.28199938  0.10245817
#>  [23,] -0.510452971  1.425988922  0.44373673  0.08111958 -0.34671743
#>  [24,] -0.560077983 -0.213008085 -1.54600267 -1.03106569 -1.23831108
#>  [25,]  0.977404667 -1.592021409 -1.01755845  0.99916984  0.28510994
#>  [26,] -1.629125258  0.465248477 -1.17695186  0.03070972  0.91674552
#>  [27,]  1.441482108  0.236861557 -0.47114320 -0.11882087  0.48121030
#>  [28,] -0.078167819 -1.346637570 -0.16054381  0.33622316 -1.40691962
#>  [29,] -1.557020081  0.406391898 -0.83677479  0.45440199  0.47565727
#>  [30,] -0.144901410 -0.892106228  0.81769966 -0.88723719 -0.01806373
#>  [31,] -0.113397531  0.725747879 -1.99249007 -0.43708627 -2.35766252
#>  [32,]  1.171072003 -0.830661179 -0.82864747  1.41360487  0.26292100
#>  [33,]  0.664748305  0.369451089 -0.29272553 -1.08021658 -1.81911394
#>               [,6]        [,7]         [,8]         [,9]        [,10]
#>   [1,] -0.38405505  0.82986990  0.779058511  0.377527632  1.055489791
#>   [2,] -0.91492998  0.45836733  0.062989630  0.561231277 -0.384393882
#>   [3,] -0.03371052  1.19718486 -0.584200000  0.563072574  0.335120613
#>   [4,]  0.45672410  3.41658350 -1.238277859  0.582205103 -1.563221687
#>   [5,] -1.65244291 -1.45868692  0.507257633 -1.238122250 -0.003341659
#>   [6,] -1.09735342  0.55940779  0.169472503  0.811788213 -2.252765669
#>   [7,]  1.62411795 -0.68127554  1.191842896 -1.170241943  0.032763902
#>   [8,]  1.02315334 -0.58309731 -0.956444732 -0.429314307  0.357821320
#>   [9,] -0.83093243 -0.01034424  0.416664540  0.763403239  0.435820582
#>  [10,]  1.36556853 -0.37802493  1.599322182 -1.010715022  0.115821305
#>  [11,]  0.80573608 -0.66147450 -1.798797833 -0.263650873 -1.108266465
#>  [12,] -0.73262381  1.16701860  1.178710326  0.447637739 -0.319454757
#>  [13,]  2.79139105 -1.42170851  0.015547954 -0.716420145  0.446566105
#>  [14,] -0.82363176 -0.35339752 -0.091823720 -0.575060782 -0.725225164
#>  [15,] -0.11058686  1.65331563  0.265607839 -0.571274814 -0.785259385
#>  [16,] -0.06203942 -1.21170331  2.493213451  1.070731236 -1.037361797
#>  [17,] -0.64009321  0.31195554  0.688151917  0.585506371  2.800855198
#>  [18,]  0.02991015 -0.10582713  0.226357926 -0.165945560 -1.435952094
#>  [19,] -0.05783307 -0.04132792  2.426110151 -2.351070061 -1.109465449
#>  [20,]  0.78116784  0.60321200 -0.885944281 -1.071623151 -0.397540409
#>  [21,]  0.54637750 -2.00313715 -0.508706745 -0.592869995 -0.892439836
#>  [22,]  0.12515989  0.78656726 -0.784771770  0.810592835  1.301096130
#>  [23,] -0.17440466  0.04567514  1.447642130  0.446670774  1.707819428
#>  [24,] -0.61106573  0.77733620  1.566122926 -0.178380306 -0.989475012
#>  [25,] -0.41277711 -0.48506536  0.386813592  0.176565093  0.626820129
#>  [26,] -0.64689305 -0.65641750 -0.467030857  0.067146156  1.108211608
#>  [27,] -0.63099660 -1.03158975  0.932162659  0.949644975  0.346254588
#>  [28,]  0.31031813  0.21645737  0.080427647  0.657262824  0.495908525
#>  [29,]  0.28985033 -0.53753753 -0.546236724 -1.488419179  1.217693410
#>  [30,]  0.72003221 -0.36295234 -1.309956993  1.296125490 -0.496634292
#>  [31,]  1.51523012  1.23725555 -0.230082228  1.594211659  1.619798926
#>  [32,]  0.41526602 -1.01726933  0.671044043  0.117995997  0.172082249
#>  [33,] -0.99077446  1.48030713 -0.979645047  0.580339136  0.483357678
#>              [,11]       [,12]        [,13]        [,14]        [,15]
#>   [1,]  0.66136294 -0.92183071 -0.624303274 -0.644246020  0.728510278
#>   [2,]  0.47265632 -1.19208406 -0.370013863  1.538927940  0.143135336
#>   [3,] -1.51047747  0.27900234 -0.077540331  0.464444602  0.735007480
#>   [4,]  0.62962843  1.09875224 -1.358055973 -0.396372577 -1.563274805
#>   [5,]  1.97410025 -1.96201997 -0.532899071  1.728229240 -0.150092044
#>   [6,]  0.78521038  1.00374199 -0.072529718  0.100606693 -1.102943975
#>   [7,] -0.98501714  0.03181446 -0.241451522  1.706060276 -0.420299736
#>   [8,] -1.07634189  1.45273545 -2.257056349  1.122892065  0.217446332
#>   [9,]  1.82190764  0.76041628 -0.875416341 -0.002360484  0.498430522
#>  [10,]  0.05332355 -0.44158350 -0.525487199 -1.184683191 -0.144439531
#>  [11,] -1.98405256  0.27320346 -0.707596974  0.590873757  1.922456157
#>  [12,] -0.48967758  1.72972937  0.012514270  0.334691113  0.295957018
#>  [13,] -1.05693736 -1.66507185 -1.454392786  0.448838374 -0.505181565
#>  [14,]  1.80152682  0.25998171  0.685794504  0.095894333  1.533553857
#>  [15,] -2.32460834 -0.62008630 -0.758934330 -0.667718235  0.536839590
#>  [16,] -0.67731704  0.84913784  0.785390364 -0.028279790 -1.077810271
#>  [17,]  0.32222375  0.16161225  0.570447269  2.028022764  0.424933403
#>  [18,] -0.97354441  0.41403496 -1.829130561  1.859784531 -1.727437273
#>  [19,] -0.10985584  0.85547910  1.051338972  0.295588479  0.269475956
#>  [20,] -0.52163125  0.36942291  0.343508467  0.718415602  0.956087170
#>  [21,]  0.91839360  1.23716418  1.833569279 -0.624176854  0.501745139
#>  [22,] -1.66776612  0.37697545  0.374038756 -1.041709285 -0.765720211
#>  [23,] -0.65896689  0.33851676 -0.930925528  0.182683587  0.175175296
#>  [24,]  1.27759399  0.60027227 -0.074902586  1.732727736 -0.764315366
#>  [25,] -1.13612394  1.56133590  3.230084027  1.495408563  0.235781413
#>  [26,] -0.05311085  0.32893668 -0.558310616 -0.162191972 -0.357766954
#>  [27,]  1.47343149  0.08195542 -0.879421467 -0.064098263 -0.221312431
#>  [28,] -1.56365447  1.03961979 -0.294167653 -0.140010987  0.390930897
#>  [29,]  0.92884002  0.38055342  0.425885301  1.381694047  0.557335430
#>  [30,] -0.89473209  0.04219835 -0.235945040  0.319948056  0.752582245
#>  [31,]  0.70559931  1.46236610 -0.967542049 -1.271100026  0.548245374
#>  [32,] -0.24072529 -0.58356932 -2.259142514  0.200931390 -0.420281399
#>  [33,]  0.90506677  0.15151383 -0.236836185 -0.796168770  0.929499806
#>               [,16]        [,17]        [,18]       [,19]        [,20]
#>   [1,]  1.229151380  0.162529060  0.846792044 -0.54538276  2.533394121
#>   [2,]  0.841498872 -0.623712296 -0.425884776 -0.34989820  0.542952495
#>   [3,] -0.375238846  0.068443253 -0.508724212  0.09880536 -1.822325570
#>   [4,] -0.042795894 -1.181104817  0.711973941  0.92793759  0.589158094
#>   [5,] -0.414746893  0.830417221 -1.103986196 -0.01498096 -0.157101037
#>   [6,] -0.484867343  0.834838940 -0.440229505 -1.66716170  0.226495723
#>   [7,]  0.250493454  0.057345380  0.718699408 -0.80731128 -0.251075435
#>   [8,]  0.672082061 -0.231102719  0.120840993 -0.24622582 -0.244313281
#>   [9,] -1.341723099  0.117272779  1.533939615 -1.26281921 -0.468235633
#>  [10,]  0.476677359 -0.198517705  0.508862549  0.57583798 -1.417076846
#>  [11,] -0.170670938  1.604059851  0.559256018  1.80448012 -0.605505247
#>  [12,]  0.350888764  0.622627239  1.509496992  1.94283212 -1.961482592
#>  [13,]  0.417988218  1.043826676 -0.634455032 -1.45355385  1.569132209
#>  [14,]  1.177722371  0.054590907 -2.186234513 -1.24564833 -2.511014118
#>  [15,] -0.046418655  0.257994179  0.438799539  1.57843853 -1.883020120
#>  [16,] -0.603563219  0.566130896 -0.728395994 -0.99660635  0.708536887
#>  [17,] -0.548769035 -0.681069788 -1.320440487  1.17710395 -0.160106230
#>  [18,]  2.210983849 -0.301324412  0.486696681  1.16240969  0.513745487
#>  [19,]  1.931561883  0.712879199 -0.368430456 -0.97451728  0.350714441
#>  [20,] -0.018384219  0.485791951  0.514669555 -0.47295057 -0.642237352
#>  [21,] -0.837604933  0.380573930 -1.956397294 -0.87146667  0.110169683
#>  [22,] -0.197287990 -1.018452897 -1.365241597 -1.37137936 -0.899546500
#>  [23,]  0.525506804  1.425708953  1.997409118  0.35000924  1.090104919
#>  [24,] -0.204529296 -0.415665341 -1.308096471  2.30193607 -0.618825387
#>  [25,]  0.472572451 -2.431133432 -0.935507094 -1.49806548 -1.702833013
#>  [26,]  0.220440251 -0.531486647 -1.243899277 -0.92150437  1.885997558
#>  [27,] -1.347536350 -0.551638287 -1.595721847 -1.45420988 -0.593865492
#>  [28,]  2.018903943  1.384493212  0.463965726 -0.15771745  0.511385605
#>  [29,]  0.819129262 -0.669621229 -1.098413437 -0.40274539 -1.236636140
#>  [30,]  0.282806263  0.465170646  0.071818793  0.09477477 -0.223191053
#>  [31,]  0.904952472 -2.425426706  0.026630990 -0.92552622  0.559222539
#>  [32,]  0.802745171 -0.589627177  0.855834074  0.22908418  0.298695804
#>  [33,] -2.152559591 -0.288922067 -1.556605058 -0.62663231 -1.013654415
#>              [,21]       [,22]        [,23]        [,24]       [,25]
#>   [1,]  0.68152731  2.37830743 -1.383126342 -1.665627881 -0.05498046
#>   [2,]  0.08225710  0.57906647  0.585307679  1.549108015  0.90285405
#>   [3,] -0.16220751 -1.14548092  0.287447056 -0.165877449  2.04189569
#>   [4,] -0.59216415 -2.19155572  0.889347402 -1.167062521 -0.27516066
#>   [5,]  2.31565635  1.08420512 -0.864231635 -0.108979790  1.18439621
#>   [6,]  1.35635494 -1.19273442  1.656954019  0.152519396  0.88037677
#>   [7,] -0.40059127  0.74750239 -0.950862482  0.541991357 -0.43225434
#>   [8,]  0.67729530  0.09152017 -1.292362467 -0.072836987 -0.99123061
#>   [9,] -0.04657728  0.96454624 -1.230676327  0.089162496 -0.71209717
#>  [10,] -0.37867064 -1.58843784 -0.833205852  1.647307569  0.78902332
#>  [11,] -0.76722358  0.19647637  0.270120629  1.533078962  0.22194581
#>  [12,] -1.19212310 -0.50571951 -0.736354223 -0.927015668 -0.94278583
#>  [13,]  0.29249214  0.72358801 -1.355981637 -0.620559366  0.53347325
#>  [14,]  1.46739055  0.65092741  0.454494480  0.850989627 -0.32024428
#>  [15,] -0.12138262  0.19588494  1.164784686 -0.914337720  0.84373306
#>  [16,]  1.77008851  2.32914679  0.319506092  0.876481034 -0.31035283
#>  [17,]  1.31565176 -0.73739048  0.531890477  1.295589801  0.04686414
#>  [18,] -0.56003035  0.31152199 -0.907320815 -0.017522630  0.18166942
#>  [19,] -0.35919452  0.74477597  0.348268328 -0.992089740  0.79947190
#>  [20,] -1.64587522  0.33210742  0.554308292  1.262656650 -1.27358221
#>  [21,] -0.23785671  0.27657385  1.527051496 -1.148667484 -0.60926197
#>  [22,] -0.51092949  0.09811944  1.292508200 -1.425464900 -1.22212894
#>  [23,] -1.86756158  1.32018299  1.013907116  0.310897701  0.98838750
#>  [24,] -3.29630284  1.19650789  1.230631657 -0.501132616 -0.09424353
#>  [25,] -0.60939489  1.37784406  1.905504932 -0.912219151 -0.88161503
#>  [26,]  0.32975480 -0.67917195 -0.753910478  0.407325765  1.93691286
#>  [27,] -0.51239460 -0.07239001  1.701483078  0.062322106  0.76113552
#>  [28,]  0.92201698  0.13090221  1.057507794 -1.436635411  1.17159637
#>  [29,] -1.31195135  0.23273891 -0.096640527 -1.995171698  2.01018452
#>  [30,]  0.92042003  1.90234722 -0.464100979  0.752579866 -0.73558323
#>  [31,]  1.21790424 -1.82804352 -1.018030646 -1.179690585 -0.18398538
#>  [32,]  1.63650087 -0.19763980 -0.783729512  0.500305064  0.13295332
#>  [33,] -0.22946503  0.10915973 -1.147348585  0.890461353  1.80801160
#>               [,26]       [,27]        [,28]        [,29]        [,30]
#>   [1,]  0.331069506 -0.11299751  0.600021908 -0.582136597 -0.422257722
#>   [2,] -0.143319948  0.46062078 -0.258460445 -0.139313016  0.050965889
#>   [3,] -0.422762109  0.82697415  0.879391960 -0.152402622  1.997653280
#>   [4,] -1.623582056 -2.27200053 -0.063796680 -1.662419328 -1.074652222
#>   [5,]  0.757364341 -0.00157246  0.546789273  0.884922939 -0.246421783
#>   [6,] -1.159268603 -0.77614411 -1.115413680  0.540655224 -0.055241607
#>   [7,]  1.033524774  1.70031936  0.352369628  0.059725184 -1.053670704
#>   [8,] -1.016270482  0.26226225 -0.549848492  0.785580270 -1.109107311
#>   [9,] -0.128141835 -1.83162398 -0.452803343  0.038890961 -0.793165762
#>  [10,]  1.164733354 -0.60357277 -2.274671146  1.135235923 -1.001016510
#>  [11,] -1.367600589 -0.45119593  1.029825834 -2.152842459  0.495105632
#>  [12,]  0.366362117  0.02769601  0.247460195  1.332896799 -1.068265638
#>  [13,] -0.570480065  0.72664223 -0.985933545 -1.194545857  0.292043530
#>  [14,]  0.242052627  0.53499734 -0.185763272  0.137148559 -0.858818387
#>  [15,] -0.512904976 -0.11378227 -0.031736725  1.310401118 -0.476651942
#>  [16,]  0.434934722  1.39960650  0.643336319 -1.927013074  0.120247797
#>  [17,] -0.240065275 -1.46483544 -0.845026027  0.979315470  0.699545396
#>  [18,] -0.656438211  0.97828024 -0.973428107  0.009325942 -0.438481545
#>  [19,]  0.247975337  0.19250885  1.398137769  1.694200821  1.124704307
#>  [20,]  1.301982272 -0.20455464  1.104662600  0.244768880  0.613794330
#>  [21,]  0.546329576 -1.68484263  0.771321644 -0.979729682  0.547940383
#>  [22,] -1.027053557  0.43130267  1.464010994  1.026916467  0.148024923
#>  [23,]  1.529344726 -0.50478589  0.193536205 -0.502161797  0.259211834
#>  [24,]  2.987622550  0.77786238 -0.236107448 -0.126447283 -0.328900351
#>  [25,]  0.804314123 -0.75215798  0.119074124  0.014839226  0.514864809
#>  [26,] -1.371501212 -1.76998393  0.686458002  1.224460830  0.940134699
#>  [27,]  0.689250698  0.62386500 -1.238719649 -0.341494003  0.863164962
#>  [28,] -0.707769706  0.24401149  0.047306787  1.308174896 -0.963125636
#>  [29,]  0.110056511  0.08433814  1.053399597 -1.318959555 -1.289932972
#>  [30,]  0.004249235 -1.58907256  0.138291951 -0.302740012  0.703601703
#>  [31,]  1.074271901  1.30133210  0.291794003 -0.647088023 -0.858709982
#>  [32,]  0.692285704 -1.41728027 -1.336646215 -0.319207364  0.024551316
#>  [33,] -1.039880212  0.43590052 -0.181972972  1.981178981  1.436608311
#>  [ reached getOption("max.print") -- omitted 67 rows ]
#> 
#> $y
#>                [,1]         [,2]         [,3]         [,4]          [,5]
#>   [1,]  0.280022591  0.260455560 -0.052070621 -0.011986059  -2.069602587
#>   [2,]  0.075111330 -0.476234097  0.232998791 -0.408169232   0.126094716
#>   [3,] -0.332834861 -0.008120404  0.100741644  0.223934595  -0.039109235
#>   [4,] -0.005595405 -0.025784713 -0.043117760  0.153083173   7.281995242
#>   [5,]  0.186422570  0.418998401  0.384864241 -0.217332541   0.139632685
#>   [6,] -0.181031203  0.383944229  0.261514529 -0.352583776   0.748821916
#>   [7,] -0.118172898  0.386022259  0.148908254  0.133183461   0.656258680
#>   [8,]  0.238678468  0.083269262 -0.262656562  0.230841249   0.134311365
#>   [9,] -0.162344901  0.237349073 -0.018293059  0.206544602   0.217472256
#>  [10,] -0.101323520  0.403539451  0.083374528 -0.382712516   0.364328105
#>  [11,]  0.261098218 -0.125836760 -0.089807617 -0.075514127   3.698343951
#>  [12,] -0.140147754  0.101600761  0.452740420 -0.208855095  -1.497877616
#>  [13,] -0.448314027  0.390857027 -0.012026314 -0.202583157   0.019883201
#>  [14,] -0.004463626 -0.030517256  0.149259641  0.186284889  -0.360258066
#>  [15,] -0.004431743 -0.262141313  0.340163451 -0.252758299  -0.140818253
#>  [16,] -0.197716784 -0.019648209  0.350325970  0.218751935  -0.079655138
#>  [17,] -0.231862398  0.056030569  0.297243368  0.022039111  -0.058528578
#>  [18,] -0.361524299  0.014777113  0.467260114 -0.163000968  -0.133905191
#>  [19,] -0.309753863 -0.350288246  0.110768197  0.200449650  -0.136628487
#>  [20,]  0.268556944  0.006603687  0.023850536 -0.025034574  -0.065758068
#>  [21,] -0.031742511 -0.010888670  0.007429619 -0.343417914   3.108967266
#>  [22,] -0.016025598 -0.111763964 -0.466915710 -0.268538728   2.745223969
#>  [23,]  0.266975254 -0.022916071  0.010888151 -0.100550078  -0.158178879
#>  [24,]  0.059707461  0.221631659  0.166086697 -0.010441173   5.478825717
#>  [25,] -0.188712197  0.062967328  0.297470349  0.221082903   9.005921571
#>  [26,]  0.117879841  0.067716082  0.260499835  0.148774077  -2.720934497
#>  [27,]  0.033795922 -0.165330659 -0.123926723  0.144074656   2.659732204
#>  [28,]  0.067716725 -0.320675330  0.180240567 -0.096370892  -0.786107018
#>                [,6]         [,7]          [,8]         [,9]        [,10]
#>   [1,] -0.167719312 -0.363723370  0.0292564826 -0.070230957  0.254006833
#>   [2,]  4.127583203  0.294069042 -0.0899764753 -0.450507890  0.072827412
#>   [3,] -8.900400410 -0.326606024  0.2819843297  0.018320042 -0.206180374
#>   [4,] -0.058458581 -0.090189478  0.0790295704 -0.010146600 -0.082412861
#>   [5,] 10.310798616  0.318965048  0.2890999641  0.094073268  0.181929497
#>   [6,] -1.952863897  0.224457776  0.4383526573 -0.102180000  0.372792890
#>   [7,] -0.229207864 -0.409812044 -0.2408124003  0.084707381  0.595929353
#>   [8,]  3.797759872  0.011045372 -0.3574336557  0.166148718 -0.413558853
#>   [9,]  1.078379097  0.062156297 -0.1022562083 -0.066252739  0.173232956
#>  [10,]  5.328880963  0.202792295 -0.2717103630  0.316564012 -0.452518004
#>  [11,]  0.130805932 -0.117029657  0.1333652589 -0.063701540 -0.123153830
#>  [12,]  0.277678703  0.502471359  0.4219383741  0.101987630 -0.063263269
#>  [13,] -3.521166721  0.062817859  0.4615190301  0.076476646  0.068392071
#>  [14,] -0.258752195  0.008456300 -0.4625604556 -0.030506881  0.219548599
#>  [15,] -1.266939872  0.151268844 -0.1880236574  0.192300347  0.003413389
#>  [16,] -4.045075002  0.025348737 -0.3387331614  0.123545075  0.069155586
#>  [17,]  5.533853770 -0.027394415 -0.0191981301  0.193417978 -0.144410589
#>  [18,]  4.362825072 -0.029405059 -0.0105327934 -0.126476652 -0.055932006
#>  [19,] -0.222130677 -0.190447629  0.2400857871  0.317101975  0.246538625
#>  [20,]  4.646557155  0.004279126  0.2416039718 -0.191677539 -0.186482969
#>  [21,]  0.186471019 -0.050660261 -0.2142185657 -0.057836170 -0.278604020
#>  [22,] -0.114073503 -0.453862861 -0.1224332952 -0.285462836 -0.068371708
#>  [23,]  5.522536139  0.243179648 -0.2349261915 -0.266352188 -0.325576392
#>  [24,]  0.278438330  0.025166178 -0.1809729794  0.342831208 -0.307674873
#>  [25,]  0.150453770 -0.123589307 -0.0054951832 -0.131683146 -0.117724676
#>  [26,] -0.021138913  0.089661923  0.0375909301  0.120110207 -0.199744978
#>  [27,]  0.172165023  0.024627292 -0.0193526082  0.350046589  0.499761719
#>  [28,]  5.092988210 -0.178292958 -0.1061697446 -0.283674120 -0.121755713
#>                [,11]         [,12]         [,13]        [,14]        [,15]
#>   [1,]  -2.197161927  0.3385686866  -2.156083533  0.338338970  0.076010608
#>   [2,]   0.306890537  3.3489441182  -0.201678193  0.145898224 -0.370456884
#>   [3,]  -0.053938118 -9.1472968581  -0.115184128  0.007247377 -0.186453448
#>   [4,]   7.186999317  0.1177301222   7.160529159  0.270175835 -0.164419104
#>   [5,]  -0.161403516 10.6662699534  -0.009703933 -0.083804944 -0.018127201
#>   [6,]  -0.012350377 -1.8733647255   0.594714729 -0.172090760 -0.002800628
#>   [7,]   0.447170294  0.0978273225   0.481968748 -0.339626461 -0.346024645
#>   [8,]  -0.108657477  4.2276716916  -0.124948335 -0.113579135 -0.127217908
#>   [9,]   0.762631228  1.3249896040   0.131782593 -0.135921527 -0.333576343
#>  [10,]  -0.072583024  5.6619310150   0.108909113 -0.011552810 -0.281216256
#>  [11,]   3.499580853  0.2041312082   3.005375232 -0.047101208 -0.097219498
#>  [12,]  -1.669511328 -0.0001960611  -1.967844083  0.471543362  0.126051784
#>  [13,]  -0.295841981 -3.1384435011   0.261464326  0.026930344 -0.197028811
#>  [14,]  -0.207816293 -0.0695549791  -0.413832234 -0.088008198 -0.214739973
#>  [15,]  -0.162009001 -1.4340242919   0.187513602  0.058977113  0.444848390
#>  [16,]   0.035678393 -3.9823208644  -0.069737199  0.028991736 -0.233674776
#>  [17,]   0.060085573  5.3574364972   0.698250183  0.229701158  0.429913098
#>  [18,]  -0.319224819  4.3687715380   0.088810774 -0.019966740  0.102072217
#>  [19,]   0.247537231 -0.1427023274  -0.046236640  0.103862132  0.210904125
#>  [20,]  -0.207604889  4.9829494564  -0.013989235  0.248229078  0.373864308
#>  [21,]   2.844468725 -0.3578970284   2.688312254 -0.014821714  0.070950570
#>  [22,]   2.860784662 -0.1892492020   2.961859576  0.073392406  0.330311344
#>  [23,]  -0.378852434  5.3161769741   0.218477012 -0.007805915  0.488935647
#>  [24,]   5.715123901 -0.2085144937   5.221017885  0.405086949 -0.254830396
#>  [25,]   8.457987053 -0.1427011307   8.823936786  0.001376154  0.268525718
#>  [26,]  -3.102689811 -0.1579088309  -3.117788123  0.326992030  0.237849860
#>  [27,]   2.684815624 -0.3151725733   2.350156570 -0.290143498 -0.300229593
#>  [28,]  -0.198123874  4.8819921887  -0.223311028  0.207522282  0.217768511
#>               [,16]        [,17]        [,18]         [,19]         [,20]
#>   [1,]  -2.38853369  -2.29060983 -0.026820738   0.212201618 -0.1193212526
#>   [2,]   3.63149781   3.90219106 -0.176097233  -3.463011975  3.8289064376
#>   [3,]  -8.89358334  -8.93623566  0.308299172   8.820326989 -9.0222976093
#>   [4,]   7.11834959   7.23113531  0.157023127  -0.023860721  0.3370141987
#>   [5,]  10.04847138  10.41044782 -0.096982364 -10.077450063 10.5008556023
#>   [6,]  -1.56124238  -1.76260582  0.409999250   1.281892078 -1.6171886936
#>   [7,]  -0.05372046   0.29887326  0.296126767  -0.120835275 -0.2401309116
#>   [8,]   4.01241521   3.64713644  0.357566826  -4.409527315  3.6107547829
#>   [9,]   0.81871040   0.82130149  0.133660851  -0.825337060  0.6567576026
#>  [10,]   4.67847519   5.39025737  0.200842882  -5.244279105  5.1367449008
#>  [11,]   3.31428535   3.17528816  0.227903819   0.052791337  0.0511362603
#>  [12,]  -1.57790144  -1.63630192 -0.225174591   0.324365592 -0.2769776516
#>  [13,]  -2.55247234  -2.88967924  0.034427082   2.892854005 -2.8910027647
#>  [14,]  -0.35165714  -0.42005931  0.130934667  -0.147877635 -0.0158011320
#>  [15,]  -1.31914533  -1.34951672 -0.399027684   0.747274745 -0.9439351251
#>  [16,]  -3.25401626  -4.08245439  0.295720042   3.758631454 -3.9589403367
#>  [17,]   5.22646442   5.64071351 -0.365982322  -5.504680232  5.8443358187
#>  [18,]   4.47732664   5.05940545 -0.165172616  -4.353407712  4.4014088832
#>  [19,]   0.16895762   0.21541428 -0.394817909   0.155362195 -0.2499091046
#>  [20,]   4.40484977   4.66738711 -0.123646114  -4.536141180  5.0390477279
#>  [21,]   2.96426791   3.03373283  0.001959678  -0.129128326  0.1740218609
#>  [22,]   2.52342135   3.02234807 -0.058662836   0.217101251  0.0229101046
#>  [23,]   5.67843687   5.53210222 -0.395939185  -5.594974118  5.6530888482
#>  [24,]   5.42960954   5.49539892 -0.297347102  -0.008767759  0.3902010853
#>  [25,]   8.57829890   9.13980868  0.008000988   0.167941026  0.2572240505
#>  [26,]  -3.05599671  -2.91566538 -0.144030014   0.411034196 -0.0271341652
#>  [27,]   2.93535410   2.22959610 -0.147169419   0.261913732  0.1416562176
#>  [28,]   4.83679552   4.82660359  0.135027424  -4.508284455  4.5687694117
#>                [,21]         [,22]        [,23]        [,24]        [,25]
#>   [1,]  0.1820737744   2.700411021   0.05595977  0.208385588   2.56416200
#>   [2,] -0.0924163890  -0.133591956  -3.73639726 -0.194906317  -3.75461761
#>   [3,]  0.2511458278  -0.308924383   9.03254973  0.009877864   9.31489979
#>   [4,]  0.2393085919  -7.211391872   0.04201511  0.044664609  -7.43899221
#>   [5,]  0.0834556945  -0.139889503 -10.47409876  0.188406263 -10.49299910
#>   [6,] -0.1189068856  -0.026920361   1.56127568  0.196749622   1.74390778
#>   [7,] -0.0064579564  -0.376803983  -0.42485623 -0.300945560  -0.42093223
#>   [8,]  0.1498031211   0.070880843  -4.11908714  0.233776043  -4.01347098
#>   [9,]  0.0419613369   0.416954117  -0.74895160 -0.182829802  -1.07521755
#>  [10,] -0.1115906759   0.057097326  -5.40368192  0.158963972  -5.13195931
#>  [11,] -0.0519109470  -3.455347021  -0.18579702  0.216725572  -3.20046197
#>  [12,]  0.2031182502   2.008703590  -0.22063381 -0.011602421   2.38378180
#>  [13,] -0.0046365198  -0.137320511   3.46045277 -0.180322416   2.92859980
#>  [14,]  0.2237490812   0.552129481   0.02460903  0.218530319   0.72505518
#>  [15,]  0.3592434349   0.042883777   1.35916899  0.188431848   1.00381031
#>  [16,] -0.0644336683  -0.010042664   3.94983879  0.082800264   3.45949573
#>  [17,] -0.2529329031  -0.203347650  -5.32561995 -0.200517531  -5.45086626
#>  [18,] -0.1437259221   0.269234757  -4.51391898 -0.056434216  -4.51878796
#>  [19,] -0.1922843472  -0.072623898   0.31295157 -0.055638147  -0.03340977
#>  [20,]  0.0251232392   0.197985670  -4.65818498  0.060614545  -4.99215849
#>  [21,] -0.1293958078  -3.011870045   0.07854629  0.310254456  -2.82097842
#>  [22,] -0.1265764442  -2.868746970   0.01205677 -0.623889619  -2.99951404
#>  [23,] -0.5684608147   0.123136067  -5.36605130  0.349859643  -5.55943903
#>  [24,] -0.2356165660  -6.000398099   0.07669193 -0.033769045  -5.35718780
#>  [25,] -0.2525338236  -8.958666030  -0.31095049  0.157300526  -8.54410068
#>  [26,]  0.1378879700   3.126704307   0.31904534  0.076545609   2.46799190
#>  [27,]  0.0085621610  -2.750150264   0.40787248 -0.125045866  -2.68219065
#>  [28,] -0.2019197391   0.058497617  -4.83823122 -0.374848744  -4.93214095
#>              [,26]        [,27]       [,28]        [,29]        [,30]
#>   [1,] -0.16441318   0.30296491  -2.7790103  0.118079068 -0.446163210
#>   [2,] -0.26485577  -3.47467713   4.0799316 -0.136883791 -0.250445933
#>   [3,]  0.20589720   9.22720955  -9.0135510  0.047766118  0.036788667
#>   [4,] -0.22642708  -0.08550643   6.9468210 -0.045678391  0.379817737
#>   [5,]  0.04473910 -10.59392666  10.7642699 -0.404019107 -0.326587111
#>   [6,] -0.10524030   1.81520133  -1.6266910 -0.088670523  0.289975938
#>   [7,]  0.16499822   0.20683936   0.5738025  0.256268035 -0.014666120
#>   [8,] -0.01458322  -4.35764516   3.9105412 -0.063025697 -0.167948735
#>   [9,] -0.30635781  -0.78321115   0.4580421 -0.185960849 -0.208398469
#>  [10,] -0.20967941  -5.29461382   5.1478673 -0.142437286  0.028098263
#>  [11,] -0.09981581   0.04001402   3.5445115  0.147042082 -0.116824946
#>  [12,]  0.13508329   0.09166625  -1.8256235 -0.100438406 -0.096620810
#>  [13,]  0.06857765   3.25645680  -2.8649033  0.102056609 -0.194457110
#>  [14,]  0.06435974  -0.10241776  -0.2954468 -0.072988239 -0.551239286
#>  [15,] -0.12074932   1.36028677  -1.1371238 -0.154438748  0.088124770
#>  [16,]  0.08486989   3.98119719  -3.4302168  0.049181706  0.006878444
#>  [17,] -0.30753429  -5.91932716   5.4499428 -0.070319600  0.257604565
#>  [18,]  0.08043066  -4.22822373   4.6774180  0.062010836  0.445790462
#>  [19,]  0.22960759  -0.36336097  -0.2869947 -0.220753577 -0.004444637
#>  [20,]  0.18873331  -4.55335181   4.2741918  0.070738372 -0.094369954
#>  [21,]  0.15568047  -0.23436702   2.9787949 -0.160927049  0.060649464
#>  [22,] -0.21644997  -0.34549056   2.5910863 -0.097079397 -0.049902070
#>  [23,]  0.06185495  -5.63492496   5.5557106  0.119014112 -0.334483904
#>  [24,] -0.04957482   0.16959829   5.5974295 -0.159721047 -0.177234471
#>  [25,]  0.15664571  -0.26612296   8.5515564 -0.029526594  0.362238056
#>  [26,] -0.06581258  -0.19948307  -2.5055952  0.371314320 -0.249999705
#>  [27,]  0.01625863   0.21437435   2.3295979  0.203135839 -0.084056331
#>  [28,] -0.09409848  -4.60084586   4.8250870  0.147560806 -0.117239861
#>               [,31]         [,32]         [,33]         [,34]         [,35]
#>   [1,] -0.264559399 -0.3694905773   2.426459640   2.412168880 -0.4757249783
#>   [2,] -0.056533836 -0.0966986212   0.389857381  -0.099546121  0.4613062674
#>   [3,] -0.006654932  0.0245538733  -0.231844394  -0.100624980 -0.1897267496
#>   [4,]  0.063471501  0.0267445428  -7.561151547  -7.428074863  0.1437596239
#>   [5,] -0.384570546 -0.0294134737  -0.164314193  -0.166412666 -0.2592273259
#>   [6,] -0.562697581 -0.2024778324  -0.324247961   0.092649627 -0.5705894679
#>   [7,] -0.004754350  0.1519629747  -0.728829463  -0.172470009  0.1055612123
#>   [8,]  0.291654664  0.0555180832  -0.184849317  -0.165707903  0.5148781599
#>   [9,]  0.336950474 -0.1255386150   0.058720087  -0.071548013  0.1360936798
#>  [10,]  0.004948127 -0.0038466792   0.315372226  -0.091410924 -0.1159161691
#>  [11,]  0.098433701 -0.0160780637  -3.259269841  -3.676941292 -0.0004000427
#>  [12,] -0.139814336  0.0200775600   1.764408654   1.581198454 -0.0386016185
#>  [13,]  0.140966305 -0.0719756399   0.068697630   0.417134611  0.5345354984
#>  [14,]  0.165439490  0.1009594502   0.430852329   0.453417063 -0.3806452451
#>  [15,]  0.041005094 -0.4483420861   0.475678283  -0.016907868 -0.2398327461
#>  [16,] -0.027654417  0.1218515117  -0.345363912   0.245590860 -0.1417498435
#>  [17,] -0.054204392 -0.1391372000  -0.206158910  -0.112630354  0.3397254947
#>  [18,] -0.169360365 -0.0127128403   0.096748062   0.097889615 -0.3665519187
#>  [19,] -0.005972815  0.1973472408  -0.056692824  -0.062336080  0.2075933939
#>  [20,]  0.074477127  0.3335322415  -0.027190460   0.398385878 -0.3360391495
#>  [21,]  0.114171299  0.1258711724  -2.795987923  -3.109882191  0.2751307968
#>  [22,]  0.085712747 -0.0550862435  -2.756393896  -2.831501682 -0.2293568726
#>  [23,] -0.539371903 -0.0287150161   0.037515323  -0.349188980  0.1731376809
#>  [24,] -0.036130929  0.3125112518  -5.178609638  -5.323671390 -0.0883157647
#>  [25,] -0.212032263 -0.0344300007  -8.893178686  -8.634787481 -0.4310103654
#>  [26,]  0.166254547 -0.1103625889   2.603970196   2.948478014 -0.2118960236
#>  [27,] -0.349865652 -0.3370717429  -2.325412880  -2.642089831 -0.1232608630
#>  [28,]  0.116191868 -0.0245818019   0.043533752   0.031285960 -0.0607415406
#>  [ reached getOption("max.print") -- omitted 72 rows ]
#> 
#> $a
#> $a[[1]]
#>       [,1] [,2] [,3] [,4] [,5]      [,6] [,7] [,8] [,9] [,10] [,11]     [,12]
#>  [1,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>  [2,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>  [3,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>  [4,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>  [5,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>  [6,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>  [7,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>  [8,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>  [9,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [10,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [11,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [12,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [13,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [14,]    0    0    0    0    0  2.592556    0    0    0     0     0  2.592556
#> [15,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [16,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [17,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [18,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [19,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [20,]    0    0    0    0    0  1.605957    0    0    0     0     0  1.605957
#> [21,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [22,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [23,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [24,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [25,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [26,]    0    0    0    0    0  3.656523    0    0    0     0     0  3.656523
#> [27,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#> [28,]    0    0    0    0    0  0.000000    0    0    0     0     0  0.000000
#>       [,13] [,14] [,15]     [,16]     [,17] [,18]     [,19]     [,20] [,21]
#>  [1,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>  [2,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>  [3,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>  [4,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>  [5,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>  [6,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>  [7,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>  [8,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>  [9,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [10,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [11,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [12,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [13,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [14,]     0     0     0  2.592556  2.592556     0 -2.592556  2.592556     0
#> [15,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [16,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [17,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [18,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [19,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [20,]     0     0     0  1.605957  1.605957     0 -1.605957  1.605957     0
#> [21,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [22,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [23,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [24,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [25,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [26,]     0     0     0  3.656523  3.656523     0 -3.656523  3.656523     0
#> [27,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#> [28,]     0     0     0  0.000000  0.000000     0  0.000000  0.000000     0
#>       [,22]     [,23] [,24]     [,25] [,26]     [,27]     [,28] [,29] [,30]
#>  [1,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>  [2,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>  [3,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>  [4,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>  [5,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>  [6,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>  [7,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>  [8,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>  [9,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [10,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [11,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [12,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [13,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [14,]     0 -2.592556     0 -2.592556     0 -2.592556  2.592556     0     0
#> [15,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [16,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [17,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [18,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [19,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [20,]     0 -1.605957     0 -1.605957     0 -1.605957  1.605957     0     0
#> [21,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [22,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [23,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [24,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [25,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [26,]     0 -3.656523     0 -3.656523     0 -3.656523  3.656523     0     0
#> [27,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#> [28,]     0  0.000000     0  0.000000     0  0.000000  0.000000     0     0
#>       [,31] [,32] [,33] [,34] [,35]
#>  [1,]     0     0     0     0     0
#>  [2,]     0     0     0     0     0
#>  [3,]     0     0     0     0     0
#>  [4,]     0     0     0     0     0
#>  [5,]     0     0     0     0     0
#>  [6,]     0     0     0     0     0
#>  [7,]     0     0     0     0     0
#>  [8,]     0     0     0     0     0
#>  [9,]     0     0     0     0     0
#> [10,]     0     0     0     0     0
#> [11,]     0     0     0     0     0
#> [12,]     0     0     0     0     0
#> [13,]     0     0     0     0     0
#> [14,]     0     0     0     0     0
#> [15,]     0     0     0     0     0
#> [16,]     0     0     0     0     0
#> [17,]     0     0     0     0     0
#> [18,]     0     0     0     0     0
#> [19,]     0     0     0     0     0
#> [20,]     0     0     0     0     0
#> [21,]     0     0     0     0     0
#> [22,]     0     0     0     0     0
#> [23,]     0     0     0     0     0
#> [24,]     0     0     0     0     0
#> [25,]     0     0     0     0     0
#> [26,]     0     0     0     0     0
#> [27,]     0     0     0     0     0
#> [28,]     0     0     0     0     0
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> $a[[2]]
#>       [,1] [,2] [,3] [,4]      [,5] [,6] [,7] [,8] [,9] [,10]     [,11] [,12]
#>  [1,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#>  [2,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#>  [3,]    0    0    0    0 -3.517918    0    0    0    0     0 -3.517918     0
#>  [4,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#>  [5,]    0    0    0    0  1.463379    0    0    0    0     0  1.463379     0
#>  [6,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#>  [7,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#>  [8,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#>  [9,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [10,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [11,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [12,]    0    0    0    0  2.003502    0    0    0    0     0  2.003502     0
#> [13,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [14,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [15,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [16,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [17,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [18,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [19,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [20,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [21,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [22,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [23,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [24,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [25,]    0    0    0    0 -1.996790    0    0    0    0     0 -1.996790     0
#> [26,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [27,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#> [28,]    0    0    0    0  0.000000    0    0    0    0     0  0.000000     0
#>           [,13] [,14] [,15]     [,16]     [,17] [,18] [,19] [,20] [,21]
#>  [1,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#>  [2,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#>  [3,] -3.517918     0     0 -3.517918 -3.517918     0     0     0     0
#>  [4,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#>  [5,]  1.463379     0     0  1.463379  1.463379     0     0     0     0
#>  [6,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#>  [7,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#>  [8,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#>  [9,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [10,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [11,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [12,]  2.003502     0     0  2.003502  2.003502     0     0     0     0
#> [13,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [14,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [15,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [16,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [17,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [18,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [19,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [20,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [21,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [22,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [23,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [24,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [25,] -1.996790     0     0 -1.996790 -1.996790     0     0     0     0
#> [26,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [27,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#> [28,]  0.000000     0     0  0.000000  0.000000     0     0     0     0
#>           [,22] [,23] [,24]     [,25] [,26] [,27]     [,28] [,29] [,30] [,31]
#>  [1,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#>  [2,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#>  [3,]  3.517918     0     0  3.517918     0     0 -3.517918     0     0     0
#>  [4,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#>  [5,] -1.463379     0     0 -1.463379     0     0  1.463379     0     0     0
#>  [6,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#>  [7,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#>  [8,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#>  [9,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [10,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [11,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [12,] -2.003502     0     0 -2.003502     0     0  2.003502     0     0     0
#> [13,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [14,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [15,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [16,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [17,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [18,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [19,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [20,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [21,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [22,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [23,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [24,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [25,]  1.996790     0     0  1.996790     0     0 -1.996790     0     0     0
#> [26,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [27,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#> [28,]  0.000000     0     0  0.000000     0     0  0.000000     0     0     0
#>       [,32]     [,33]     [,34] [,35]
#>  [1,]     0  0.000000  0.000000     0
#>  [2,]     0  0.000000  0.000000     0
#>  [3,]     0  3.517918  3.517918     0
#>  [4,]     0  0.000000  0.000000     0
#>  [5,]     0 -1.463379 -1.463379     0
#>  [6,]     0  0.000000  0.000000     0
#>  [7,]     0  0.000000  0.000000     0
#>  [8,]     0  0.000000  0.000000     0
#>  [9,]     0  0.000000  0.000000     0
#> [10,]     0  0.000000  0.000000     0
#> [11,]     0  0.000000  0.000000     0
#> [12,]     0 -2.003502 -2.003502     0
#> [13,]     0  0.000000  0.000000     0
#> [14,]     0  0.000000  0.000000     0
#> [15,]     0  0.000000  0.000000     0
#> [16,]     0  0.000000  0.000000     0
#> [17,]     0  0.000000  0.000000     0
#> [18,]     0  0.000000  0.000000     0
#> [19,]     0  0.000000  0.000000     0
#> [20,]     0  0.000000  0.000000     0
#> [21,]     0  0.000000  0.000000     0
#> [22,]     0  0.000000  0.000000     0
#> [23,]     0  0.000000  0.000000     0
#> [24,]     0  0.000000  0.000000     0
#> [25,]     0  1.996790  1.996790     0
#> [26,]     0  0.000000  0.000000     0
#> [27,]     0  0.000000  0.000000     0
#> [28,]     0  0.000000  0.000000     0
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> 
#> $true
#>   [1] 2 1 1 2 1 1 2 1 1 1 2 2 1 2 1 1 1 1 2 1 2 2 1 2 2 2 2 1 2 2 1 2 2 1 2 1 1
#>  [38] 2 2 2 1 2 2 1 2 2 1 1 1 1 2 1 1 2 1 2 2 2 2 2 2 1 2 1 1 2 2 2 2 1 1 2 2 2
#>  [75] 1 1 2 1 2 2 1 2 2 2 1 1 1 1 1 1 2 1 2 2 1 1 2 2 1 1
#>