Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2639.369
#> 
#> $assign
#>  [1] 1 1 2 2 1 1 2 1 1 1 1 1 1 2 1 2 2 1 2 2 1 2 1 1 2 1 2 2 2 2 2 1 2 1 1 2 1 1
#> [39] 1 2 2 1 1 1 2 2 2 1 2 1
#> 
#> $a
#> $a[[1]]
#>                [,1]          [,2]         [,3]          [,4]          [,5] [,6]
#>  [1,] -4.521233e-18 -0.0198824509  0.001408816 -0.0179419396  0.0185323686    0
#>  [2,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#>  [3,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#>  [4,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#>  [5,]  4.435462e-16  0.9492716207 -0.619201300  0.9549476386 -1.0310113729    0
#>  [6,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#>  [7,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#>  [8,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#>  [9,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [10,] -2.184755e-16 -0.4066323989 -0.030949484 -0.4368617869  0.4309178511    0
#> [11,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [12,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [13,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [14,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [15,]  3.194883e-17 -0.0548338249  0.195386705 -0.0176031368  0.0463563753    0
#> [16,] -5.851295e-17 -0.1297475766 -0.062376762 -0.1356335083  0.1282919138    0
#> [17,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [18,] -2.052053e-16 -0.5044040236 -0.080387820 -0.5084985852  0.5016127294    0
#> [19,]  4.669588e-17  0.1018457074  0.097772747  0.1089470015 -0.0971698530    0
#> [20,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [21,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [22,]  4.688170e-16  1.3159550654 -0.162301756  1.2752625848 -1.3083113255    0
#> [23,]  1.080428e-19  0.0001904863  0.000410332  0.0002244936 -0.0001729785    0
#> [24,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [25,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [26,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [27,] -3.382158e-17 -0.2428104131  0.340218752 -0.1939639305  0.2423668094    0
#> [28,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [29,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [30,]  0.000000e+00  0.0000000000  0.000000000  0.0000000000  0.0000000000    0
#> [31,]  6.624956e-16  0.2406748543  0.044536097  0.5447889272 -0.4926504366    0
#>       [,7]          [,8]          [,9]         [,10]         [,11] [,12]
#>  [1,]    0 -0.0191595068 -0.0177802096  0.0168019761 -0.0207865869     0
#>  [2,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#>  [3,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#>  [4,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#>  [5,]    0  1.6057752430  0.9515873900 -1.0002874198  1.5970034390     0
#>  [6,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#>  [7,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#>  [8,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#>  [9,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [10,]    0 -0.3778590194 -0.4120934167  0.4226941615 -0.3638727844     0
#> [11,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [12,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [13,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [14,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [15,]    0 -0.2427646469 -0.0358632321  0.0327328056 -0.2636394665     0
#> [16,]    0 -0.0590589360 -0.1258584394  0.1230079489 -0.0588888292     0
#> [17,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [18,]    0 -0.3965590093 -0.4841962294  0.4769144535 -0.4040076102     0
#> [19,]    0 -0.0063116089  0.0979766558 -0.0929813258 -0.0066549094     0
#> [20,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [21,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [22,]    0  1.4115308759  1.2454495661 -1.2309756143  1.4553801890     0
#> [23,]    0 -0.0002503383  0.0001862217 -0.0001686829 -0.0002583983     0
#> [24,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [25,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [26,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [27,]    0 -0.5759497682 -0.2160842714  0.2179529838 -0.6033435647     0
#> [28,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [29,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [30,]    0  0.0000000000  0.0000000000  0.0000000000  0.0000000000     0
#> [31,]    0  0.4104767650  0.4541376734 -0.6036960279  0.2018490804     0
#>              [,13]        [,14]         [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  0.022754079  0.004938136 -0.0219274879     0     0     0     0     0
#>  [2,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#>  [3,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#>  [4,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#>  [5,] -1.587118597 -0.503359006  1.5221798556     0     0     0     0     0
#>  [6,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#>  [7,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#>  [8,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#>  [9,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [10,]  0.351506613 -0.096548288 -0.3620316420     0     0     0     0     0
#> [11,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [12,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [13,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [14,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [15,]  0.285344077  0.244546976 -0.2540156229     0     0     0     0     0
#> [16,]  0.060386093 -0.067963377 -0.0678804149     0     0     0     0     0
#> [17,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [18,]  0.417110997 -0.085885119 -0.4241665023     0     0     0     0     0
#> [19,]  0.005155621  0.100397234  0.0064025159     0     0     0     0     0
#> [20,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [21,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [22,] -1.512824304 -0.202466373  1.4780315353     0     0     0     0     0
#> [23,]  0.000261346  0.000426208 -0.0002104735     0     0     0     0     0
#> [24,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [25,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [26,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [27,]  0.632144642  0.389931817 -0.5818179553     0     0     0     0     0
#> [28,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [29,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [30,]  0.000000000  0.000000000  0.0000000000     0     0     0     0     0
#> [31,]  0.031072804  0.678514865  0.0654517759     0     0     0     0     0
#>               [,21]         [,22]         [,23] [,24] [,25] [,26] [,27]
#>  [1,]  0.0013867178  0.0034167287 -0.0204748003     0     0     0     0
#>  [2,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#>  [3,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#>  [4,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#>  [5,] -0.6325521030 -0.5972267876  0.9449514284     0     0     0     0
#>  [6,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#>  [7,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#>  [8,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#>  [9,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [10,] -0.0365603948 -0.0505044508 -0.3954978813     0     0     0     0
#> [11,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [12,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [13,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [14,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [15,]  0.2032723020  0.2201479872 -0.0670924023     0     0     0     0
#> [16,] -0.0657872714 -0.0625015972 -0.1274367262     0     0     0     0
#> [17,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [18,] -0.0872859131 -0.0720674934 -0.5016893699     0     0     0     0
#> [19,]  0.1022230991  0.0976294889  0.0991998520     0     0     0     0
#> [20,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [21,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [22,] -0.1592711991 -0.2117396410  1.3260048487     0     0     0     0
#> [23,]  0.0004273288  0.0004172859  0.0001786148     0     0     0     0
#> [24,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [25,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [26,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [27,]  0.3520739257  0.3710940857 -0.2584515015     0     0     0     0
#> [28,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [29,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [30,]  0.0000000000  0.0000000000  0.0000000000     0     0     0     0
#> [31,]  0.0615369556  0.3154699911  0.1389401339     0     0     0     0
#>               [,28] [,29]         [,30] [,31] [,32] [,33] [,34] [,35]
#>  [1,]  0.0021879268     0  3.432566e-05     0     0     0     0     0
#>  [2,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#>  [3,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#>  [4,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#>  [5,] -0.5971804141     0 -6.337783e-01     0     0     0     0     0
#>  [6,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#>  [7,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#>  [8,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#>  [9,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [10,] -0.0519851182     0 -1.716070e-02     0     0     0     0     0
#> [11,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [12,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [13,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [14,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [15,]  0.2125254491     0  1.779866e-01     0     0     0     0     0
#> [16,] -0.0664676119     0 -6.209409e-02     0     0     0     0     0
#> [17,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [18,] -0.0880168191     0 -8.566183e-02     0     0     0     0     0
#> [19,]  0.1017652073     0  9.762598e-02     0     0     0     0     0
#> [20,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [21,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [22,] -0.1658349416     0 -1.287314e-01     0     0     0     0     0
#> [23,]  0.0004263747     0  4.046543e-04     0     0     0     0     0
#> [24,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [25,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [26,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [27,]  0.3596045906     0  3.184495e-01     0     0     0     0     0
#> [28,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [29,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [30,]  0.0000000000     0  0.000000e+00     0     0     0     0     0
#> [31,]  0.2108153961     0 -1.424607e-01     0     0     0     0     0
#> 
#> $a[[2]]
#>                [,1]         [,2]        [,3]        [,4]        [,5] [,6] [,7]
#>  [1,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#>  [2,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#>  [3,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#>  [4,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#>  [5,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#>  [6,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#>  [7,] -3.595553e-17 -0.052403847 -0.06527725 -0.05247484  0.04562830    0    0
#>  [8,] -1.515187e-16 -0.367369083 -0.15871895 -0.37500995  0.34522658    0    0
#>  [9,] -1.248820e-16 -0.093103499 -0.06916480 -0.09695927  0.10139396    0    0
#> [10,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [11,]  8.664102e-17 -0.018939461  0.03587284 -0.01767711  0.00456303    0    0
#> [12,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [13,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [14,] -5.467329e-16  0.173561003 -0.34910598  0.17112399 -0.09769892    0    0
#> [15,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [16,] -6.183512e-17 -0.121589289 -0.07672561 -0.12366072  0.11296451    0    0
#> [17,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [18,] -3.937360e-17 -0.428294253  0.03877296 -0.44044232  0.40825511    0    0
#> [19,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [20,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [21,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [22,]  1.390178e-16  0.549599905 -0.10630444  0.57019778 -0.55159700    0    0
#> [23,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [24,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [25,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [26,] -1.590304e-16 -0.009302179 -0.11377548 -0.01089633  0.02463833    0    0
#> [27,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [28,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [29,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [30,]  0.000000e+00  0.000000000  0.00000000  0.00000000  0.00000000    0    0
#> [31,] -1.658788e-15 -0.261190898 -0.08903756 -0.32023206  0.57975317    0    0
#>              [,8]        [,9]         [,10]       [,11]         [,12]
#>  [1,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#>  [2,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#>  [3,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#>  [4,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#>  [5,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#>  [6,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#>  [7,]  0.01046568 -0.05038117  0.0532468429  0.01409883 -0.0018828979
#>  [8,] -0.21862146 -0.36451724  0.3607806099 -0.21851194  0.0060628831
#>  [9,] -0.02252383 -0.09708187  0.0842813511 -0.03565091 -0.0034002690
#> [10,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [11,] -0.05909236 -0.01444615  0.0254346202 -0.04662736  0.0063389949
#> [12,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [13,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [14,]  0.54615333  0.15235205 -0.2045442211  0.48533499 -0.0492331693
#> [15,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [16,] -0.04850988 -0.11999654  0.1200337090 -0.04714259  0.0003054416
#> [17,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [18,] -0.47724380 -0.42876887  0.4180087590 -0.48349765  0.0239732134
#> [19,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [20,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [21,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [22,]  0.66119720  0.56047436 -0.5223217028  0.69588862 -0.0301043775
#> [23,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [24,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [25,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [26,]  0.10878986 -0.01400542  0.0008373783  0.09404822 -0.0116536290
#> [27,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [28,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [29,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [30,]  0.00000000  0.00000000  0.0000000000  0.00000000  0.0000000000
#> [31,] -0.08502828 -0.37784645  0.0856321736 -0.40408281 -0.0565553064
#>              [,13]        [,14]       [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#>  [2,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#>  [3,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#>  [4,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#>  [5,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#>  [6,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#>  [7,] -0.005858351 -0.067433421  0.00273219     0     0     0     0     0
#>  [8,]  0.226220612 -0.185937981 -0.23526583     0     0     0     0     0
#>  [9,]  0.019360987 -0.083446682 -0.02776999     0     0     0     0     0
#> [10,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [11,]  0.063605038  0.041562890 -0.05712071     0     0     0     0     0
#> [12,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [13,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [14,] -0.561058783 -0.372098893  0.51688482     0     0     0     0     0
#> [15,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [16,]  0.052678460 -0.085150226 -0.05693913     0     0     0     0     0
#> [17,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [18,]  0.473649613  0.004886938 -0.47062385     0     0     0     0     0
#> [19,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [20,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [21,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [22,] -0.641432982 -0.048545556  0.64193829     0     0     0     0     0
#> [23,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [24,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [25,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [26,] -0.111055223 -0.123205097  0.09859429     0     0     0     0     0
#> [27,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [28,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [29,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [30,]  0.000000000  0.000000000  0.00000000     0     0     0     0     0
#> [31,] -0.063492209 -0.283154447 -0.04085471     0     0     0     0     0
#>             [,21]       [,22]        [,23] [,24]         [,25] [,26] [,27]
#>  [1,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#>  [2,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#>  [3,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#>  [4,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#>  [5,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#>  [6,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#>  [7,] -0.05802872 -0.06238710 -0.060474453     0  0.0088386094     0     0
#>  [8,] -0.14725306 -0.16460067 -0.391924905     0  0.0060431926     0     0
#>  [9,] -0.06024667 -0.08545023 -0.093773773     0  0.0001382914     0     0
#> [10,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [11,]  0.02837897  0.04969387 -0.024464434     0 -0.0008670362     0     0
#> [12,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [13,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [14,] -0.28504263 -0.41661719  0.193359851     0  0.0342941321     0     0
#> [15,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [16,] -0.06964384 -0.07735725 -0.131908512     0  0.0062116706     0     0
#> [17,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [18,]  0.02089714  0.02633406 -0.439933676     0 -0.0293281213     0     0
#> [19,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [20,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [21,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [22,] -0.07900641 -0.06112459  0.546355059     0  0.0577507099     0     0
#> [23,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [24,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [25,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [26,] -0.09564195 -0.13102556 -0.008346864     0  0.0101957041     0     0
#> [27,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [28,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [29,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [30,]  0.00000000  0.00000000  0.000000000     0  0.0000000000     0     0
#> [31,] -0.03927910 -0.44726796 -0.100309982     0 -0.1384464272     0     0
#>             [,28] [,29]       [,30] [,31] [,32] [,33] [,34] [,35]
#>  [1,]  0.00000000     0  0.00000000     0     0     0     0     0
#>  [2,]  0.00000000     0  0.00000000     0     0     0     0     0
#>  [3,]  0.00000000     0  0.00000000     0     0     0     0     0
#>  [4,]  0.00000000     0  0.00000000     0     0     0     0     0
#>  [5,]  0.00000000     0  0.00000000     0     0     0     0     0
#>  [6,]  0.00000000     0  0.00000000     0     0     0     0     0
#>  [7,] -0.06304734     0 -0.06813949     0     0     0     0     0
#>  [8,] -0.16085732     0 -0.18494785     0     0     0     0     0
#>  [9,] -0.07261452     0 -0.08881371     0     0     0     0     0
#> [10,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [11,]  0.03794575     0  0.04772190     0     0     0     0     0
#> [12,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [13,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [14,] -0.35156737     0 -0.41224891     0     0     0     0     0
#> [15,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [16,] -0.07620890     0 -0.08553647     0     0     0     0     0
#> [17,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [18,]  0.02671146     0  0.01456209     0     0     0     0     0
#> [19,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [20,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [21,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [22,] -0.08098072     0 -0.05502035     0     0     0     0     0
#> [23,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [24,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [25,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [26,] -0.11443687     0 -0.13313264     0     0     0     0     0
#> [27,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [28,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [29,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [30,]  0.00000000     0  0.00000000     0     0     0     0     0
#> [31,] -0.18599648     0 -0.38441918     0     0     0     0     0
#> 
#> 
#> $BIC
#> [1] 6753.57
#>