Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2645.009
#> 
#> $assign
#>  [1] 2 1 2 2 1 2 2 2 1 2 1 1 1 1 2 2 2 1 1 2 1 2 1 2 1 2 1 2 1 2 1 2 1 2 2 2 2 1
#> [39] 1 1 1 2 2 1 1 2 2 1 1 2
#> 
#> $a
#> $a[[1]]
#>              [,1]         [,2]         [,3] [,4]        [,5] [,6]        [,7]
#>  [1,] -0.18010338 -1.281930155  1.287129273    0 -1.48190927    0 -1.31598255
#>  [2,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#>  [3,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#>  [4,] -0.10377239 -0.006367569  0.002025945    0 -0.06196960    0 -0.03183108
#>  [5,] -0.52110165 -0.237993825  0.278839426    0 -0.88647216    0 -0.21193632
#>  [6,]  0.78784767  0.267094530 -0.273266873    0  0.92198415    0  0.36581209
#>  [7,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#>  [8,] -0.06827246 -0.162497994  0.168715736    0 -0.25553578    0 -0.15916599
#>  [9,]  0.83551900 -0.104598650  0.057942831    0  0.81432294    0 -0.10740134
#> [10,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [11,]  0.10342894 -0.008101585  0.008426245    0  0.07027083    0  0.00696819
#> [12,]  0.67771440  0.312650895 -0.336216579    0  0.98413092    0  0.35310664
#> [13,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [14,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [15,] -0.58660885  0.074448612 -0.051342823    0 -0.51454147    0  0.05217577
#> [16,] -1.10188292 -0.683495528  0.723949611    0 -1.79173825    0 -0.74690893
#> [17,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [18,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [19,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [20,] -0.63246654 -0.193270248  0.222917346    0 -0.86217914    0 -0.21008904
#> [21,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [22,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [23,] -0.04406883 -0.858117659  0.868182764    0 -0.97079925    0 -0.85371611
#> [24,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [25,] -1.03577645  0.350640129 -0.311958794    0 -0.67195759    0  0.30976412
#> [26,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [27,]  0.91049201 -1.224861999  1.150059376    0 -0.10976062    0 -1.30725679
#> [28,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [29,]  1.29551453  0.242507208 -0.298744162    0  1.58299323    0  0.28560680
#> [30,]  0.00000000  0.000000000  0.000000000    0  0.00000000    0  0.00000000
#> [31,]  0.49443625  0.931739091 -0.822513335    0  0.70235054    0  1.29112468
#>              [,8]        [,9] [,10] [,11] [,12] [,13] [,14] [,15]        [,16]
#>  [1,]  0.16004933  0.13791073     0     0     0     0     0     0  0.166391954
#>  [2,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#>  [3,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#>  [4,]  0.04152198  0.10428202     0     0     0     0     0     0  0.068812125
#>  [5,]  0.62740085  0.48739267     0     0     0     0     0     0  0.574328105
#>  [6,] -0.57711889 -0.76644470     0     0     0     0     0     0 -0.665375024
#>  [7,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#>  [8,]  0.08828863  0.05962226     0     0     0     0     0     0  0.078457786
#>  [9,] -0.87903683 -0.80470032     0     0     0     0     0     0 -0.851919058
#> [10,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [11,] -0.06808434 -0.10242183     0     0     0     0     0     0 -0.083191130
#> [12,] -0.61856689 -0.64807036     0     0     0     0     0     0 -0.639040547
#> [13,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [14,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [15,]  0.55270296  0.56966580     0     0     0     0     0     0  0.562891022
#> [16,]  1.02040733  1.04756241     0     0     0     0     0     0  1.046878495
#> [17,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [18,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [19,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [20,]  0.62816660  0.60396329     0     0     0     0     0     0  0.624283012
#> [21,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [22,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [23,]  0.09914642  0.01367720     0     0     0     0     0     0  0.073017770
#> [24,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [25,]  0.96126607  1.01325100     0     0     0     0     0     0  0.986088778
#> [26,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [27,] -1.11537991 -0.89759906     0     0     0     0     0     0 -1.015419049
#> [28,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [29,] -1.25635909 -1.24375652     0     0     0     0     0     0 -1.262300079
#> [30,]  0.00000000  0.00000000     0     0     0     0     0     0  0.000000000
#> [31,]  0.39143416 -0.51347576     0     0     0     0     0     0 -0.004972212
#>       [,17]       [,18]       [,19] [,20] [,21] [,22] [,23] [,24]       [,25]
#>  [1,]     0 -0.15471451  1.44134488     0     0     0     0     0 -1.15180024
#>  [2,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#>  [3,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#>  [4,]     0 -0.07050149  0.11312585     0     0     0     0     0  0.09338792
#>  [5,]     0 -0.59716047  0.73309460     0     0     0     0     0  0.24636277
#>  [6,]     0  0.68588005 -1.05071778     0     0     0     0     0 -0.47699910
#>  [7,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#>  [8,]     0 -0.07981559  0.22496378     0     0     0     0     0 -0.10341176
#>  [9,]     0  0.89026382 -0.70761039     0     0     0     0     0 -0.89986592
#> [10,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [11,]     0  0.08618542 -0.09607818     0     0     0     0     0 -0.10735839
#> [12,]     0  0.66053471 -0.97445889     0     0     0     0     0 -0.32185893
#> [13,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [14,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [15,]     0 -0.58716384  0.50167130     0     0     0     0     0  0.63435948
#> [16,]     0 -1.07991781  1.75571758     0     0     0     0     0  0.34215419
#> [17,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [18,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [19,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [20,]     0 -0.64785913  0.80764821     0     0     0     0     0  0.40083958
#> [21,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [22,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [23,]     0 -0.06482272  0.88401375     0     0     0     0     0 -0.84522777
#> [24,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [25,]     0 -1.03148897  0.67102950     0     0     0     0     0  1.34667973
#> [26,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [27,]     0  1.07965785  0.33847390     0     0     0     0     0 -2.12351614
#> [28,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [29,]     0  1.31154035 -1.50588655     0     0     0     0     0 -0.98009722
#> [30,]     0  0.00000000  0.00000000     0     0     0     0     0  0.00000000
#> [31,]     0 -0.02426475 -1.48132189     0     0     0     0     0  0.47147600
#>             [,26]       [,27]       [,28] [,29] [,30] [,31] [,32] [,33]
#>  [1,] -0.21421563 -1.29409083  1.30672737     0     0     0     0     0
#>  [2,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#>  [3,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#>  [4,] -0.05093220  0.02875164  0.01483904     0     0     0     0     0
#>  [5,] -0.61441006 -0.32804302  0.23461946     0     0     0     0     0
#>  [6,]  0.61459817  0.16827168 -0.30415245     0     0     0     0     0
#>  [7,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#>  [8,] -0.09098844 -0.17996182  0.16345172     0     0     0     0     0
#>  [9,]  0.85903261 -0.04685522  0.10291127     0     0     0     0     0
#> [10,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [11,]  0.07248921 -0.02650982  0.00302813     0     0     0     0     0
#> [12,]  0.63418227  0.30562806 -0.33150979     0     0     0     0     0
#> [13,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [14,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [15,] -0.54919093  0.07447262 -0.06570514     0     0     0     0     0
#> [16,] -1.05059001 -0.68376737  0.71524778     0     0     0     0     0
#> [17,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [18,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [19,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [20,] -0.63033839 -0.21735957  0.20330473     0     0     0     0     0
#> [21,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [22,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [23,] -0.12495163 -0.90643803  0.86589007     0     0     0     0     0
#> [24,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [25,] -0.94845238  0.36295738 -0.33700526     0     0     0     0     0
#> [26,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [27,]  1.02206167 -1.07840194  1.25938953     0     0     0     0     0
#> [28,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [29,]  1.25947805  0.27061559 -0.26416158     0     0     0     0     0
#> [30,]  0.00000000  0.00000000  0.00000000     0     0     0     0     0
#> [31,] -0.21285828  0.40630051 -1.05584491     0     0     0     0     0
#>             [,34]       [,35]
#>  [1,] -1.29907033 -1.44257502
#>  [2,]  0.00000000  0.00000000
#>  [3,]  0.00000000  0.00000000
#>  [4,]  0.02495775 -0.07586338
#>  [5,] -0.32812253 -0.83540536
#>  [6,]  0.18538894  0.94913439
#>  [7,]  0.00000000  0.00000000
#>  [8,] -0.17994522 -0.24310950
#>  [9,] -0.04180107  0.78219622
#> [10,]  0.00000000  0.00000000
#> [11,] -0.02399118  0.07727195
#> [12,]  0.31486576  0.97154977
#> [13,]  0.00000000  0.00000000
#> [14,]  0.00000000  0.00000000
#> [15,]  0.06794528 -0.50962804
#> [16,] -0.69844696 -1.76183730
#> [17,]  0.00000000  0.00000000
#> [18,]  0.00000000  0.00000000
#> [19,]  0.00000000  0.00000000
#> [20,] -0.22343580 -0.83958649
#> [21,]  0.00000000  0.00000000
#> [22,]  0.00000000  0.00000000
#> [23,] -0.90593800 -0.92790327
#> [24,]  0.00000000  0.00000000
#> [25,]  0.35117982 -0.67407931
#> [26,]  0.00000000  0.00000000
#> [27,] -1.08232957 -0.15334394
#> [28,]  0.00000000  0.00000000
#> [29,]  0.28416791  1.54976301
#> [30,]  0.00000000  0.00000000
#> [31,]  0.45337447  0.90159315
#> 
#> $a[[2]]
#>              [,1]        [,2]         [,3] [,4]        [,5] [,6]        [,7]
#>  [1,] -0.12947334 -0.53482538  0.617873745    0 -0.66385967    0 -0.57620227
#>  [2,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#>  [3,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#>  [4,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#>  [5,]  1.04546689  0.35992119 -0.274174896    0  1.30784258    0  0.35992059
#>  [6,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#>  [7,] -0.08130578  0.18554616 -0.188869540    0  0.10307164    0  0.18472654
#>  [8,]  0.39815337 -0.03468692  0.061757989    0  0.33426567    0 -0.03279270
#>  [9,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [10,]  0.08381910  0.33522733 -0.323745161    0  0.39872496    0  0.33413314
#> [11,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [12,] -0.62531641  0.62860083 -0.634817068    0  0.01705481    0  0.61176678
#> [13,]  0.49804095  0.12949708 -0.106810446    0  0.58827169    0  0.13712920
#> [14,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [15,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [16,] -0.79803619 -0.52905857  0.562520555    0 -1.27443546    0 -0.57200287
#> [17,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [18,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [19,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [20,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [21,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [22,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [23,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [24,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [25,] -0.45372368  0.32183818 -0.365331798    0 -0.10483848    0  0.32588063
#> [26,] -0.08510688 -0.45093030  0.321010950    0 -0.47437506    0 -0.39965606
#> [27,]  0.18231613  0.01300158  0.004419661    0  0.17933013    0  0.01180624
#> [28,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [29,]  0.43146219  0.17721180 -0.148546169    0  0.56948744    0  0.18016229
#> [30,]  0.00000000  0.00000000  0.000000000    0  0.00000000    0  0.00000000
#> [31,]  0.16210329  0.37757415  0.020182423    0  0.39252912    0  0.21438807
#>              [,8]        [,9] [,10] [,11] [,12] [,13] [,14] [,15]       [,16]
#>  [1,]  0.12928197  0.17875142     0     0     0     0     0     0  0.12664041
#>  [2,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#>  [3,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#>  [4,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#>  [5,] -0.99725187 -0.97570188     0     0     0     0     0     0 -1.01121680
#>  [6,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#>  [7,]  0.09432083  0.08809998     0     0     0     0     0     0  0.08594929
#>  [8,] -0.39453997 -0.38386798     0     0     0     0     0     0 -0.39150635
#>  [9,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [10,] -0.05598422 -0.06085617     0     0     0     0     0     0 -0.07062369
#> [11,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [12,]  0.67664321  0.65889917     0     0     0     0     0     0  0.63942281
#> [13,] -0.48601253 -0.48205616     0     0     0     0     0     0 -0.48630154
#> [14,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [15,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [16,]  0.78621251  0.81751784     0     0     0     0     0     0  0.78165260
#> [17,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [18,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [19,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [20,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [21,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [22,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [23,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [24,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [25,]  0.46431517  0.43912221     0     0     0     0     0     0  0.45274715
#> [26,] -0.00273575 -0.04338689     0     0     0     0     0     0  0.04701676
#> [27,] -0.17639637 -0.17024273     0     0     0     0     0     0 -0.17746351
#> [28,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [29,] -0.41244512 -0.40711252     0     0     0     0     0     0 -0.41763274
#> [30,]  0.00000000  0.00000000     0     0     0     0     0     0  0.00000000
#> [31,]  0.03636806  0.18899029     0     0     0     0     0     0 -0.07770208
#>       [,17]       [,18]       [,19] [,20] [,21] [,22] [,23] [,24]      [,25]
#>  [1,]     0 -0.08545875  0.69243925     0     0     0     0     0 -0.4059382
#>  [2,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#>  [3,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#>  [4,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#>  [5,]     0  0.99121028 -1.24431647     0     0     0     0     0 -0.6633204
#>  [6,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#>  [7,]     0 -0.07995848 -0.11043134     0     0     0     0     0  0.2769679
#>  [8,]     0  0.37892408 -0.30900864     0     0     0     0     0 -0.4347414
#>  [9,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [10,]     0  0.07529012 -0.39716134     0     0     0     0     0  0.2701282
#> [11,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [12,]     0 -0.59456652 -0.05490314     0     0     0     0     0  1.2966586
#> [13,]     0  0.46745608 -0.56286914     0     0     0     0     0 -0.3653545
#> [14,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [15,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [16,]     0 -0.72202867  1.26019079     0     0     0     0     0  0.2677147
#> [17,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [18,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [19,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [20,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [21,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [22,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [23,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [24,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [25,]     0 -0.44077782  0.06765565     0     0     0     0     0  0.7884433
#> [26,]     0 -0.11025989  0.43851977     0     0     0     0     0 -0.4247205
#> [27,]     0  0.17450767 -0.16669821     0     0     0     0     0 -0.1672582
#> [28,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [29,]     0  0.40651265 -0.54578099     0     0     0     0     0 -0.2455649
#> [30,]     0  0.00000000  0.00000000     0     0     0     0     0  0.0000000
#> [31,]     0  0.26430031 -0.27067029     0     0     0     0     0  0.3466795
#>             [,26]        [,27]        [,28] [,29] [,30] [,31] [,32] [,33]
#>  [1,] -0.13886524 -0.592817241  0.569918847     0     0     0     0     0
#>  [2,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#>  [3,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#>  [4,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#>  [5,]  0.97388588  0.323807744 -0.291905635     0     0     0     0     0
#>  [6,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#>  [7,] -0.07842975  0.187748687 -0.182695989     0     0     0     0     0
#>  [8,]  0.37375464 -0.046129595  0.052266528     0     0     0     0     0
#>  [9,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [10,]  0.07426578  0.331583352 -0.318213558     0     0     0     0     0
#> [11,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [12,] -0.60081525  0.627737925 -0.617565691     0     0     0     0     0
#> [13,]  0.46869000  0.123441051 -0.109438770     0     0     0     0     0
#> [14,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [15,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [16,] -0.76499999 -0.565603371  0.530602523     0     0     0     0     0
#> [17,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [18,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [19,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [20,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [21,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [22,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [23,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [24,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [25,] -0.42536235  0.344829872 -0.341585217     0     0     0     0     0
#> [26,] -0.04591904 -0.373495580  0.357918535     0     0     0     0     0
#> [27,]  0.16967738  0.004798787 -0.001044086     0     0     0     0     0
#> [28,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [29,]  0.40328840  0.166744604 -0.152450233     0     0     0     0     0
#> [30,]  0.00000000  0.000000000  0.000000000     0     0     0     0     0
#> [31,]  0.05547294  0.132001060 -0.121884017     0     0     0     0     0
#>             [,34]       [,35]
#>  [1,] -0.48340902 -0.63462290
#>  [2,]  0.00000000  0.00000000
#>  [3,]  0.00000000  0.00000000
#>  [4,]  0.00000000  0.00000000
#>  [5,]  0.35990331  1.28482899
#>  [6,]  0.00000000  0.00000000
#>  [7,]  0.18828234  0.09897173
#>  [8,] -0.03854075  0.32987126
#>  [9,]  0.00000000  0.00000000
#> [10,]  0.33904045  0.38840535
#> [11,]  0.00000000  0.00000000
#> [12,]  0.65755066  0.01158741
#> [13,]  0.11898685  0.57624392
#> [14,]  0.00000000  0.00000000
#> [15,]  0.00000000  0.00000000
#> [16,] -0.47372601 -1.23679222
#> [17,]  0.00000000  0.00000000
#> [18,]  0.00000000  0.00000000
#> [19,]  0.00000000  0.00000000
#> [20,]  0.00000000  0.00000000
#> [21,]  0.00000000  0.00000000
#> [22,]  0.00000000  0.00000000
#> [23,]  0.00000000  0.00000000
#> [24,]  0.00000000  0.00000000
#> [25,]  0.32013013 -0.10978369
#> [26,] -0.52255305 -0.47563962
#> [27,]  0.01420993  0.17710169
#> [28,]  0.00000000  0.00000000
#> [29,]  0.17348956  0.55828867
#> [30,]  0.00000000  0.00000000
#> [31,]  0.59768108  0.42824883
#> 
#> 
#> $BIC
#> [1] 7105.197
#>