Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2699.595
#> 
#> $assign
#>  [1] 2 1 2 2 2 1 1 1 2 2 1 1 1 1 2 2 1 2 1 1 1 2 1 2 2 2 2 2 1 1 2 1 1 2 1 2 2 2
#> [39] 2 1 2 2 2 1 1 1 1 2 2 2
#> 
#> $a
#> $a[[1]]
#>                [,1]         [,2]          [,3] [,4] [,5]        [,6]
#>  [1,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#>  [2,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#>  [3,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#>  [4,] -3.691073e-16  0.691125389 -1.358208e-17    0    0 -0.15440440
#>  [5,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#>  [6,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#>  [7,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#>  [8,]  2.324961e-17  0.058064723  3.115477e-17    0    0 -0.81812909
#>  [9,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [10,] -3.887889e-16  0.893696326  3.762535e-17    0    0 -1.38363479
#> [11,]  7.697859e-17 -0.034811203  2.543142e-17    0    0 -1.37419481
#> [12,]  1.665172e-18  0.009571766  8.313921e-18    0    0  0.12755372
#> [13,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [14,]  1.070868e-15 -1.885460369  1.168349e-16    0    0  1.62413020
#> [15,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [16,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [17,]  2.993081e-16 -0.231896381  1.752236e-16    0    0  0.86123468
#> [18,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [19,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [20,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [21,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [22,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [23,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [24,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [25,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [26,]  7.491214e-17 -0.129574740  3.000408e-18    0    0 -0.20754808
#> [27,] -1.718890e-16  0.304842906 -1.906311e-17    0    0 -0.32839588
#> [28,] -6.863424e-17  0.160598144  1.019351e-17    0    0 -0.12782107
#> [29,]  0.000000e+00  0.000000000  0.000000e+00    0    0  0.00000000
#> [30,]  9.812513e-18 -0.033872751 -7.046401e-19    0    0  0.31368335
#> [31,]  3.699327e-16  0.013233126  3.344025e-16    0    0  0.08523193
#>               [,7] [,8] [,9] [,10] [,11]       [,12]        [,13]       [,14]
#>  [1,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#>  [2,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#>  [3,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#>  [4,]  0.716066658    0    0     0     0  0.59717449 -0.581225918 -0.58626945
#>  [5,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#>  [6,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#>  [7,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#>  [8,]  0.046006243    0    0     0     0 -0.71319460  0.717034652  0.68886432
#>  [9,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [10,]  0.904979636    0    0     0     0 -0.34884330  0.369935021  0.31734006
#> [11,] -0.059742782    0    0     0     0 -1.34617990  1.362455302  1.33610126
#> [12,]  0.011920846    0    0     0     0  0.13384770 -0.139740446 -0.14472685
#> [13,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [14,] -1.934937358    0    0     0     0 -0.46424400  0.363645078  0.33116964
#> [15,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [16,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [17,] -0.230130308    0    0     0     0  0.60582528 -0.698341597 -0.80788050
#> [18,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [19,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [20,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [21,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [22,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [23,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [24,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [25,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [26,] -0.138226162    0    0     0     0 -0.33790626  0.339414499  0.33839624
#> [27,]  0.311742287    0    0     0     0  0.01213284  0.005508541  0.01041786
#> [28,]  0.164582806    0    0     0     0  0.05344158 -0.053476401 -0.06443404
#> [29,]  0.000000000    0    0     0     0  0.00000000  0.000000000  0.00000000
#> [30,] -0.029933778    0    0     0     0  0.26304181 -0.269319841 -0.26599465
#> [31,]  0.008444379    0    0     0     0  0.15961669 -0.300522910 -0.52921158
#>       [,15]        [,16]      [,17]       [,18]      [,19]       [,20]
#>  [1,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#>  [2,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#>  [3,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#>  [4,]     0 -0.573652368 -0.1424193  0.56167727 -1.3054212 -0.74249782
#>  [5,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#>  [6,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#>  [7,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#>  [8,]     0  0.725381633 -0.7691572 -0.71828195  0.6805029 -0.01238631
#>  [9,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [10,]     0  0.393420845 -1.2964604 -0.39499223 -0.5294014 -0.88463748
#> [11,]     0  1.365924836 -1.2970335 -1.35935849  1.4351861  0.11534665
#> [12,]     0 -0.136686819  0.1220311  0.13876561 -0.1512415 -0.01482472
#> [13,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [14,]     0  0.370616159  1.5388092 -0.31688811  2.3258422  1.98372429
#> [15,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [16,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [17,]     0 -0.646515811  0.8391948  0.69501033 -0.4435523  0.25092389
#> [18,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [19,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [20,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [21,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [22,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [23,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [24,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [25,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [26,]     0  0.337497364 -0.1970493 -0.33521129  0.4807189  0.15157660
#> [27,]     0  0.004151522 -0.3110991 -0.01290451 -0.3102560 -0.31736498
#> [28,]     0 -0.047881464 -0.1181224  0.04852742 -0.2172594 -0.16435106
#> [29,]     0  0.000000000  0.0000000  0.00000000  0.0000000  0.00000000
#> [30,]     0 -0.269286784  0.2964818  0.26961794 -0.2411234  0.02017808
#> [31,]     0 -0.197800336  0.1304898  0.28429973 -0.2567204  0.06696598
#>             [,21]         [,22] [,23] [,24]        [,25]        [,26] [,27]
#>  [1,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#>  [2,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#>  [3,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#>  [4,]  0.13070075  0.6588687010     0     0  0.728180834 -0.582162415     0
#>  [5,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#>  [6,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#>  [7,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#>  [8,]  0.72804354  0.0501331839     0     0  0.046332396  0.705061011     0
#>  [9,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [10,]  1.21307408  0.8381568304     0     0  0.916402592  0.352814166     0
#> [11,]  1.26551442 -0.0176323457     0     0 -0.048684545  1.335781436     0
#> [12,] -0.13215036 -0.0009829512     0     0  0.006458729 -0.135021804     0
#> [13,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [14,] -1.60029928 -1.8919517058     0     0 -2.020591035  0.410275853     0
#> [15,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [16,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [17,] -1.03832295 -0.3781191194     0     0 -0.318601693 -0.644334185     0
#> [18,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [19,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [20,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [21,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [22,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [23,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [24,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [25,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [26,]  0.19828206 -0.1178444973     0     0 -0.136916523  0.334240481     0
#> [27,]  0.32113141  0.3077128857     0     0  0.326712681 -0.003580816     0
#> [28,]  0.09923998  0.1447870508     0     0  0.163271094 -0.053067255     0
#> [29,]  0.00000000  0.0000000000     0     0  0.000000000  0.000000000     0
#> [30,] -0.29372140 -0.0390219801     0     0 -0.034834582 -0.262510347     0
#> [31,] -0.53156251 -0.2570202316     0     0 -0.132935417 -0.229752531     0
#>             [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35]
#>  [1,]  0.00000000     0     0     0     0     0     0     0
#>  [2,]  0.00000000     0     0     0     0     0     0     0
#>  [3,]  0.00000000     0     0     0     0     0     0     0
#>  [4,] -0.67754501     0     0     0     0     0     0     0
#>  [5,]  0.00000000     0     0     0     0     0     0     0
#>  [6,]  0.00000000     0     0     0     0     0     0     0
#>  [7,]  0.00000000     0     0     0     0     0     0     0
#>  [8,] -0.10652986     0     0     0     0     0     0     0
#>  [9,]  0.00000000     0     0     0     0     0     0     0
#> [10,] -0.95443138     0     0     0     0     0     0     0
#> [11,] -0.02968719     0     0     0     0     0     0     0
#> [12,] -0.01091912     0     0     0     0     0     0     0
#> [13,]  0.00000000     0     0     0     0     0     0     0
#> [14,]  1.83352754     0     0     0     0     0     0     0
#> [15,]  0.00000000     0     0     0     0     0     0     0
#> [16,]  0.00000000     0     0     0     0     0     0     0
#> [17,]  0.13600962     0     0     0     0     0     0     0
#> [18,]  0.00000000     0     0     0     0     0     0     0
#> [19,]  0.00000000     0     0     0     0     0     0     0
#> [20,]  0.00000000     0     0     0     0     0     0     0
#> [21,]  0.00000000     0     0     0     0     0     0     0
#> [22,]  0.00000000     0     0     0     0     0     0     0
#> [23,]  0.00000000     0     0     0     0     0     0     0
#> [24,]  0.00000000     0     0     0     0     0     0     0
#> [25,]  0.00000000     0     0     0     0     0     0     0
#> [26,]  0.11873631     0     0     0     0     0     0     0
#> [27,] -0.29854554     0     0     0     0     0     0     0
#> [28,] -0.16982293     0     0     0     0     0     0     0
#> [29,]  0.00000000     0     0     0     0     0     0     0
#> [30,]  0.04448462     0     0     0     0     0     0     0
#> [31,] -0.24262447     0     0     0     0     0     0     0
#> 
#> $a[[2]]
#>                [,1]        [,2] [,3] [,4] [,5]        [,6]        [,7]
#>  [1,] -2.069938e-16  0.11887950    0    0    0  0.31134327  0.08831427
#>  [2,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#>  [3,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#>  [4,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#>  [5,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#>  [6,]  1.244624e-18  0.03011803    0    0    0  0.01520995  0.03061026
#>  [7,] -1.694090e-17  0.02218316    0    0    0  0.07599678  0.02054396
#>  [8,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#>  [9,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [10,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [11,]  4.379736e-16 -0.83379237    0    0    0 -0.53399082 -0.76737787
#> [12,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [13,] -3.384262e-16 -0.66804661    0    0    0  1.77502235 -0.69715094
#> [14,] -3.840768e-16 -1.64603035    0    0    0  2.02442743 -1.67945354
#> [15,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [16,]  8.327394e-17  0.16282132    0    0    0 -0.72505172  0.16510550
#> [17,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [18,]  3.768199e-17 -0.55565099    0    0    0 -0.26949228 -0.55402428
#> [19,]  1.041734e-16  0.18128062    0    0    0 -0.26289350  0.19501920
#> [20,] -1.543926e-17 -0.03689903    0    0    0  0.01714258 -0.03931062
#> [21,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [22,]  8.762173e-17  0.13193435    0    0    0 -0.24088065  0.14314289
#> [23,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [24,] -1.076083e-17 -0.01082443    0    0    0  0.03319490 -0.01213645
#> [25,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [26,] -3.402840e-16 -0.76092725    0    0    0  0.80110245 -0.80688633
#> [27,]  1.480570e-17  0.03772523    0    0    0 -0.03690709  0.03969313
#> [28,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [29,]  1.351663e-16  0.31038110    0    0    0 -1.10403943  0.31534974
#> [30,]  0.000000e+00  0.00000000    0    0    0  0.00000000  0.00000000
#> [31,]  9.855340e-16  0.05168071    0    0    0  0.17662122  0.22565930
#>               [,8] [,9] [,10] [,11]         [,12]         [,13]        [,14]
#>  [1,]  0.012373916    0     0     0  0.3969660072 -0.4139902775 -0.423381487
#>  [2,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#>  [3,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#>  [4,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#>  [5,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#>  [6,] -0.002466459    0     0     0  0.0412014174 -0.0426278074 -0.042661924
#>  [7,] -0.003482733    0     0     0  0.0906203988 -0.0934790031 -0.094821934
#>  [8,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#>  [9,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [10,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [11,] -0.017207842    0     0     0 -1.2474071558  1.3036022007  1.319668017
#> [12,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [13,] -0.053963221    0     0     0  1.0500714944 -1.0690412455 -1.103360887
#> [14,] -0.032969658    0     0     0  0.4028477876 -0.3901093513 -0.432637460
#> [15,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [16,]  0.036656521    0     0     0 -0.5276458575  0.5379778625  0.549970836
#> [17,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [18,]  0.032457492    0     0     0 -0.7500388266  0.7780011736  0.780255883
#> [19,] -0.005530480    0     0     0 -0.0817202043  0.0829942459  0.089968478
#> [20,]  0.002913839    0     0     0 -0.0171806709  0.0178053591  0.017001697
#> [21,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [22,] -0.002380246    0     0     0 -0.1055604611  0.1078134659  0.113843212
#> [23,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [24,] -0.000176134    0     0     0  0.0211551229 -0.0217107843 -0.022474898
#> [25,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [26,]  0.028257303    0     0     0  0.0616675253 -0.0581582203 -0.080885936
#> [27,] -0.001214719    0     0     0 -0.0004383383  0.0001463062  0.001178449
#> [28,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [29,]  0.052096178    0     0     0 -0.7473413842  0.7609069556  0.779654897
#> [30,]  0.000000000    0     0     0  0.0000000000  0.0000000000  0.000000000
#> [31,] -0.213384141    0     0     0  0.2066754563 -0.1818415968 -0.155177573
#>       [,15]         [,16]       [,17]        [,18]        [,19]       [,20]
#>  [1,]     0 -0.4073465243  0.30152480  0.411532521 -0.491909571 -0.09602558
#>  [2,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#>  [3,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#>  [4,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#>  [5,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#>  [6,]     0 -0.0416790201  0.01306685  0.040811194 -0.074622963 -0.02844075
#>  [7,]     0 -0.0915240309  0.07182891  0.090554942 -0.114891049 -0.01835652
#>  [8,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#>  [9,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [10,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [11,]     0  1.2812516373 -0.49701368 -1.284158234  2.061806967  0.76371316
#> [12,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [13,]     0 -1.0453795974  1.72162034  1.036959278 -0.365271714  0.70355905
#> [14,]     0 -0.3800847490  1.99875874  0.383617411  1.321303736  1.64815165
#> [15,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [16,]     0  0.5254375797 -0.69549148 -0.516922194  0.379117635 -0.17971231
#> [17,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [18,]     0  0.7615937443 -0.23427659 -0.749610628  1.350780822  0.52238965
#> [19,]     0  0.0818195906 -0.26156704 -0.085175790 -0.121084340 -0.18531639
#> [20,]     0  0.0172804758  0.01858522 -0.016146733  0.059330624  0.03645064
#> [21,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [22,]     0  0.1059723576 -0.23782245 -0.108047332 -0.041840810 -0.13677291
#> [23,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [24,]     0 -0.0212975288  0.03239929  0.021391278 -0.009022531  0.01175295
#> [25,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [26,]     0 -0.0589013621  0.80678879  0.073678349  0.784585164  0.76518263
#> [27,]     0  0.0002197741 -0.03722275 -0.000886261 -0.041179794 -0.03776789
#> [28,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [29,]     0  0.7431275252 -1.06203436 -0.731618756  0.452432649 -0.33375832
#> [30,]     0  0.0000000000  0.00000000  0.000000000  0.000000000  0.00000000
#> [31,]     0 -0.1631175660  0.10453730  0.096981737 -0.519251473 -0.08720795
#>             [,21]       [,22] [,23] [,24]        [,25]        [,26] [,27]
#>  [1,] -0.30463740  0.14074702     0     0  0.112672582 -0.412009890     0
#>  [2,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#>  [3,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#>  [4,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#>  [5,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#>  [6,] -0.01270828  0.02963187     0     0  0.029308163 -0.043183732     0
#>  [7,] -0.07214563  0.02181480     0     0  0.019375433 -0.092880677     0
#>  [8,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#>  [9,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [10,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [11,]  0.49513808 -0.89402870     0     0 -0.824738589  1.316413545     0
#> [12,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [13,] -1.74418104 -0.70292588     0     0 -0.732761529 -1.024360644     0
#> [14,] -2.03644172 -1.69858453     0     0 -1.717603361 -0.316188025     0
#> [15,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [16,]  0.70233862  0.18469887     0     0  0.191687932  0.522688376     0
#> [17,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [18,]  0.22857784 -0.55732492     0     0 -0.543575008  0.788311650     0
#> [19,]  0.26688647  0.17857844     0     0  0.188541250  0.074137111     0
#> [20,] -0.01940989 -0.03563193     0     0 -0.036969592  0.019050472     0
#> [21,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [22,]  0.24210853  0.13028473     0     0  0.138981753  0.100600102     0
#> [23,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [24,] -0.03286927 -0.01071293     0     0 -0.011866095 -0.020910294     0
#> [25,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [26,] -0.82618888 -0.75181563     0     0 -0.781616321 -0.025709115     0
#> [27,]  0.03813879  0.03748860     0     0  0.038705416 -0.001416344     0
#> [28,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [29,]  1.07344039  0.34317686     0     0  0.353846908  0.735987701     0
#> [30,]  0.00000000  0.00000000     0     0  0.000000000  0.000000000     0
#> [31,] -0.08892461 -0.12036648     0     0  0.005162514 -0.182364974     0
#>             [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35]
#>  [1,] -0.12062003     0     0     0     0     0     0     0
#>  [2,]  0.00000000     0     0     0     0     0     0     0
#>  [3,]  0.00000000     0     0     0     0     0     0     0
#>  [4,]  0.00000000     0     0     0     0     0     0     0
#>  [5,]  0.00000000     0     0     0     0     0     0     0
#>  [6,] -0.02929134     0     0     0     0     0     0     0
#>  [7,] -0.02013693     0     0     0     0     0     0     0
#>  [8,]  0.00000000     0     0     0     0     0     0     0
#>  [9,]  0.00000000     0     0     0     0     0     0     0
#> [10,]  0.00000000     0     0     0     0     0     0     0
#> [11,]  0.84112821     0     0     0     0     0     0     0
#> [12,]  0.00000000     0     0     0     0     0     0     0
#> [13,]  0.71669343     0     0     0     0     0     0     0
#> [14,]  1.69912533     0     0     0     0     0     0     0
#> [15,]  0.00000000     0     0     0     0     0     0     0
#> [16,] -0.18708314     0     0     0     0     0     0     0
#> [17,]  0.00000000     0     0     0     0     0     0     0
#> [18,]  0.54537038     0     0     0     0     0     0     0
#> [19,] -0.18424184     0     0     0     0     0     0     0
#> [20,]  0.03637921     0     0     0     0     0     0     0
#> [21,]  0.00000000     0     0     0     0     0     0     0
#> [22,] -0.13532590     0     0     0     0     0     0     0
#> [23,]  0.00000000     0     0     0     0     0     0     0
#> [24,]  0.01141029     0     0     0     0     0     0     0
#> [25,]  0.00000000     0     0     0     0     0     0     0
#> [26,]  0.76766066     0     0     0     0     0     0     0
#> [27,] -0.03809243     0     0     0     0     0     0     0
#> [28,]  0.00000000     0     0     0     0     0     0     0
#> [29,] -0.34652577     0     0     0     0     0     0     0
#> [30,]  0.00000000     0     0     0     0     0     0     0
#> [31,]  0.02908095     0     0     0     0     0     0     0
#> 
#> 
#> $BIC
#> [1] 7370.85
#>