Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2725.995
#> 
#> $assign
#>  [1] 1 1 1 2 1 1 1 2 1 1 1 1 2 2 2 1 1 2 1 1 2 1 2 1 1 2 2 2 2 2 2 2 2 2 2 1 2 1
#> [39] 2 2 1 1 1 2 2 1 2 1 2 1
#> 
#> $a
#> $a[[1]]
#>                [,1]        [,2] [,3]         [,4]        [,5] [,6]         [,7]
#>  [1,] -5.197503e-19  0.15682156    0  0.001893983 -0.16419759    0 -0.157140002
#>  [2,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#>  [3,]  2.683534e-19 -0.01184456    0 -0.032671398  0.04599767    0  0.042647288
#>  [4,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#>  [5,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#>  [6,]  3.742409e-18  0.46404736    0  0.092058609 -0.53132130    0 -0.541665269
#>  [7,] -3.990251e-18  0.44282088    0 -0.063977728 -0.41446268    0 -0.380745044
#>  [8,]  3.087539e-18 -0.04899616    0 -0.894382266  0.94934025    0  0.891875660
#>  [9,]  3.359672e-19  0.88129895    0 -0.532614575 -0.36589961    0 -0.369547349
#> [10,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#> [11,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#> [12,] -5.740719e-19 -1.24247590    0 -0.248079352  1.49448751    0  1.460020566
#> [13,]  1.306844e-18  0.36657521    0  0.112686059 -0.47117875    0 -0.466895404
#> [14,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#> [15,] -1.208618e-19 -0.03355874    0  0.009396392  0.02379988    0  0.024210432
#> [16,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#> [17,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#> [18,] -3.266682e-19  0.10662527    0  0.018534663 -0.12835013    0 -0.123031967
#> [19,]  8.385190e-19 -0.11918981    0  0.110486347  0.01836662    0  0.014598991
#> [20,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#> [21,] -4.386694e-18 -0.27327212    0 -0.047746153  0.29029909    0  0.311001009
#> [22,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#> [23,]  1.428196e-18  0.18662559    0 -0.198967598  0.01710928    0  0.003941631
#> [24,]  7.679149e-19  0.11264756    0 -0.007989746 -0.10028266    0 -0.103092580
#> [25,] -5.271156e-19  0.15228783    0 -0.047714326 -0.11103554    0 -0.106032357
#> [26,] -9.958211e-19 -1.69314591    0  0.226573053  1.48141474    0  1.458610790
#> [27,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#> [28,]  0.000000e+00  0.00000000    0  0.000000000  0.00000000    0  0.000000000
#>       [,8]         [,9] [,10] [,11]       [,12] [,13]       [,14] [,15] [,16]
#>  [1,]    0  0.160237232     0     0  0.16109756     0 -0.15902578     0     0
#>  [2,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#>  [3,]    0 -0.043744998     0     0 -0.04430775     0  0.04394936     0     0
#>  [4,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#>  [5,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#>  [6,]    0  0.547079968     0     0  0.52694041     0 -0.52817857     0     0
#>  [7,]    0  0.390740852     0     0  0.40438349     0 -0.39492209     0     0
#>  [8,]    0 -0.913118163     0     0 -0.91332707     0  0.91165516     0     0
#>  [9,]    0  0.373111463     0     0  0.37195385     0 -0.36291889     0     0
#> [10,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#> [11,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#> [12,]    0 -1.484246629     0     0 -1.46881030     0  1.45945078     0     0
#> [13,]    0  0.473638410     0     0  0.46391754     0 -0.46281124     0     0
#> [14,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#> [15,]    0 -0.024445546     0     0 -0.02381496     0  0.02365132     0     0
#> [16,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#> [17,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#> [18,]    0  0.125443674     0     0  0.12568845     0 -0.12437167     0     0
#> [19,]    0 -0.015245403     0     0 -0.01920801     0  0.01667468     0     0
#> [20,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#> [21,]    0 -0.311823788     0     0 -0.29115198     0  0.29479021     0     0
#> [22,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#> [23,]    0 -0.006245418     0     0 -0.01086716     0  0.01123861     0     0
#> [24,]    0  0.103965055     0     0  0.10005595     0 -0.10006934     0     0
#> [25,]    0  0.108105448     0     0  0.10969664     0 -0.10749381     0     0
#> [26,]    0 -1.480493751     0     0 -1.46615896     0  1.45185797     0     0
#> [27,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#> [28,]    0  0.000000000     0     0  0.00000000     0  0.00000000     0     0
#>       [,17]       [,18]        [,19] [,20]       [,21]        [,22] [,23] [,24]
#>  [1,]     0  0.16230341 -0.001895283     0 -0.16261285 -0.004352635     0     0
#>  [2,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#>  [3,]     0 -0.07456800 -0.031894624     0  0.04643875  0.035413423     0     0
#>  [4,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#>  [5,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#>  [6,]     0  0.62978044  0.084575305     0 -0.53103848 -0.127800229     0     0
#>  [7,]     0  0.33373198 -0.076205287     0 -0.40730181  0.073563090     0     0
#>  [8,]     0 -1.75359262 -0.882598714     0  0.96592235  0.977791910     0     0
#>  [9,]     0 -0.12697242 -0.546202767     0 -0.34734627  0.563586008     0     0
#> [10,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#> [11,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#> [12,]     0 -1.71506127 -0.219955374     0  1.48908604  0.309182400     0     0
#> [13,]     0  0.57778804  0.105002715     0 -0.47096694 -0.139026755     0     0
#> [14,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#> [15,]     0 -0.01551067  0.009920092     0  0.02337251 -0.009040534     0     0
#> [16,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#> [17,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#> [18,]     0  0.14298467  0.015824825     0 -0.12764335 -0.022107047     0     0
#> [19,]     0  0.08767610  0.112731060     0  0.01462268 -0.121946527     0     0
#> [20,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#> [21,]     0 -0.35264338 -0.045263479     0  0.29107936  0.076150864     0     0
#> [22,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#> [23,]     0 -0.19401938 -0.199904277     0  0.02215248  0.209538616     0     0
#> [24,]     0  0.09573557 -0.009651526     0 -0.09951562  0.003001067     0     0
#> [25,]     0  0.06370428 -0.050964118     0 -0.10845844  0.050587757     0     0
#> [26,]     0 -1.26584850  0.259598832     0  1.46245654 -0.201528372     0     0
#> [27,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#> [28,]     0  0.00000000  0.000000000     0  0.00000000  0.000000000     0     0
#>       [,25] [,26]         [,27]        [,28]        [,29]       [,30]
#>  [1,]     0     0  0.0002282356 -0.001266028  0.001116301 -0.16021020
#>  [2,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#>  [3,]     0     0  0.0310028099 -0.030765041  0.031391933  0.07538986
#>  [4,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#>  [5,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#>  [6,]     0     0 -0.1294160729  0.097587344 -0.118146464 -0.59378803
#>  [7,]     0     0  0.0911146345 -0.079219609  0.089556605 -0.34229874
#>  [8,]     0     0  0.8795380340 -0.858621655  0.886157101  1.76132421
#>  [9,]     0     0  0.5239567627 -0.524387862  0.535613828  0.15196949
#> [10,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#> [11,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#> [12,]     0     0  0.2716979870 -0.231862195  0.257333180  1.67065151
#> [13,]     0     0 -0.1292759632  0.110518672 -0.123223017 -0.55808654
#> [14,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#> [15,]     0     0 -0.0080769061  0.009035423 -0.008677236  0.01387034
#> [16,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#> [17,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#> [18,]     0     0 -0.0174908981  0.015987828 -0.016881016 -0.14126571
#> [19,]     0     0 -0.1168749377  0.111188981 -0.117058534 -0.08586160
#> [20,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#> [21,]     0     0  0.0873366652 -0.058341964  0.077033220  0.31949366
#> [22,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#> [23,]     0     0  0.1870697402 -0.190347519  0.192001982  0.20603604
#> [24,]     0     0  0.0003802607 -0.006253844  0.002948728 -0.08803339
#> [25,]     0     0  0.0504821637 -0.049460302  0.051418085 -0.06222216
#> [26,]     0     0 -0.1986514606  0.232064488 -0.219848676  1.20902546
#> [27,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#> [28,]     0     0  0.0000000000  0.000000000  0.000000000  0.00000000
#>              [,31]        [,32] [,33] [,34] [,35]
#>  [1,] -0.152462556 -0.001202078     0     0     0
#>  [2,]  0.000000000  0.000000000     0     0     0
#>  [3,]  0.043067504  0.031403785     0     0     0
#>  [4,]  0.000000000  0.000000000     0     0     0
#>  [5,]  0.000000000  0.000000000     0     0     0
#>  [6,] -0.544343461 -0.116022225     0     0     0
#>  [7,] -0.358494389  0.078290735     0     0     0
#>  [8,]  0.919943718  0.877503100     0     0     0
#>  [9,] -0.328944793  0.519032006     0     0     0
#> [10,]  0.000000000  0.000000000     0     0     0
#> [11,]  0.000000000  0.000000000     0     0     0
#> [12,]  1.444058390  0.265336051     0     0     0
#> [13,] -0.466601707 -0.123330605     0     0     0
#> [14,]  0.000000000  0.000000000     0     0     0
#> [15,]  0.023376847 -0.008368660     0     0     0
#> [16,]  0.000000000  0.000000000     0     0     0
#> [17,]  0.000000000  0.000000000     0     0     0
#> [18,] -0.120625750 -0.018155605     0     0     0
#> [19,]  0.005357364 -0.112946337     0     0     0
#> [20,]  0.000000000  0.000000000     0     0     0
#> [21,]  0.317673311  0.072711615     0     0     0
#> [22,]  0.000000000  0.000000000     0     0     0
#> [23,]  0.012746895  0.189147106     0     0     0
#> [24,] -0.102151408  0.002797846     0     0     0
#> [25,] -0.099399691  0.048378804     0     0     0
#> [26,]  1.414190305 -0.200645463     0     0     0
#> [27,]  0.000000000  0.000000000     0     0     0
#> [28,]  0.000000000  0.000000000     0     0     0
#>  [ reached getOption("max.print") -- omitted 3 rows ]
#> 
#> $a[[2]]
#>                [,1]        [,2] [,3]        [,4]        [,5] [,6]        [,7]
#>  [1,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#>  [2,]  1.326459e-18 -0.02793304    0  0.01564076  0.01426684    0  0.01266360
#>  [3,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#>  [4,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#>  [5,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#>  [6,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#>  [7,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#>  [8,] -1.645108e-16 -1.27434542    0  0.64568011  0.63318848    0  0.61178540
#>  [9,]  4.048543e-17 -0.60661037    0  0.48002319  0.15758216    0  0.13806610
#> [10,]  6.908557e-17  0.44299254    0 -0.15525401 -0.29261542    0 -0.27928666
#> [11,] -8.566158e-18 -0.23727271    0  0.13988617  0.10506570    0  0.09771761
#> [12,] -8.270357e-16 -2.04536447    0  0.40853258  1.48467635    0  1.51956969
#> [13,] -6.539119e-18  0.74938650    0 -0.54895996 -0.22587714    0 -0.20769795
#> [14,] -5.339619e-18 -0.10324954    0  0.04614426  0.06153311    0  0.05683431
#> [15,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#> [16,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#> [17,] -9.791336e-17 -0.17997049    0 -0.09827730  0.27040974    0  0.26261497
#> [18,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#> [19,]  1.340743e-16 -0.67357197    0  0.85712743 -0.15071814    0 -0.15380977
#> [20,] -2.572594e-16  0.43623031    0 -0.57713760  0.05133575    0  0.09638498
#> [21,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#> [22,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#> [23,] -6.075189e-16  0.90303274    0 -2.03171873  1.01192704    0  1.01537492
#> [24,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#> [25,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#> [26,] -9.503590e-16 -1.85908507    0 -0.55464032  2.30323010    0  2.26678238
#> [27,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#> [28,]  0.000000e+00  0.00000000    0  0.00000000  0.00000000    0  0.00000000
#>       [,8]        [,9] [,10] [,11]       [,12] [,13]       [,14] [,15] [,16]
#>  [1,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#>  [2,]    0 -0.01226475     0     0 -0.01410939     0  0.01095663     0     0
#>  [3,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#>  [4,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#>  [5,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#>  [6,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#>  [7,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#>  [8,]    0 -0.61049223     0     0 -0.66965637     0  0.61143181     0     0
#>  [9,]    0 -0.13320624     0     0 -0.16228418     0  0.11218473     0     0
#> [10,]    0  0.27748732     0     0  0.30262019     0 -0.27596879     0     0
#> [11,]    0 -0.09626436     0     0 -0.10856323     0  0.09149700     0     0
#> [12,]    0 -1.54412235     0     0 -1.61765359     0  1.66203735     0     0
#> [13,]    0  0.20390219     0     0  0.23777133     0 -0.18726348     0     0
#> [14,]    0 -0.05583606     0     0 -0.06228678     0  0.05309963     0     0
#> [15,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#> [16,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#> [17,]    0 -0.26233364     0     0 -0.27737485     0  0.26973427     0     0
#> [18,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#> [19,]    0  0.15573669     0     0  0.14022270     0 -0.18020725     0     0
#> [20,]    0 -0.11143896     0     0 -0.07982803     0  0.17545823     0     0
#> [21,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#> [22,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#> [23,]    0 -1.02432248     0     0 -1.02363479     0  1.11435903     0     0
#> [24,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#> [25,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#> [26,]    0 -2.27440651     0     0 -2.39689151     0  2.36843253     0     0
#> [27,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#> [28,]    0  0.00000000     0     0  0.00000000     0  0.00000000     0     0
#>       [,17]         [,18]       [,19] [,20]       [,21]       [,22] [,23] [,24]
#>  [1,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#>  [2,]     0 -0.0004205921  0.01229119     0  0.01383464 -0.01370520     0     0
#>  [3,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#>  [4,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#>  [5,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#>  [6,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#>  [7,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#>  [8,]     0  0.0087626394  0.63872101     0  0.64498585 -0.65092212     0     0
#>  [9,]     0  0.2955712894  0.44578297     0  0.15266369 -0.45250239     0     0
#> [10,]     0  0.1422731055 -0.13845612     0 -0.29545644  0.15273547     0     0
#> [11,]     0  0.0280005583  0.12866962     0  0.10475245 -0.13353136     0     0
#> [12,]     0 -0.9407084893  0.60807242     0  1.56203131 -0.58616107     0     0
#> [13,]     0 -0.3013332854 -0.52442545     0 -0.22451014  0.52865658     0     0
#> [14,]     0 -0.0194782499  0.03751341     0  0.06098739 -0.04203336     0     0
#> [15,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#> [16,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#> [17,]     0 -0.3629920935 -0.10706135     0  0.27508918  0.09012076     0     0
#> [18,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#> [19,]     0  0.9821585593  0.86141887     0 -0.15540383 -0.83159239     0     0
#> [20,]     0 -0.5491884874 -0.45956151     0  0.07954888  0.48559377     0     0
#> [21,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#> [22,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#> [23,]     0 -2.9496425880 -2.01392605     0  1.04450810  1.92631980     0     0
#> [24,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#> [25,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#> [26,]     0 -2.7680478196 -0.53610760     0  2.36273368  0.43319415     0     0
#> [27,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#> [28,]     0  0.0000000000  0.00000000     0  0.00000000  0.00000000     0     0
#>       [,25] [,26]       [,27]       [,28]       [,29]        [,30]       [,31]
#>  [1,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#>  [2,]     0     0 -0.01292938  0.01561769 -0.01341353  0.001641728  0.01445838
#>  [3,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#>  [4,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#>  [5,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#>  [6,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#>  [7,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#>  [8,]     0     0 -0.63557304  0.65542036 -0.65880688  0.018643542  0.64970737
#>  [9,]     0     0 -0.43827718  0.47410402 -0.45179692 -0.275760260  0.16040806
#> [10,]     0     0  0.14670315 -0.16153987  0.15337069 -0.154224335 -0.29921506
#> [11,]     0     0 -0.12905636  0.14017954 -0.13356135 -0.020899955  0.10729007
#> [12,]     0     0 -0.59941430  0.46742819 -0.62931406  0.941251240  1.53376792
#> [13,]     0     0  0.51412782 -0.54483094  0.53035824  0.280479346 -0.23108155
#> [14,]     0     0 -0.03984841  0.04685482 -0.04149798  0.023216328  0.06265357
#> [15,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#> [16,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#> [17,]     0     0  0.09010516 -0.08739330  0.09018297  0.367444957  0.27645931
#> [18,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#> [19,]     0     0 -0.81581548  0.83572816 -0.83703176 -0.966712946 -0.15268180
#> [20,]     0     0  0.45940844 -0.55614720  0.47162273  0.515993487  0.05830040
#> [21,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#> [22,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#> [23,]     0     0  1.88723717 -1.95575945  1.93038218  2.920161200  1.03447153
#> [24,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#> [25,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#> [26,]     0     0  0.42663774 -0.46083509  0.41784884  2.795307836  2.36060498
#> [27,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#> [28,]     0     0  0.00000000  0.00000000  0.00000000  0.000000000  0.00000000
#>             [,32] [,33] [,34] [,35]
#>  [1,]  0.00000000     0     0     0
#>  [2,] -0.01323918     0     0     0
#>  [3,]  0.00000000     0     0     0
#>  [4,]  0.00000000     0     0     0
#>  [5,]  0.00000000     0     0     0
#>  [6,]  0.00000000     0     0     0
#>  [7,]  0.00000000     0     0     0
#>  [8,] -0.62627485     0     0     0
#>  [9,] -0.43842787     0     0     0
#> [10,]  0.14596872     0     0     0
#> [11,] -0.12888166     0     0     0
#> [12,] -0.55195082     0     0     0
#> [13,]  0.51154859     0     0     0
#> [14,] -0.04041561     0     0     0
#> [15,]  0.00000000     0     0     0
#> [16,]  0.00000000     0     0     0
#> [17,]  0.08968702     0     0     0
#> [18,]  0.00000000     0     0     0
#> [19,] -0.80833431     0     0     0
#> [20,]  0.47398284     0     0     0
#> [21,]  0.00000000     0     0     0
#> [22,]  0.00000000     0     0     0
#> [23,]  1.87883246     0     0     0
#> [24,]  0.00000000     0     0     0
#> [25,]  0.00000000     0     0     0
#> [26,]  0.44113907     0     0     0
#> [27,]  0.00000000     0     0     0
#> [28,]  0.00000000     0     0     0
#>  [ reached getOption("max.print") -- omitted 3 rows ]
#> 
#> 
#> $BIC
#> [1] 7634.899
#>