Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2725.088
#> 
#> $assign
#>  [1] 2 1 1 1 1 2 1 1 2 1 2 2 1 1 2 1 1 1 2 1 1 1 1 1 2 1 2 2 1 1 2 1 2 1 1 2 1 1
#> [39] 2 1 1 2 2 2 2 2 1 1 2 1
#> 
#> $a
#> $a[[1]]
#>              [,1]        [,2] [,3] [,4]        [,5]        [,6]       [,7] [,8]
#>  [1,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#>  [2,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#>  [3,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#>  [4,] -0.12334736  0.07342008    0    0 -0.12307624 -0.04681977  0.1977094    0
#>  [5,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#>  [6,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#>  [7,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#>  [8,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#>  [9,] -0.40771117  0.07428982    0    0 -0.40650865 -0.31967035  0.4862642    0
#> [10,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [11,] -0.19521186  0.34363084    0    0 -0.17842408  0.11934466  0.5216584    0
#> [12,]  0.28759962  0.12506731    0    0  0.32227400  0.33669634 -0.2033441    0
#> [13,]  0.12254169 -0.16697169    0    0  0.17508790 -0.13751683 -0.3435491    0
#> [14,] -0.27748735  0.62139739    0    0 -0.27898497  0.34370483  0.8991929    0
#> [15,]  0.23321108  1.18607973    0    0  0.21296317  1.41629666  0.9597278    0
#> [16,]  0.05270724 -0.57404629    0    0  0.04363622 -0.49408925 -0.6133011    0
#> [17,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [18,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [19,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [20,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [21,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [22,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [23,] -0.10620945 -0.30359957    0    0 -0.11562629 -0.38153408 -0.1835418    0
#> [24,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [25,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [26,] -0.15885231 -0.39128692    0    0 -0.16490005 -0.52338247 -0.2204718    0
#> [27,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [28,]  0.37279099  0.02626470    0    0  0.34254509  0.43514047 -0.3214884    0
#> [29,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [30,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.0000000    0
#> [31,] -0.68379460  0.16223597    0    0 -1.04542016  0.13570112  1.2240768    0
#>       [,9] [,10] [,11]       [,12] [,13] [,14] [,15] [,16]       [,17] [,18]
#>  [1,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#>  [2,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#>  [3,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#>  [4,]    0     0     0 -0.12121637     0     0     0     0 -0.12369100     0
#>  [5,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#>  [6,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#>  [7,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#>  [8,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#>  [9,]    0     0     0 -0.39967103     0     0     0     0 -0.40408219     0
#> [10,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [11,]    0     0     0 -0.20140932     0     0     0     0 -0.21374273     0
#> [12,]    0     0     0  0.26330144     0     0     0     0  0.25533065     0
#> [13,]    0     0     0  0.09516725     0     0     0     0  0.08928686     0
#> [14,]    0     0     0 -0.27476656     0     0     0     0 -0.29029020     0
#> [15,]    0     0     0  0.22983755     0     0     0     0  0.20830304     0
#> [16,]    0     0     0  0.05988295     0     0     0     0  0.07505728     0
#> [17,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [18,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [19,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [20,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [21,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [22,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [23,]    0     0     0 -0.09714485     0     0     0     0 -0.08901982     0
#> [24,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [25,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [26,]    0     0     0 -0.14966305     0     0     0     0 -0.14053297     0
#> [27,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [28,]    0     0     0  0.37908978     0     0     0     0  0.38676303     0
#> [29,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [30,]    0     0     0  0.00000000     0     0     0     0  0.00000000     0
#> [31,]    0     0     0 -0.49225748     0     0     0     0 -0.42837502     0
#>             [,19] [,20] [,21]       [,22]       [,23] [,24]       [,25] [,26]
#>  [1,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#>  [2,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#>  [3,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#>  [4,] -0.12336018     0     0 -0.07643639  0.12090340     0  0.06741848     0
#>  [5,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#>  [6,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#>  [7,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#>  [8,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#>  [9,] -0.40613563     0     0 -0.08627196  0.39791514     0  0.05816198     0
#> [10,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [11,] -0.20186022     0     0 -0.32073829  0.21676667     0  0.33876176     0
#> [12,]  0.27528800     0     0 -0.06117524 -0.22950518     0  0.15313571     0
#> [13,]  0.10951297     0     0  0.24715203 -0.04856004     0 -0.12868118     0
#> [14,] -0.28156132     0     0 -0.62463182  0.27488383     0  0.59711070     0
#> [15,]  0.22509472     0     0 -1.19031620 -0.23893809     0  1.15828142     0
#> [16,]  0.06056046     0     0  0.55406168 -0.07089375     0 -0.56524060     0
#> [17,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [18,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [19,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [20,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [21,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [22,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [23,] -0.09996302     0     0  0.28127205  0.08615142     0 -0.30661596     0
#> [24,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [25,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [26,] -0.15227219     0     0  0.37076713  0.14072002     0 -0.39249720     0
#> [27,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [28,]  0.37812054     0     0 -0.05728292 -0.40278349     0  0.02228320     0
#> [29,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [30,]  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000     0
#> [31,] -0.58520724     0     0 -0.72268887  0.16695975     0 -0.07460025     0
#>       [,27] [,28]       [,29]       [,30]       [,31]       [,32]       [,33]
#>  [1,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [2,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [3,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [4,]     0     0  0.07193758  0.12392923  0.11839362  0.07391529  0.07299471
#>  [5,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [6,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [7,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [8,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [9,]     0     0  0.06571251  0.40943409  0.39063738  0.07754135  0.06655942
#> [10,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [11,]     0     0  0.33276885  0.22310395  0.21274544  0.32097730  0.37526201
#> [12,]     0     0  0.10617970 -0.23154040 -0.22261099  0.07564354  0.18796997
#> [13,]     0     0 -0.20986884 -0.03942566 -0.04133697 -0.23328033 -0.09549649
#> [14,]     0     0  0.62905143  0.27730570  0.26637860  0.61975567  0.63576542
#> [15,]     0     0  1.23173937 -0.26031420 -0.24357203  1.19439557  1.21209877
#> [16,]     0     0 -0.57727074 -0.06942946 -0.06732643 -0.55645374 -0.60833209
#> [17,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [18,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [19,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [20,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [21,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [22,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [23,]     0     0 -0.30459247  0.08984896  0.08539243 -0.28657981 -0.33236024
#> [24,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [25,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [26,]     0     0 -0.39795259  0.14743292  0.13984949 -0.37676322 -0.42039685
#> [27,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [28,]     0     0  0.05945485 -0.42038025 -0.39898880  0.05910958 -0.00417980
#> [29,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [30,]     0     0  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [31,]     0     0  0.44092040  0.10794695  0.12578806  0.62809919 -0.36440966
#>             [,34]       [,35]
#>  [1,]  0.00000000  0.00000000
#>  [2,]  0.00000000  0.00000000
#>  [3,]  0.00000000  0.00000000
#>  [4,] -0.07130418 -0.07024380
#>  [5,]  0.00000000  0.00000000
#>  [6,]  0.00000000  0.00000000
#>  [7,]  0.00000000  0.00000000
#>  [8,]  0.00000000  0.00000000
#>  [9,] -0.06362069 -0.06369063
#> [10,]  0.00000000  0.00000000
#> [11,] -0.33002687 -0.31589001
#> [12,] -0.10394085 -0.08381667
#> [13,]  0.21469097  0.23559076
#> [14,] -0.62775396 -0.61626403
#> [15,] -1.23471711 -1.21647637
#> [16,]  0.57596455  0.55946749
#> [17,]  0.00000000  0.00000000
#> [18,]  0.00000000  0.00000000
#> [19,]  0.00000000  0.00000000
#> [20,]  0.00000000  0.00000000
#> [21,]  0.00000000  0.00000000
#> [22,]  0.00000000  0.00000000
#> [23,]  0.30406534  0.29249712
#> [24,]  0.00000000  0.00000000
#> [25,]  0.00000000  0.00000000
#> [26,]  0.39800904  0.38574812
#> [27,]  0.00000000  0.00000000
#> [28,] -0.06436744 -0.07548540
#> [29,]  0.00000000  0.00000000
#> [30,]  0.00000000  0.00000000
#> [31,] -0.47506705 -0.63978563
#> 
#> $a[[2]]
#>              [,1]        [,2] [,3] [,4]        [,5]        [,6]         [,7]
#>  [1,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#>  [2,] -0.67502770  0.25107910    0    0 -0.67525970 -0.41206527  0.911865871
#>  [3,] -0.53962394  0.29595585    0    0 -0.54184405 -0.22430867  0.817662239
#>  [4,]  0.08369737  0.05391094    0    0  0.08077209  0.13662661 -0.029144245
#>  [5,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#>  [6,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#>  [7,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#>  [8,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#>  [9,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [10,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [11,] -0.42090693 -0.03348759    0    0 -0.41222596 -0.43182701  0.371838708
#> [12,]  0.85843926 -0.08398098    0    0  0.86603887  0.87384023 -0.991525061
#> [13,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [14,] -0.70721365 -0.07106121    0    0 -0.69906599 -0.79125552  0.638962087
#> [15,] -0.48506525  0.47679885    0    0 -0.49125935  0.03317842  0.930040437
#> [16,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [17,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [18,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [19,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [20,]  0.13658096  0.47017827    0    0  0.11257140  0.56380782  0.354278188
#> [21,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [22,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [23,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [24,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [25,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [26,]  0.00000000  0.00000000    0    0  0.00000000  0.00000000  0.000000000
#> [27,] -0.19217856  0.02409923    0    0 -0.19213770 -0.17621842  0.219223194
#> [28,]  0.02591741  0.02025960    0    0  0.02504410  0.04703664 -0.006153735
#> [29,]  0.36697021  0.70499377    0    0  0.34445200  1.11850466  0.307372347
#> [30,]  0.23876222  0.44776666    0    0  0.22254896  0.70255478  0.197247774
#> [31,] -0.04153716 -0.02516709    0    0 -0.03277961 -0.01240663 -0.014455644
#>       [,8] [,9] [,10] [,11]       [,12] [,13] [,14] [,15] [,16]       [,17]
#>  [1,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#>  [2,]    0    0     0     0 -0.69430860     0     0     0     0 -0.64859110
#>  [3,]    0    0     0     0 -0.55117401     0     0     0     0 -0.51509395
#>  [4,]    0    0     0     0  0.08755676     0     0     0     0  0.08338509
#>  [5,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#>  [6,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#>  [7,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#>  [8,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#>  [9,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [10,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [11,]    0    0     0     0 -0.43233748     0     0     0     0 -0.41091886
#> [12,]    0    0     0     0  0.91398349     0     0     0     0  0.83393789
#> [13,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [14,]    0    0     0     0 -0.73878409     0     0     0     0 -0.69135721
#> [15,]    0    0     0     0 -0.48636838     0     0     0     0 -0.45548147
#> [16,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [17,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [18,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [19,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [20,]    0    0     0     0  0.13909930     0     0     0     0  0.14904841
#> [21,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [22,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [23,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [24,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [25,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [26,]    0    0     0     0  0.00000000     0     0     0     0  0.00000000
#> [27,]    0    0     0     0 -0.20116664     0     0     0     0 -0.18639516
#> [28,]    0    0     0     0  0.02744646     0     0     0     0  0.02595427
#> [29,]    0    0     0     0  0.40345601     0     0     0     0  0.38233099
#> [30,]    0    0     0     0  0.25893036     0     0     0     0  0.24826104
#> [31,]    0    0     0     0 -0.03060389     0     0     0     0 -0.04089299
#>       [,18]       [,19] [,20] [,21]       [,22]       [,23] [,24]       [,25]
#>  [1,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#>  [2,]     0 -0.68882395     0     0 -0.22483159  0.66848722     0  0.25394139
#>  [3,]     0 -0.54492000     0     0 -0.27744661  0.53556983     0  0.30169681
#>  [4,]     0  0.08573027     0     0 -0.06086732 -0.07793647     0  0.05533316
#>  [5,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#>  [6,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#>  [7,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#>  [8,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#>  [9,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [10,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [11,]     0 -0.42084867     0     0  0.06104755  0.39813709     0 -0.03312002
#> [12,]     0  0.94216034     0     0  0.06068510 -0.88952292     0 -0.06492871
#> [13,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [14,]     0 -0.73649403     0     0  0.10888216  0.69194306     0 -0.07873629
#> [15,]     0 -0.47585632     0     0 -0.46523485  0.48297650     0  0.48994968
#> [16,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [17,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [18,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [19,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [20,]     0  0.11621057     0     0 -0.50599379 -0.08448952     0  0.47364214
#> [21,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [22,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [23,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [24,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [25,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [26,]     0  0.00000000     0     0  0.00000000  0.00000000     0  0.00000000
#> [27,]     0 -0.20277678     0     0 -0.01658506  0.19283786     0  0.02215672
#> [28,]     0  0.02722000     0     0 -0.02236354 -0.02446242     0  0.02095409
#> [29,]     0  0.40543624     0     0 -0.74724463 -0.33702707     0  0.72858488
#> [30,]     0  0.25555857     0     0 -0.47733070 -0.21325537     0  0.46065777
#> [31,]     0 -0.01149365     0     0  0.03821680  0.01351947     0 -0.01738679
#>       [,26] [,27] [,28]       [,29]       [,30]       [,31]       [,32]
#>  [1,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#>  [2,]     0     0     0  0.23303789  0.69112813  0.64986466  0.20925267
#>  [3,]     0     0     0  0.28958061  0.55083943  0.52388648  0.26191068
#>  [4,]     0     0     0  0.06236344 -0.08487231 -0.07721552  0.05811241
#>  [5,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#>  [6,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#>  [7,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#>  [8,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#>  [9,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [10,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [11,]     0     0     0 -0.05673016  0.42461353  0.39877468 -0.05650499
#> [12,]     0     0     0 -0.03063862 -0.90793823 -0.80968539 -0.02751073
#> [13,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [14,]     0     0     0 -0.11712388  0.72735538  0.66609516 -0.11226060
#> [15,]     0     0     0  0.48913563  0.49086851  0.48075408  0.44496678
#> [16,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [17,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [18,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [19,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [20,]     0     0     0  0.50682331 -0.12275680 -0.11376343  0.47029762
#> [21,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [22,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [23,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [24,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [25,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [26,]     0     0     0  0.00000000  0.00000000  0.00000000  0.00000000
#> [27,]     0     0     0  0.01418745  0.19941635  0.18240078  0.01214304
#> [28,]     0     0     0  0.02327094 -0.02656299 -0.02369402  0.02157998
#> [29,]     0     0     0  0.78192046 -0.38091156 -0.31685495  0.72054374
#> [30,]     0     0     0  0.49536325 -0.24401247 -0.20752223  0.45725732
#> [31,]     0     0     0 -0.02296835  0.02725229  0.04192662 -0.02441547
#>              [,33]       [,34]       [,35]
#>  [1,]  0.000000000  0.00000000  0.00000000
#>  [2,]  0.231056693 -0.25906814 -0.20493523
#>  [3,]  0.285282165 -0.31242102 -0.25699410
#>  [4,]  0.058008268 -0.05690826 -0.06061986
#>  [5,]  0.000000000  0.00000000  0.00000000
#>  [6,]  0.000000000  0.00000000  0.00000000
#>  [7,]  0.000000000  0.00000000  0.00000000
#>  [8,]  0.000000000  0.00000000  0.00000000
#>  [9,]  0.000000000  0.00000000  0.00000000
#> [10,]  0.000000000  0.00000000  0.00000000
#> [11,] -0.043426662  0.02761659  0.07001983
#> [12,] -0.011222146  0.01489659  0.06121096
#> [13,]  0.000000000  0.00000000  0.00000000
#> [14,] -0.107277926  0.09252637  0.11525819
#> [15,]  0.479846539 -0.51524360 -0.43642745
#> [16,]  0.000000000  0.00000000  0.00000000
#> [17,]  0.000000000  0.00000000  0.00000000
#> [18,]  0.000000000  0.00000000  0.00000000
#> [19,]  0.000000000  0.00000000  0.00000000
#> [20,]  0.467631542 -0.46672969 -0.50044318
#> [21,]  0.000000000  0.00000000  0.00000000
#> [22,]  0.000000000  0.00000000  0.00000000
#> [23,]  0.000000000  0.00000000  0.00000000
#> [24,]  0.000000000  0.00000000  0.00000000
#> [25,]  0.000000000  0.00000000  0.00000000
#> [26,]  0.000000000  0.00000000  0.00000000
#> [27,]  0.013241072 -0.01737009 -0.01425419
#> [28,]  0.022028762 -0.02204407 -0.02198804
#> [29,]  0.750078538 -0.76913183 -0.72340841
#> [30,]  0.471520329 -0.48042826 -0.46446855
#> [31,] -0.006803227 -0.00585719  0.04685553
#> 
#> 
#> $BIC
#> [1] 7210.587
#>