Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2674.008
#> 
#> $assign
#>  [1] 1 2 2 2 1 1 1 1 1 1 2 2 1 2 1 1 2 1 1 1 2 1 2 2 2 1 2 2 2 1 1 2 2 1 2 1 2 1
#> [39] 2 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $a
#> $a[[1]]
#>                [,1]          [,2]         [,3] [,4]         [,5] [,6]
#>  [1,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#>  [2,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#>  [3,]  8.712042e-18 -1.245330e-17  0.016366553    0 -0.015199005    0
#>  [4,]  9.473085e-18 -3.216623e-17  0.038682944    0 -0.038302915    0
#>  [5,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#>  [6,] -3.420596e-16 -3.570851e-17  0.129023437    0 -0.160785048    0
#>  [7,] -1.052778e-16  3.571497e-16 -0.493242846    0  0.475397425    0
#>  [8,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#>  [9,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [10,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [11,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [12,] -9.525864e-17 -1.045679e-16  0.209846250    0 -0.208730214    0
#> [13,] -1.247937e-15  5.802819e-16 -0.417374365    0  0.315945204    0
#> [14,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [15,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [16,]  9.509239e-17 -1.249093e-16  0.154157834    0 -0.143477612    0
#> [17,]  4.235239e-18 -6.530121e-18  0.008150846    0 -0.007678497    0
#> [18,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [19,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [20,]  2.873977e-18  1.816280e-16 -0.221719230    0  0.226953759    0
#> [21,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [22,] -3.408327e-17  1.859253e-16 -0.251895781    0  0.246619448    0
#> [23,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [24,]  1.917055e-17 -6.757370e-16  0.978914141    0 -0.961211962    0
#> [25,]  8.691231e-18 -1.577967e-17  0.014772544    0 -0.014903946    0
#> [26,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [27,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [28,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [29,] -2.746185e-18  4.775854e-18 -0.005547871    0  0.005341068    0
#> [30,]  0.000000e+00  0.000000e+00  0.000000000    0  0.000000000    0
#> [31,]  4.657302e-16 -3.008700e-16  0.038021898    0 -0.049533266    0
#>               [,7] [,8] [,9] [,10]       [,11] [,12]        [,13] [,14]
#>  [1,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#>  [2,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#>  [3,]  0.014507444    0    0     0 -0.04389531     0  0.027738516     0
#>  [4,]  0.039742573    0    0     0 -0.03636821     0 -0.001403491     0
#>  [5,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#>  [6,]  0.188565464    0    0     0  0.99517998     0 -1.126407086     0
#>  [7,] -0.470112931    0    0     0  0.76883758     0 -0.281116432     0
#>  [8,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#>  [9,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [10,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [11,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [12,]  0.204991062    0    0     0 -0.05300693     0 -0.158870615     0
#> [13,] -0.267775787    0    0     0  3.71929065     0 -3.329439772     0
#> [14,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [15,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [16,]  0.138676489    0    0     0 -0.41646820     0  0.265070074     0
#> [17,]  0.007511531    0    0     0 -0.01910670     0  0.011099761     0
#> [18,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [19,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [20,] -0.242758881    0    0     0 -0.02007082     0  0.236404028     0
#> [21,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [22,] -0.248720797    0    0     0  0.28253554     0 -0.034041664     0
#> [23,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [24,]  0.964003336    0    0     0 -0.93585073     0 -0.034382889     0
#> [25,]  0.016382556    0    0     0 -0.01131731     0 -0.002739417     0
#> [26,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [27,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [28,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [29,] -0.005411473    0    0     0  0.01008697     0 -0.004677128     0
#> [30,]  0.000000000    0    0     0  0.00000000     0  0.000000000     0
#> [31,]  0.112269935    0    0     0 -0.12217001     0  0.112817841     0
#>              [,15] [,16]        [,17] [,18]         [,19]        [,20] [,21]
#>  [1,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#>  [2,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#>  [3,] -0.027269607     0 -0.027553874     0  0.0277394369  0.026702636     0
#>  [4,]  0.005788532     0  0.004702662     0  0.0012550288 -0.006908546     0
#>  [5,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#>  [6,]  1.136019033     0  1.124116708     0 -1.1089678058 -1.141916934     0
#>  [7,]  0.260660858     0  0.273125960     0 -0.2906640984 -0.243363924     0
#>  [8,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#>  [9,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [10,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [11,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [12,]  0.151214099     0  0.143515742     0 -0.1617449200 -0.160585279     0
#> [13,]  3.257071039     0  3.243870654     0 -3.3386261506 -3.252419997     0
#> [14,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [15,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [16,] -0.256875171     0 -0.259440173     0  0.2673053718  0.251909346     0
#> [17,] -0.010622439     0 -0.010778298     0  0.0112744472  0.010359830     0
#> [18,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [19,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [20,] -0.266557237     0 -0.258783304     0  0.2157987336  0.272827327     0
#> [21,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [22,]  0.018714477     0  0.025768170     0 -0.0429751187 -0.010194213     0
#> [23,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [24,]  0.074739718     0  0.045682615     0 -0.0094642479 -0.109974759     0
#> [25,]  0.006260415     0  0.005871169     0 -0.0005941079 -0.006511163     0
#> [26,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [27,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [28,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [29,]  0.004102328     0  0.004223756     0 -0.0049756103 -0.003943690     0
#> [30,]  0.000000000     0  0.000000000     0  0.0000000000  0.000000000     0
#> [31,]  0.025727983     0  0.028507717     0  0.1947546834 -0.013553622     0
#>              [,22]        [,23]         [,24]         [,25]        [,26]
#>  [1,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#>  [2,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#>  [3,]  0.027006430 -0.014641885  0.0130086748  2.703060e-02 -0.014003147
#>  [4,] -0.002762110 -0.039682657 -0.0422495131 -4.187831e-05 -0.037928129
#>  [5,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#>  [6,] -1.136913760 -0.180121216 -1.3158120061 -1.118351e+00 -0.165838193
#>  [7,] -0.257891128  0.470368541  0.1961337374 -2.684751e-01  0.447584783
#>  [8,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#>  [9,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [10,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [11,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [12,] -0.173313549 -0.202569196 -0.3630339336 -1.753162e-01 -0.189690634
#> [13,] -3.334654886  0.292034736 -3.0391751899 -3.341635e+00  0.298972525
#> [14,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [15,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [16,]  0.258960846 -0.140091153  0.1235030836  2.613615e-01 -0.134375223
#> [17,]  0.010772838 -0.007566729  0.0034308569  1.095789e-02 -0.007246720
#> [18,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [19,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [20,]  0.244177349  0.240930468  0.4861715940  2.231126e-01  0.229594223
#> [21,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [22,] -0.022691093  0.248218314  0.2191690663 -3.217129e-02  0.236147759
#> [23,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [24,] -0.083141689 -0.960234445 -1.0115541251 -5.571403e-02 -0.910738890
#> [25,] -0.002891904 -0.016402913 -0.0199257458 -7.461509e-04 -0.015850481
#> [26,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [27,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [28,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [29,] -0.004492802  0.005436305  0.0008890919 -4.797163e-03  0.005215853
#> [30,]  0.000000000  0.000000000  0.0000000000  0.000000e+00  0.000000000
#> [31,]  0.139328527 -0.117331962 -0.0287746996  2.194369e-01 -0.126130765
#>              [,27] [,28] [,29]         [,30]        [,31]        [,32] [,33]
#>  [1,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#>  [2,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#>  [3,] -0.042454025     0     0 -2.273221e-06  0.013911944 -0.014824783     0
#>  [4,] -0.038181732     0     0 -1.899458e-03  0.038858413 -0.040722199     0
#>  [5,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#>  [6,]  0.938538799     0     0 -7.078674e-04  0.191149719 -0.181851867     0
#>  [7,]  0.755841871     0     0  2.775083e-03 -0.455378338  0.476771821     0
#>  [8,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#>  [9,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [10,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [11,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [12,] -0.050740065     0     0  9.641786e-03  0.199280274 -0.201857519     0
#> [13,]  3.594226356     0     0  4.357601e-02 -0.242283154  0.310438136     0
#> [14,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [15,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [16,] -0.405035022     0     0 -2.385973e-03  0.133251687 -0.142567500     0
#> [17,] -0.018668775     0     0 -1.471113e-04  0.007242164 -0.007704492     0
#> [18,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [19,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [20,]  0.004338861     0     0  1.350813e-02 -0.239143579  0.247736068     0
#> [21,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [22,]  0.284141874     0     0  4.216613e-03 -0.242026023  0.252386471     0
#> [23,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [24,] -0.938253635     0     0 -4.037165e-03  0.937811969 -0.972504268     0
#> [25,] -0.013251109     0     0 -1.871884e-03  0.016155956 -0.017165516     0
#> [26,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [27,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [28,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [29,]  0.010080344     0     0  2.495594e-04 -0.005256507  0.005577858     0
#> [30,]  0.000000000     0     0  0.000000e+00  0.000000000  0.000000000     0
#> [31,] -0.201061960     0     0 -8.583090e-02  0.118054627 -0.145115335     0
#>       [,34] [,35]
#>  [1,]     0     0
#>  [2,]     0     0
#>  [3,]     0     0
#>  [4,]     0     0
#>  [5,]     0     0
#>  [6,]     0     0
#>  [7,]     0     0
#>  [8,]     0     0
#>  [9,]     0     0
#> [10,]     0     0
#> [11,]     0     0
#> [12,]     0     0
#> [13,]     0     0
#> [14,]     0     0
#> [15,]     0     0
#> [16,]     0     0
#> [17,]     0     0
#> [18,]     0     0
#> [19,]     0     0
#> [20,]     0     0
#> [21,]     0     0
#> [22,]     0     0
#> [23,]     0     0
#> [24,]     0     0
#> [25,]     0     0
#> [26,]     0     0
#> [27,]     0     0
#> [28,]     0     0
#> [29,]     0     0
#> [30,]     0     0
#> [31,]     0     0
#> 
#> $a[[2]]
#>                [,1]          [,2]        [,3]          [,4]        [,5]
#>  [1,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#>  [2,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#>  [3,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#>  [4,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#>  [5,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#>  [6,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#>  [7,] -2.368268e-16  9.190392e-17  0.20287008 -1.564421e-31 -0.22552464
#>  [8,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#>  [9,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [10,] -1.706430e-16  5.379762e-16  0.69934242 -4.278612e-31 -0.75738254
#> [11,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [12,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [13,]  3.635116e-16  3.636033e-16  0.26166850 -2.130672e-31 -0.25724555
#> [14,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [15,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [16,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [17,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [18,]  2.315036e-16  3.750030e-16  0.37598793  2.388586e-32 -0.41727042
#> [19,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [20,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [21,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [22,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [23,] -3.916722e-16  8.772854e-16  1.30586170  9.095129e-34 -1.49899723
#> [24,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [25,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [26,] -4.394822e-16  7.414775e-16  0.87973035 -1.703215e-30 -0.84537697
#> [27,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [28,]  8.886695e-19 -1.005623e-17 -0.01867629 -3.344173e-32  0.02460369
#> [29,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [30,]  0.000000e+00  0.000000e+00  0.00000000  0.000000e+00  0.00000000
#> [31,]  3.827668e-17 -3.246107e-17  0.01418949  4.567672e-31 -0.06059820
#>                [,6]        [,7]          [,8]          [,9]         [,10]
#>  [1,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#>  [2,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#>  [3,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#>  [4,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#>  [5,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#>  [6,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#>  [7,]  4.160493e-33  0.24901101  1.911067e-32  7.081561e-33 -1.691582e-33
#>  [8,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#>  [9,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [10,] -3.444962e-33  0.72818721  7.476072e-32  1.400660e-32 -1.643606e-32
#> [11,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [12,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [13,] -8.482460e-33  0.17634770  3.925256e-32 -1.088556e-33 -1.396347e-32
#> [14,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [15,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [16,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [17,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [18,] -1.791384e-32  0.34185297  3.619324e-32  1.451146e-33 -1.290118e-32
#> [19,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [20,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [21,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [22,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [23,] -3.377163e-32  1.43816460  1.039607e-31  2.376085e-32 -2.439115e-32
#> [24,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [25,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [26,]  4.254450e-32  0.87303027  1.389602e-31  2.619092e-32 -2.405299e-32
#> [27,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [28,]  1.791094e-33 -0.02217593 -1.613114e-34 -1.205622e-34  2.214985e-34
#> [29,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [30,]  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00  0.000000e+00
#> [31,] -1.777434e-32  0.04229535 -1.744336e-32 -2.402022e-33  1.857643e-33
#>               [,11]         [,12]        [,13]         [,14]         [,15]
#>  [1,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#>  [2,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#>  [3,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#>  [4,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#>  [5,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#>  [6,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#>  [7,] -0.5679919741  8.745903e-34  0.349116399  1.462465e-33 -0.3259820178
#>  [8,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#>  [9,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [10,] -0.7395135784  6.341704e-34  0.010355727 -1.330899e-33 -0.0119927272
#> [11,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [12,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [13,]  0.5475751732 -1.263746e-33 -0.807691414 -5.143697e-33  0.7490259401
#> [14,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [15,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [16,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [17,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [18,]  0.1844018295 -9.673462e-34 -0.557354310 -3.235363e-33  0.5222877693
#> [19,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [20,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [21,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [22,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [23,] -1.6333084784  9.754864e-34  0.295028752  9.827994e-34 -0.2620089470
#> [24,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [25,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [26,] -1.1955985484  2.289140e-33  0.233321146 -3.263768e-33 -0.2442940340
#> [27,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [28,]  0.0190679226  2.261144e-35 -0.001161799 -7.783473e-35  0.0002047903
#> [29,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [30,]  0.0000000000  0.000000e+00  0.000000000  0.000000e+00  0.0000000000
#> [31,] -0.0008980201 -4.105923e-34  0.002477292  1.110640e-33  0.0073854701
#>               [,16]         [,17]         [,18]        [,19]        [,20] [,21]
#>  [1,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#>  [2,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#>  [3,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#>  [4,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#>  [5,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#>  [6,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#>  [7,] -9.255059e-34 -0.3563979712 -3.690018e-34  0.350179644  0.353236415     0
#>  [8,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#>  [9,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [10,] -9.903610e-35 -0.0412417209  2.501579e-34  0.004744022  0.007974301     0
#> [11,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [12,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [13,]  1.994141e-33  0.7857828358  1.152098e-33 -0.807759000 -0.803476799     0
#> [14,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [15,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [16,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [17,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [18,]  1.517189e-33  0.5480538521  8.165429e-34 -0.577791765 -0.592079538     0
#> [19,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [20,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [21,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [22,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [23,] -4.403913e-34 -0.3102210509 -3.273809e-35  0.233209807  0.189359335     0
#> [24,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [25,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [26,] -1.331417e-33 -0.3281382208  2.389585e-34  0.302113211  0.379012149     0
#> [27,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [28,] -2.004456e-35 -0.0002577987  2.513637e-36  0.001893206  0.004609372     0
#> [29,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [30,]  0.000000e+00  0.0000000000  0.000000e+00  0.000000000  0.000000000     0
#> [31,]  2.509246e-34  0.0201133644 -8.136340e-35 -0.027785650 -0.056793796     0
#>              [,22]       [,23]       [,24]        [,25]       [,26]       [,27]
#>  [1,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#>  [2,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#>  [3,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#>  [4,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#>  [5,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#>  [6,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#>  [7,]  0.330716589 -0.23648364  0.17873563  0.367175957 -0.23125810 -0.59406006
#>  [8,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#>  [9,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [10,]  0.002948745 -0.75076027 -0.63146169  0.049697370 -0.74541482 -0.81351760
#> [11,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [12,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [13,] -0.767523159 -0.23367302 -1.08196997 -0.792209395 -0.22354929  0.53299582
#> [14,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [15,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [16,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [17,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [18,] -0.542463496 -0.37852346 -0.93267963 -0.572574580 -0.40780112  0.14846518
#> [19,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [20,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [21,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [22,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [23,]  0.237262120 -1.43850788 -0.90954514  0.268975515 -1.53292622 -1.79005170
#> [24,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [25,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [26,]  0.256619341 -0.92617409 -0.53804765  0.432325953 -0.76833639 -1.27097460
#> [27,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [28,]  0.000764938  0.02128101  0.01695676  0.003477084  0.02697245  0.02186689
#> [29,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [30,]  0.000000000  0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> [31,] -0.015259462 -0.02761816 -0.02099678 -0.055796868 -0.08802154 -0.01082993
#>       [,28] [,29] [,30]       [,31]       [,32] [,33] [,34] [,35]
#>  [1,]     0     0     0  0.00000000  0.00000000     0     0     0
#>  [2,]     0     0     0  0.00000000  0.00000000     0     0     0
#>  [3,]     0     0     0  0.00000000  0.00000000     0     0     0
#>  [4,]     0     0     0  0.00000000  0.00000000     0     0     0
#>  [5,]     0     0     0  0.00000000  0.00000000     0     0     0
#>  [6,]     0     0     0  0.00000000  0.00000000     0     0     0
#>  [7,]     0     0     0  0.22220309 -0.22184654     0     0     0
#>  [8,]     0     0     0  0.00000000  0.00000000     0     0     0
#>  [9,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [10,]     0     0     0  0.72982147 -0.68401789     0     0     0
#> [11,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [12,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [13,]     0     0     0  0.24036594 -0.19417745     0     0     0
#> [14,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [15,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [16,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [17,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [18,]     0     0     0  0.38798559 -0.34034575     0     0     0
#> [19,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [20,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [21,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [22,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [23,]     0     0     0  1.42310236 -1.33973854     0     0     0
#> [24,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [25,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [26,]     0     0     0  0.85270733 -0.81459681     0     0     0
#> [27,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [28,]     0     0     0 -0.02234675  0.02071365     0     0     0
#> [29,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [30,]     0     0     0  0.00000000  0.00000000     0     0     0
#> [31,]     0     0     0  0.04419008 -0.03875661     0     0     0
#> 
#> 
#> $BIC
#> [1] 7319.677
#>