Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr()
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1),
alt_iter = 1, anneal_iter = 3, em_iter = 3)
#> mixedLSR Start: 1 
#> EM Step...
#> Simulated Annealing Step
#> Full Cycle 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -3692.833
#> 
#> $assign
#>   [1] 2 1 2 1 2 1 1 1 1 2 2 1 1 1 1 2 2 1 2 2 1 1 1 1 2 1 1 1 1 2 2 2 1 1 2 2 1
#>  [38] 2 2 2 1 2 2 1 2 1 1 1 1 1 1 1 1 2 2 1 2 1 1 1 1 2 2 2 2 1 2 2 1 1 2 1 2 1
#>  [75] 1 1 2 2 1 1 1 2 1 1 2 1 1 1 2 1 1 2 1 2 1 1 1 1 1 2
#> 
#> $a
#> $a[[1]]
#>               [,1]          [,2]         [,3]         [,4] [,5]        [,6]
#>  [1,] -0.998483833  3.497213e-16 -0.611817978 -0.596413005    0  0.40626005
#>  [2,]  2.714897050 -9.225386e-16 -0.614286438 -0.645515295    0 -3.41012374
#>  [3,] -0.018210672  1.088238e-18 -1.465637788 -1.439932356    0 -1.42594819
#>  [4,] -0.221774165  6.868136e-17 -0.091787456 -0.081851204    0  0.14887487
#>  [5,]  0.000000000  0.000000e+00  0.000000000  0.000000000    0  0.00000000
#>  [6,]  0.034598523 -9.244084e-18 -0.683297235 -0.675842234    0 -0.71735544
#>  [7,]  0.017928017  1.299377e-18  0.269959704  0.259875176    0  0.23691825
#>  [8,]  0.298452913 -1.398546e-16  0.164983875  0.188745642    0 -0.08062768
#>  [9,]  0.008258589 -1.973517e-18 -0.630855871 -0.622439135    0 -0.63539909
#> [10,]  0.510217104 -1.784542e-16 -0.381505848 -0.379452305    0 -0.89625538
#> [11,] -0.239960290  9.444876e-17 -0.028651842 -0.034953548    0  0.19776988
#> [12,] -0.066981676  1.259475e-17 -0.122425705 -0.111559153    0 -0.03542780
#> [13,] -1.594239952  5.544414e-16  0.670551933  0.674258160    0  2.28862434
#> [14,] -0.217176204  8.136814e-17 -0.288656850 -0.287255487    0 -0.07468525
#> [15,]  0.000000000  0.000000e+00  0.000000000  0.000000000    0  0.00000000
#> [16,]  0.000000000  0.000000e+00  0.000000000  0.000000000    0  0.00000000
#> [17,]  0.529553819 -1.950065e-16 -0.194556791 -0.187573794    0 -0.71429941
#> [18,] -0.328582442  1.843776e-16  0.876811188  0.811018694    0  1.08783137
#> [19,] -0.046230628  1.717242e-17  0.007350157  0.006763189    0  0.05256053
#> [20,] -0.169567336  6.042563e-17  0.096681594  0.095549033    0  0.26615857
#> [21,]  0.617972289 -2.207154e-16  1.141508097  1.124243975    0  0.51200630
#> [22,] -0.443009714  1.615323e-16 -0.093319038 -0.094148804    0  0.34631755
#> [23,] -0.483269569  1.779180e-16 -0.517561218 -0.513759596    0 -0.03726805
#> [24,]  2.401383999 -8.707045e-16  0.663290586  0.661593476    0 -1.72926708
#> [25,] -0.404894400  1.465380e-16  0.523329968  0.514566768    0  0.92178996
#> [26,]  0.146061530 -4.499578e-17  0.022162593  0.015987561    0 -0.13640823
#> [27,]  0.126894444 -5.508970e-17  0.045705313  0.052685150    0 -0.06576301
#> [28,] -2.444353789  8.946563e-16 -0.927029848 -0.928290059    0  1.49660235
#>               [,7]        [,8] [,9]        [,10] [,11] [,12]        [,13] [,14]
#>  [1,] -0.605806235  0.99295595    0 -0.585879529     0     0 -0.999599347     0
#>  [2,] -0.634812845 -2.77134946    0 -0.704277953     0     0  2.698689356     0
#>  [3,] -1.476061525  0.01327789    0 -1.426182727     0     0 -0.015590307     0
#>  [4,] -0.093209939  0.23666253    0 -0.067091611     0     0 -0.216261297     0
#>  [5,]  0.000000000  0.00000000    0  0.000000000     0     0  0.000000000     0
#>  [6,] -0.686780440 -0.04622957    0 -0.677259068     0     0  0.032448253     0
#>  [7,]  0.273957771 -0.02854417    0  0.247971344     0     0  0.013244164     0
#>  [8,]  0.151208535 -0.23536778    0  0.235170934     0     0  0.321308896     0
#>  [9,] -0.634424712 -0.01596259    0 -0.621127997     0     0  0.007360653     0
#> [10,] -0.388672081 -0.51453889    0 -0.382253753     0     0  0.510277040     0
#> [11,] -0.022998480  0.22163293    0 -0.047499601     0     0 -0.246804284     0
#> [12,] -0.126164098  0.08477601    0 -0.095281595     0     0 -0.060166176     0
#> [13,]  0.688741768  1.60821800    0  0.689645717     0     0 -1.592673351     0
#> [14,] -0.286504018  0.20522295    0 -0.291736893     0     0 -0.220879343     0
#> [15,]  0.000000000  0.00000000    0  0.000000000     0     0  0.000000000     0
#> [16,]  0.000000000  0.00000000    0  0.000000000     0     0  0.000000000     0
#> [17,] -0.204072148 -0.51501870    0 -0.177613872     0     0  0.535962234     0
#> [18,]  0.908567070  0.21632891    0  0.712827112     0     0 -0.372607056     0
#> [19,]  0.008172647  0.04461018    0  0.005726774     0     0 -0.046889452     0
#> [20,]  0.099257584  0.16876342    0  0.095041144     0     0 -0.170315346     0
#> [21,]  1.140944651 -0.59973120    0  1.120485071     0     0  0.621770984     0
#> [22,] -0.087439560  0.43117356    0 -0.099150551     0     0 -0.447476415     0
#> [23,] -0.513076810  0.46336348    0 -0.519850413     0     0 -0.489440267     0
#> [24,]  0.634014543 -2.34424822    0  0.681175095     0     0  2.422550375     0
#> [25,]  0.531998763  0.40187195    0  0.508961052     0     0 -0.407970224     0
#> [26,]  0.022948977 -0.15665245    0  0.006094907     0     0  0.142259528     0
#> [27,]  0.041170901 -0.10793555    0  0.066399165     0     0  0.133810391     0
#> [28,] -0.895932564  2.36915945    0 -0.958890049     0     0 -2.471366375     0
#>       [,15]         [,16] [,17]        [,18] [,19]       [,20] [,21] [,22]
#>  [1,]     0  0.0061880207     0  1.007703716     0 -0.98987311     0     0
#>  [2,]     0 -0.0246732342     0 -2.791854952     0  2.75951268     0     0
#>  [3,]     0  0.0147424062     0  0.047579805     0  0.01214172     0     0
#>  [4,]     0  0.0086161016     0  0.247327411     0 -0.22595925     0     0
#>  [5,]     0  0.0000000000     0  0.000000000     0  0.00000000     0     0
#>  [6,]     0  0.0023083782     0 -0.035256458     0  0.05200257     0     0
#>  [7,]     0 -0.0082866952     0 -0.040972841     0  0.01615327     0     0
#>  [8,]     0  0.0279230554     0 -0.206930832     0  0.27629688     0     0
#>  [9,]     0  0.0036322062     0 -0.004177558     0  0.02321834     0     0
#> [10,]     0  0.0014903787     0 -0.508385399     0  0.52179829     0     0
#> [11,]     0 -0.0080734506     0  0.213235832     0 -0.23470499     0     0
#> [12,]     0  0.0101056560     0  0.097391135     0 -0.07069667     0     0
#> [13,]     0  0.0016505459     0  1.602356945     0 -1.62031617     0     0
#> [14,]     0 -0.0020142373     0  0.206674433     0 -0.20875253     0     0
#> [15,]     0  0.0000000000     0  0.000000000     0  0.00000000     0     0
#> [16,]     0  0.0000000000     0  0.000000000     0  0.00000000     0     0
#> [17,]     0  0.0083526259     0 -0.503667407     0  0.53143772     0     0
#> [18,]     0 -0.0636774525     0  0.135831087     0 -0.31113423     0     0
#> [19,]     0 -0.0007860428     0  0.043676163     0 -0.04609140     0     0
#> [20,]     0 -0.0012224305     0  0.166294276     0 -0.17209671     0     0
#> [21,]     0 -0.0055904212     0 -0.620229899     0  0.59281957     0     0
#> [22,]     0 -0.0038467538     0  0.428288535     0 -0.43965049     0     0
#> [23,]     0 -0.0027157512     0  0.466984632     0 -0.46915236     0     0
#> [24,]     0  0.0156364598     0 -2.336273053     0  2.38235512     0     0
#> [25,]     0 -0.0068590261     0  0.388038708     0 -0.41633272     0     0
#> [26,]     0 -0.0056062224     0 -0.163127543     0  0.14982630     0     0
#> [27,]     0  0.0083729312     0 -0.099376686     0  0.12050690     0     0
#> [28,]     0 -0.0211787405     0  2.358392639     0 -2.41449895     0     0
#>              [,23] [,24]       [,25] [,26]       [,27] [,28]        [,29] [,30]
#>  [1,]  1.001626620     0  1.59305346     0  1.00290701     0  0.609291584     0
#>  [2,] -2.709686407     0 -2.08852689     0 -2.76625133     0  0.576316103     0
#>  [3,]  0.004528815     0  1.45426083     0  0.02822325     0  1.468088412     0
#>  [4,]  0.214075776     0  0.30786255     0  0.23616794     0  0.103330980     0
#>  [5,]  0.000000000     0  0.00000000     0  0.00000000     0  0.000000000     0
#>  [6,] -0.036129337     0  0.63675379     0 -0.03808309     0  0.677361854     0
#>  [7,] -0.009149921     0 -0.28036840     0 -0.02971548     0 -0.279187947     0
#>  [8,] -0.333496585     0 -0.46925215     0 -0.24685190     0 -0.117296638     0
#>  [9,] -0.011180601     0  0.61091413     0 -0.00879320     0  0.627668804     0
#> [10,] -0.515708697     0 -0.13303259     0 -0.51189369     0  0.380296428     0
#> [11,]  0.251431915     0  0.26985066     0  0.22528096     0  0.014576133     0
#> [12,]  0.056249879     0  0.18343121     0  0.08366397     0  0.136457069     0
#> [13,]  1.605255143     0  0.92631378     0  1.60493453     0 -0.664194719     0
#> [14,]  0.222037909     0  0.50071611     0  0.21041310     0  0.280568704     0
#> [15,]  0.000000000     0  0.00000000     0  0.00000000     0  0.000000000     0
#> [16,]  0.000000000     0  0.00000000     0  0.00000000     0  0.000000000     0
#> [17,] -0.543478569     0 -0.33859526     0 -0.51691148     0  0.206952931     0
#> [18,]  0.402722224     0 -0.51253163     0  0.22477476     0 -0.966455686     0
#> [19,]  0.047531585     0  0.03905582     0  0.04491838     0 -0.008659507     0
#> [20,]  0.172289116     0  0.07408523     0  0.16849243     0 -0.097860347     0
#> [21,] -0.619491383     0 -1.73557198     0 -0.61536983     0 -1.131883524     0
#> [22,]  0.451591786     0  0.53398241     0  0.43519041     0  0.084283974     0
#> [23,]  0.491801877     0  0.99087212     0  0.47275896     0  0.504137540     0
#> [24,] -2.442252239     0 -3.04748506     0 -2.36655353     0 -0.620341107     0
#> [25,]  0.415301043     0 -0.10932711     0  0.39854931     0 -0.528228893     0
#> [26,] -0.140989799     0 -0.16512772     0 -0.15584502     0 -0.030201484     0
#> [27,] -0.137730351     0 -0.17422137     0 -0.11147029     0 -0.031323521     0
#> [28,]  2.492513977     0  3.35161267     0  2.39671016     0  0.871388560     0
#>              [,31]        [,32]        [,33] [,34]        [,35]
#>  [1,] -1.001553009  0.601131053 -0.605266421     0 -0.594182149
#>  [2,]  2.696810464  0.621332809 -0.619599362     0 -0.664541581
#>  [3,] -0.018858659  1.447586385 -1.451697170     0 -1.437782081
#>  [4,] -0.216065450  0.088467816 -0.088466675     0 -0.077355944
#>  [5,]  0.000000000  0.000000000  0.000000000     0  0.000000000
#>  [6,]  0.030634403  0.675901678 -0.678186367     0 -0.677310614
#>  [7,]  0.013469393 -0.265510247  0.265725085     0  0.256513867
#>  [8,]  0.323900634 -0.167736556  0.171993014     0  0.203752457
#>  [9,]  0.005775723  0.623662230 -0.625673771     0 -0.622969706
#> [10,]  0.509571377  0.378895738 -0.378950108     0 -0.380817201
#> [11,] -0.247589796  0.029156798 -0.030690489     0 -0.039014832
#> [12,] -0.059864257  0.118992570 -0.118575379     0 -0.106600249
#> [13,] -1.591554051 -0.669004388  0.667837533     0  0.679840866
#> [14,] -0.221990843  0.285375508 -0.287217566     0 -0.289148818
#> [15,]  0.000000000  0.000000000  0.000000000     0  0.000000000
#> [16,]  0.000000000  0.000000000  0.000000000     0  0.000000000
#> [17,]  0.536288538  0.192611151 -0.191192581     0 -0.184612689
#> [18,] -0.374650143 -0.856472231  0.852410339     0  0.781175849
#> [19,] -0.046949992 -0.007268828  0.007102302     0  0.006437353
#> [20,] -0.170219009 -0.095984451  0.095812133     0  0.095501708
#> [21,]  0.625084707 -1.126672726  1.131961981     0  1.124883088
#> [22,] -0.448279455  0.091693472 -0.093532060     0 -0.095955756
#> [23,] -0.491427257  0.511206465 -0.514645879     0 -0.516556073
#> [24,]  2.427001997 -0.651843214  0.661869548     0  0.669239587
#> [25,] -0.407159213 -0.517881449  0.518177867     0  0.513501677
#> [26,]  0.142023526 -0.020385031  0.020245189     0  0.012915015
#> [27,]  0.134598883 -0.046450924  0.047804327     0  0.057120838
#> [28,] -2.476996861  0.913794272 -0.925507574     0 -0.939829876
#>  [ reached getOption("max.print") -- omitted 3 rows ]
#> 
#> $a[[2]]
#>              [,1]          [,2]          [,3]          [,4] [,5]        [,6]
#>  [1,]  0.64866803 -2.546184e-16  0.4047598601  0.4080216097    0 -0.21320135
#>  [2,]  0.31085800 -2.208969e-16 -0.0820730452 -0.0804657987    0 -0.38313914
#>  [3,]  0.07040136 -6.271677e-17 -0.0045195033 -0.0041724089    0 -0.07913437
#>  [4,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#>  [5,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#>  [6,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#>  [7,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#>  [8,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#>  [9,] -0.02120606 -3.009627e-17 -0.1223598495 -0.1224484814    0 -0.10385203
#> [10,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [11,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [12,]  0.01206140 -7.403981e-18  0.0008240478  0.0008855463    0 -0.01088729
#> [13,] -1.61822693  1.529374e-15  1.4037817485  1.3952833362    0  2.99963456
#> [14,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [15,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [16,] -0.41205433  1.590499e-16 -0.2445965502 -0.2466764505    0  0.14547035
#> [17,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [18,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [19,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [20,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [21,] -0.30577411  2.787213e-16  0.3296198766  0.3279762114    0  0.61975584
#> [22,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [23,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [24,]  0.38741361 -3.240962e-16 -0.0254642422 -0.0235365503    0 -0.42806855
#> [25,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [26,]  0.15010104 -4.022744e-17  0.0326752548  0.0334732824    0 -0.09612960
#> [27,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#> [28,]  0.00000000  0.000000e+00  0.0000000000  0.0000000000    0  0.00000000
#>               [,7]        [,8] [,9]        [,10] [,11] [,12]       [,13] [,14]
#>  [1,]  0.437258418 -0.62590249    0  0.442031756     0     0  0.63342108     0
#>  [2,] -0.074045026 -0.30306862    0 -0.066670889     0     0  0.30172392     0
#>  [3,] -0.011432394 -0.07790125    0 -0.016376999     0     0  0.07474850     0
#>  [4,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#>  [5,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#>  [6,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#>  [7,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#>  [8,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#>  [9,] -0.126061638  0.01958364    0 -0.123967093     0     0 -0.02166849     0
#> [10,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [11,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [12,]  0.001106839 -0.01185680    0  0.001222686     0     0  0.01182944     0
#> [13,]  1.403939738  1.60448533    0  1.358193058     0     0 -1.57479371     0
#> [14,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [15,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [16,] -0.268588286  0.39392016    0 -0.274423752     0     0 -0.39972679     0
#> [17,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [18,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [19,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [20,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [21,]  0.314277088  0.28619520    0  0.292502256     0     0 -0.28525802     0
#> [22,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [23,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [24,] -0.052599191 -0.41740091    0 -0.072182905     0     0  0.40375579     0
#> [25,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [26,]  0.059370977 -0.12397689    0  0.076130931     0     0  0.13171253     0
#> [27,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#> [28,]  0.000000000  0.00000000    0  0.000000000     0     0  0.00000000     0
#>       [,15] [,16] [,17]       [,18] [,19]       [,20] [,21] [,22]       [,23]
#>  [1,]     0     0     0 -0.62035406     0  0.64814757     0     0 -0.62422533
#>  [2,]     0     0     0 -0.30186925     0  0.31767495     0     0 -0.30900731
#>  [3,]     0     0     0 -0.07543776     0  0.06084703     0     0 -0.08061827
#>  [4,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#>  [5,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#>  [6,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#>  [7,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#>  [8,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#>  [9,]     0     0     0  0.01874538     0 -0.01799051     0     0  0.01692806
#> [10,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [11,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [12,]     0     0     0 -0.01175695     0  0.01206040     0     0 -0.01201534
#> [13,]     0     0     0  1.59986183     0 -1.66074267     0     0  1.66212330
#> [14,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [15,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [16,]     0     0     0  0.39131895     0 -0.41639083     0     0  0.39246609
#> [17,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [18,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [19,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [20,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [21,]     0     0     0  0.28964157     0 -0.33563697     0     0  0.29543392
#> [22,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [23,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [24,]     0     0     0 -0.40638483     0  0.34771852     0     0 -0.43031124
#> [25,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [26,]     0     0     0 -0.12785705     0  0.17609426     0     0 -0.12115551
#> [27,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#> [28,]     0     0     0  0.00000000     0  0.00000000     0     0  0.00000000
#>       [,24]       [,25] [,26]       [,27] [,28]        [,29] [,30]       [,31]
#>  [1,]     0 -1.05318263     0 -0.61775411     0 -0.432446095     0  0.62632654
#>  [2,]     0 -0.22953826     0 -0.30128713     0  0.073691040     0  0.30688014
#>  [3,]     0 -0.06613958     0 -0.07908321     0  0.006103449     0  0.07200443
#>  [4,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#>  [5,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#>  [6,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#>  [7,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#>  [8,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#>  [9,]     0  0.14321859     0  0.01854170     0  0.125055764     0 -0.01792301
#> [10,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [11,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [12,]     0 -0.01290403     0 -0.01178253     0 -0.001163622     0  0.01185253
#> [13,]     0  0.22166738     0  1.60555741     0 -1.383676152     0 -1.62608710
#> [14,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [15,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [16,]     0  0.65648135     0  0.38815479     0  0.263472484     0 -0.39701511
#> [17,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [18,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [19,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [20,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [21,]     0 -0.02252517     0  0.28370981     0 -0.319947413     0 -0.30347781
#> [22,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [23,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [24,]     0 -0.36321191     0 -0.42188039     0  0.029720258     0  0.39283851
#> [25,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [26,]     0 -0.18261126     0 -0.11868799     0 -0.046533861     0  0.13990848
#> [27,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#> [28,]     0  0.00000000     0  0.00000000     0  0.000000000     0  0.00000000
#>               [,32]        [,33] [,34]        [,35]
#>  [1,] -0.3986475523  0.409748057     0  0.423766149
#>  [2,]  0.0867414901 -0.083808253     0 -0.073737539
#>  [3,]  0.0041926936  0.008345857     0 -0.011538401
#>  [4,]  0.0000000000  0.000000000     0  0.000000000
#>  [5,]  0.0000000000  0.000000000     0  0.000000000
#>  [6,]  0.0000000000  0.000000000     0  0.000000000
#>  [7,]  0.0000000000  0.000000000     0  0.000000000
#>  [8,]  0.0000000000  0.000000000     0  0.000000000
#>  [9,]  0.1228964703 -0.124646056     0 -0.122949274
#> [10,]  0.0000000000  0.000000000     0  0.000000000
#> [11,]  0.0000000000  0.000000000     0  0.000000000
#> [12,] -0.0006878181  0.001014562     0  0.001022984
#> [13,] -1.4320692174  1.405574937     0  1.378093991
#> [14,]  0.0000000000  0.000000000     0  0.000000000
#> [15,]  0.0000000000  0.000000000     0  0.000000000
#> [16,]  0.2401105546 -0.242296403     0 -0.260256140
#> [17,]  0.0000000000  0.000000000     0  0.000000000
#> [18,]  0.0000000000  0.000000000     0  0.000000000
#> [19,]  0.0000000000  0.000000000     0  0.000000000
#> [20,]  0.0000000000  0.000000000     0  0.000000000
#> [21,] -0.3378052524  0.355443779     0  0.308132645
#> [22,]  0.0000000000  0.000000000     0  0.000000000
#> [23,]  0.0000000000  0.000000000     0  0.000000000
#> [24,]  0.0252218021  0.029689666     0 -0.053371121
#> [25,]  0.0000000000  0.000000000     0  0.000000000
#> [26,] -0.0279145754  0.003083718     0  0.057357642
#> [27,]  0.0000000000  0.000000000     0  0.000000000
#> [28,]  0.0000000000  0.000000000     0  0.000000000
#>  [ reached getOption("max.print") -- omitted 3 rows ]
#> 
#> 
#> $BIC
#> [1] 10834.94
#>