Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2719.982
#> 
#> $assign
#>  [1] 2 2 1 2 1 2 2 1 1 2 2 1 1 1 1 1 2 2 2 1 2 2 2 2 2 1 1 2 1 2 1 1 2 1 1 1 1 1
#> [39] 2 1 1 1 2 1 1 2 1 2 1 1
#> 
#> $a
#> $a[[1]]
#>                [,1]        [,2]        [,3]        [,4] [,5]         [,6]
#>  [1,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#>  [2,] -2.593534e-16  0.35388632 -0.33364849 -0.47367751    0 -0.206731440
#>  [3,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#>  [4,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#>  [5,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#>  [6,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#>  [7,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#>  [8,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#>  [9,] -1.145723e-16 -0.02954055  0.03095655 -0.18078117    0 -0.218641493
#> [10,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [11,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [12,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [13,]  1.228638e-16  0.05533930 -0.05597482  0.19069341    0  0.251365331
#> [14,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [15,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [16,] -1.721498e-15  0.26320988 -0.25131909 -2.59432661    0 -2.448360871
#> [17,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [18,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [19,]  1.121085e-16 -0.06769655  0.06878983  0.15297681    0  0.088512658
#> [20,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [21,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [22,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [23,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [24,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [25,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [26,]  1.774131e-17 -0.03459323  0.03384578  0.02758654    0 -0.002778353
#> [27,] -1.082387e-16 -0.03237307  0.02975083 -0.14720491    0 -0.174562814
#> [28,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [29,] -5.584018e-17 -0.04287610  0.04178937 -0.07996576    0 -0.120280372
#> [30,]  0.000000e+00  0.00000000  0.00000000  0.00000000    0  0.000000000
#> [31,]  2.646230e-16  0.52797495 -0.42349584 -0.19298121    0 -0.006839073
#>              [,7] [,8] [,9]       [,10]        [,11]       [,12] [,13] [,14]
#>  [1,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#>  [2,] -0.51578035    0    0 -0.29053280  0.011179820 -0.32548953     0     0
#>  [3,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#>  [4,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#>  [5,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#>  [6,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#>  [7,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#>  [8,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#>  [9,] -0.17839564    0    0  0.04363629  0.007735391  0.04196653     0     0
#> [10,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [11,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [12,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [13,]  0.18546387    0    0 -0.06902968 -0.008901962 -0.06885295     0     0
#> [14,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [15,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [16,] -2.63835618    0    0 -0.17265926 -0.021745494 -0.17938440     0     0
#> [17,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [18,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [19,]  0.16155652    0    0  0.07594939  0.015290091  0.07284675     0     0
#> [20,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [21,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [22,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [23,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [24,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [25,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [26,]  0.03164039    0    0  0.03432628  0.002933830  0.03551869     0     0
#> [27,] -0.14445166    0    0  0.02804601 -0.005286196  0.03383483     0     0
#> [28,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [29,] -0.07558865    0    0  0.04475948  0.002107919  0.04721756     0     0
#> [30,]  0.00000000    0    0  0.00000000  0.000000000  0.00000000     0     0
#> [31,] -0.25028823    0    0 -0.11788958  0.273851731 -0.31404062     0     0
#>       [,15]        [,16]       [,17] [,18]       [,19]        [,20]       [,21]
#>  [1,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#>  [2,]     0 -0.007656695  0.31684326     0  0.53841754 -0.180665380 -0.50442613
#>  [3,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#>  [4,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#>  [5,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#>  [6,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#>  [7,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#>  [8,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#>  [9,]     0 -0.013516069 -0.02313572     0  0.17392204 -0.227757592 -0.18680435
#> [10,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [11,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [12,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [13,]     0  0.015997646  0.04554305     0 -0.17875586  0.263992969  0.19633274
#> [14,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [15,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [16,]     0 -0.056146533  0.25784046     0  2.66082907 -2.392316966 -2.64115186
#> [17,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [18,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [19,]     0 -0.010262120 -0.05961017     0 -0.17052323  0.067456237  0.15062290
#> [20,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [21,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [22,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [23,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [24,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [25,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [26,]     0 -0.003057736 -0.03037007     0 -0.03496103 -0.008900437  0.02809726
#> [27,]     0 -0.002234955 -0.02726585     0  0.14332115 -0.172282075 -0.14551109
#> [28,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [29,]     0 -0.006360499 -0.03651209     0  0.07175930 -0.126003866 -0.08055370
#> [30,]     0  0.000000000  0.00000000     0  0.00000000  0.000000000  0.00000000
#> [31,]     0 -0.197920885  0.49202165     0  0.21519976 -0.203372755 -0.36765085
#>       [,22]      [,23]        [,24] [,25] [,26] [,27]        [,28]       [,29]
#>  [1,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#>  [2,]     0 -0.4613586 -0.026713240     0     0     0 -0.158578494  0.47336513
#>  [3,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#>  [4,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#>  [5,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#>  [6,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#>  [7,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#>  [8,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#>  [9,]     0 -0.1787710 -0.014350438     0     0     0 -0.222572873  0.17558858
#> [10,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [11,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [12,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [13,]     0  0.1890426  0.015941401     0     0     0  0.258878553 -0.18445426
#> [14,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [15,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [16,]     0 -2.5165596 -0.090210000     0     0     0 -2.366346828  2.54942473
#> [17,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [18,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [19,]     0  0.1434251 -0.006735212     0     0     0  0.071660760 -0.15285803
#> [20,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [21,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [22,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [23,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [24,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [25,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [26,]     0  0.0255921 -0.001534140     0     0     0 -0.008738193 -0.02823403
#> [27,]     0 -0.1419571 -0.002112822     0     0     0 -0.174387054  0.14355917
#> [28,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [29,]     0 -0.0789453 -0.005484923     0     0     0 -0.125315724  0.07688863
#> [30,]     0  0.0000000  0.000000000     0     0     0  0.000000000  0.00000000
#> [31,]     0 -0.2700684 -0.231927690     0     0     0 -0.041449468  0.18489791
#>             [,30] [,31] [,32]       [,33] [,34]      [,35]
#>  [1,]  0.00000000     0     0  0.00000000     0  0.0000000
#>  [2,] -0.27420820     0     0  0.28715236     0  0.4668969
#>  [3,]  0.00000000     0     0  0.00000000     0  0.0000000
#>  [4,]  0.00000000     0     0  0.00000000     0  0.0000000
#>  [5,]  0.00000000     0     0  0.00000000     0  0.0000000
#>  [6,]  0.00000000     0     0  0.00000000     0  0.0000000
#>  [7,]  0.00000000     0     0  0.00000000     0  0.0000000
#>  [8,]  0.00000000     0     0  0.00000000     0  0.0000000
#>  [9,]  0.04139597     0     0 -0.03974259     0  0.1850884
#> [10,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [11,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [12,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [13,] -0.06544873     0     0  0.06420507     0 -0.1963375
#> [14,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [15,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [16,] -0.18289910     0     0  0.22927178     0  2.5827592
#> [17,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [18,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [19,]  0.07533615     0     0 -0.08008646     0 -0.1452420
#> [20,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [21,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [22,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [23,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [24,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [25,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [26,]  0.03303014     0     0 -0.03431149     0 -0.0253653
#> [27,]  0.02445625     0     0 -0.02238866     0  0.1466264
#> [28,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [29,]  0.04201365     0     0 -0.04173100     0  0.0823875
#> [30,]  0.00000000     0     0  0.00000000     0  0.0000000
#> [31,] -0.05994243     0     0  0.05194889     0  0.2787303
#> 
#> $a[[2]]
#>                [,1]        [,2]        [,3]       [,4] [,5]       [,6]
#>  [1,] -1.498048e-16 -0.25292804  0.27684529 -0.1089750    0 -0.3898834
#>  [2,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#>  [3,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#>  [4,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#>  [5,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#>  [6,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#>  [7,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#>  [8,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#>  [9,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [10,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [11,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [12,] -3.817799e-16 -0.77576696  0.85164310 -0.1575690    0 -1.0131589
#> [13,]  6.691125e-16  0.23998334 -0.25411360  0.8398887    0  1.1529644
#> [14,] -1.928414e-16  0.22029155 -0.25692067 -0.9285835    0 -0.7052153
#> [15,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [16,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [17,]  7.860783e-17  0.13277359 -0.13873651  0.4145769    0  0.5574948
#> [18,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [19,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [20,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [21,]  2.042714e-16  0.03688027 -0.03802703  0.2370402    0  0.2945425
#> [22,] -3.930329e-16 -0.14320115  0.12297308 -2.0516655    0 -2.2181206
#> [23,] -4.841889e-16 -0.19102660  0.20671542 -0.4027663    0 -0.6502144
#> [24,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [25,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [26,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [27,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [28,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [29,]  0.000000e+00  0.00000000  0.00000000  0.0000000    0  0.0000000
#> [30,]  4.418264e-16 -0.91203644  1.00856908  0.6722220    0 -0.2500715
#> [31,]  1.962671e-15  0.41143082 -0.47218201 -0.1188822    0  0.5241762
#>                [,7] [,8] [,9]       [,10] [,11]       [,12] [,13] [,14] [,15]
#>  [1,] -0.0971498830    0    0  0.28980191     0  0.27845620     0     0     0
#>  [2,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#>  [3,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#>  [4,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#>  [5,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#>  [6,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#>  [7,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#>  [8,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#>  [9,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [10,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [11,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [12,] -0.1154665785    0    0  0.89240940     0  0.85840249     0     0     0
#> [13,]  0.8692235278    0    0 -0.24973663     0 -0.21872722     0     0     0
#> [14,] -0.9625390399    0    0 -0.26641727     0 -0.25054833     0     0     0
#> [15,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [16,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [17,]  0.4076007714    0    0 -0.15305207     0 -0.15860687     0     0     0
#> [18,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [19,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [20,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [21,]  0.2494276733    0    0 -0.03343465     0 -0.02368343     0     0     0
#> [22,] -2.0653478345    0    0  0.15333378     0  0.18527917     0     0     0
#> [23,] -0.4228148588    0    0  0.20059891     0  0.17139625     0     0     0
#> [24,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [25,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [26,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [27,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [28,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [29,]  0.0000000000    0    0  0.00000000     0  0.00000000     0     0     0
#> [30,]  0.7910940142    0    0  1.09281353     0  1.09952147     0     0     0
#> [31,] -0.0000438803    0    0 -0.38215385     0 -0.21076157     0     0     0
#>       [,16]       [,17] [,18]      [,19]      [,20]      [,21] [,22]      [,23]
#>  [1,]     0 -0.30502299     0  0.1017564 -0.3843071 -0.1231947     0 -0.1064383
#>  [2,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#>  [3,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#>  [4,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#>  [5,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#>  [6,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#>  [7,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#>  [8,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#>  [9,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [10,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [11,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [12,]     0 -0.93372168     0  0.1258682 -0.9928230 -0.1930089     0 -0.1476721
#> [13,]     0  0.27997191     0 -0.9044060  1.1941457  0.9399950     0  0.8785360
#> [14,]     0  0.24442505     0  0.9705861 -0.7082031 -0.9307565     0 -0.9229371
#> [15,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [16,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [17,]     0  0.17702250     0 -0.4057374  0.5355689  0.3979123     0  0.3925822
#> [18,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [19,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [20,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [21,]     0  0.03960971     0 -0.2611706  0.3116140  0.2713797     0  0.2528367
#> [22,]     0 -0.24090567     0  2.0673070 -2.1592758 -2.0051359     0 -1.9841101
#> [23,]     0 -0.21435691     0  0.4503609 -0.6892534 -0.4878632     0 -0.4412087
#> [24,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [25,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [26,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [27,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [28,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [29,]     0  0.00000000     0  0.0000000  0.0000000  0.0000000     0  0.0000000
#> [30,]     0 -1.13200770     0 -0.8355795 -0.1270332  0.8030906     0  0.7756341
#> [31,]     0  0.33776917     0 -0.1402699  0.8101211  0.3689260     0  0.1669026
#>       [,24] [,25] [,26] [,27]       [,28]       [,29]       [,30] [,31] [,32]
#>  [1,]     0     0     0     0 -0.36204257  0.11193143  0.27812650     0     0
#>  [2,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#>  [3,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#>  [4,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#>  [5,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#>  [6,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#>  [7,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#>  [8,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#>  [9,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [10,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [11,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [12,]     0     0     0     0 -0.93763415  0.15565841  0.85852346     0     0
#> [13,]     0     0     0     0  1.02385072 -0.91181607 -0.19802800     0     0
#> [14,]     0     0     0     0 -0.71073025  0.97502932 -0.24704146     0     0
#> [15,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [16,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [17,]     0     0     0     0  0.57056887 -0.42025891 -0.16826858     0     0
#> [18,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [19,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [20,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [21,]     0     0     0     0  0.25231262 -0.26100898 -0.01564785     0     0
#> [22,]     0     0     0     0 -2.26305727  2.11180283  0.21606737     0     0
#> [23,]     0     0     0     0 -0.54153300  0.45137725  0.15134247     0     0
#> [24,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [25,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [26,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [27,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [28,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [29,]     0     0     0     0  0.00000000  0.00000000  0.00000000     0     0
#> [30,]     0     0     0     0 -0.36485391 -0.78681542  1.14477604     0     0
#> [31,]     0     0     0     0 -0.06512377 -0.08018904 -0.06947902     0     0
#>             [,33] [,34]       [,35]
#>  [1,] -0.27320589     0  0.09492051
#>  [2,]  0.00000000     0  0.00000000
#>  [3,]  0.00000000     0  0.00000000
#>  [4,]  0.00000000     0  0.00000000
#>  [5,]  0.00000000     0  0.00000000
#>  [6,]  0.00000000     0  0.00000000
#>  [7,]  0.00000000     0  0.00000000
#>  [8,]  0.00000000     0  0.00000000
#>  [9,]  0.00000000     0  0.00000000
#> [10,]  0.00000000     0  0.00000000
#> [11,]  0.00000000     0  0.00000000
#> [12,] -0.83926417     0  0.11096808
#> [13,]  0.24113520     0 -0.86739966
#> [14,]  0.23759367     0  0.94465972
#> [15,]  0.00000000     0  0.00000000
#> [16,]  0.00000000     0  0.00000000
#> [17,]  0.15058287     0 -0.39392557
#> [18,]  0.00000000     0  0.00000000
#> [19,]  0.00000000     0  0.00000000
#> [20,]  0.00000000     0  0.00000000
#> [21,]  0.03289771     0 -0.25038632
#> [22,] -0.17535257     0  2.00970408
#> [23,] -0.18963509     0  0.42826748
#> [24,]  0.00000000     0  0.00000000
#> [25,]  0.00000000     0  0.00000000
#> [26,]  0.00000000     0  0.00000000
#> [27,]  0.00000000     0  0.00000000
#> [28,]  0.00000000     0  0.00000000
#> [29,]  0.00000000     0  0.00000000
#> [30,] -1.02502193     0 -0.81042623
#> [31,]  0.33132706     0 -0.09169535
#> 
#> 
#> $BIC
#> [1] 6883.501
#>