Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr()
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1),
alt_iter = 1, anneal_iter = 3, em_iter = 3)
#> mixedLSR Start: 1 
#> EM Step...
#> Simulated Annealing Step.
#> Full Cycle 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2917.416
#> 
#> $assign
#>   [1] 1 2 2 2 2 2 2 1 2 1 1 1 2 2 1 1 2 2 2 1 1 1 1 1 2 2 1 1 1 2 1 2 2 1 1 1 1
#>  [38] 1 1 1 1 2 1 2 1 1 2 2 2 1 1 1 2 2 1 2 1 2 1 2 1 2 2 2 1 2 1 2 2 2 2 1 2 2
#>  [75] 1 1 1 1 1 2 1 1 2 2 2 1 2 1 1 1 1 1 1 1 2 1 2 2 1 2
#> 
#> $a
#> $a[[1]]
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,] -1.575502e-17  1.462792e-17  2.352170e-17  0.1330470555 -1.299932e-01
#>  [2,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#>  [3,] -6.603431e-18  3.719863e-17 -3.504583e-17  0.1026683490 -1.020311e-01
#>  [4,]  1.376283e-16 -9.500922e-17 -2.116817e-16 -2.3086985127  2.289677e+00
#>  [5,]  1.395417e-17 -8.224289e-17  3.800146e-17  0.9778288472 -9.874638e-01
#>  [6,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#>  [7,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#>  [8,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#>  [9,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#> [10,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#> [11,]  3.899343e-18  6.558648e-18 -2.048673e-17 -0.0189393640  1.798257e-02
#> [12,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#> [13,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#> [14,] -7.351085e-17  2.732966e-16 -1.545216e-16 -0.0022070825  1.871227e-02
#> [15,] -1.161634e-16 -6.439438e-17  3.539171e-16  2.7103773537 -2.698160e+00
#> [16,] -3.534201e-18  6.269626e-18  1.674749e-18  0.0135109607 -1.274248e-02
#> [17,] -7.358631e-19 -3.573821e-18  7.239761e-18  0.0001329462  9.534358e-05
#> [18,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#> [19,]  1.575177e-17 -3.977314e-17  4.919837e-18  0.0588934478 -6.302824e-02
#> [20,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#> [21,]  5.439861e-17  7.811903e-17 -2.683011e-16 -0.2311274271  2.176849e-01
#> [22,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#> [23,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#> [24,]  4.752917e-17  3.041180e-17 -1.913548e-16  0.0788639813 -9.218086e-02
#> [25,] -1.002863e-16  4.615608e-17  1.894971e-16  1.5922977380 -1.577575e+00
#> [26,]  6.500216e-17 -2.703457e-16  1.722053e-16  0.1297169983 -1.449104e-01
#> [27,] -7.633110e-17 -3.647106e-16  7.423031e-16  0.0229212081  6.346563e-04
#> [28,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000  0.000000e+00
#>              [,6]          [,7] [,8]          [,9] [,10]       [,11] [,12]
#>  [1,]  0.03478529 -0.1320709433    0 -0.1325916202     0 -0.03976431     0
#>  [2,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#>  [3,]  0.12107220 -0.1047805371    0 -0.1054335566     0 -0.12251041     0
#>  [4,] -0.30831798  2.2999720836    0  2.3164409419     0  0.36194591     0
#>  [5,] -0.15109660 -0.9755394213    0 -0.9865325624     0  0.14457634     0
#>  [6,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#>  [7,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#>  [8,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#>  [9,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#> [10,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#> [11,]  0.02614749  0.0178745281    0  0.0178752057     0 -0.02469177     0
#> [12,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#> [13,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#> [14,]  0.77408137 -0.0079606860    0 -0.0036068792     0 -0.78566349     0
#> [15,] -0.13966919 -2.6932940173    0 -2.7153152001     0  0.08364304     0
#> [16,]  0.01601047 -0.0134916292    0 -0.0134344737     0 -0.01690028     0
#> [17,] -0.01293368  0.0002574157    0  0.0002890939     0  0.01260952     0
#> [18,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#> [19,] -0.09853159 -0.0583013207    0 -0.0598082838     0  0.10116820     0
#> [20,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#> [21,]  0.32412069  0.2172483495    0  0.2168934486     0 -0.30451278     0
#> [22,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#> [23,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#> [24,]  0.18647703 -0.0886144582    0 -0.0918929097     0 -0.17280539     0
#> [25,]  0.14018249 -1.5839529127    0 -1.5950595147     0 -0.17930578     0
#> [26,] -0.76609465 -0.1187502846    0 -0.1239694628     0  0.77438567     0
#> [27,] -1.32103111  0.0170838037    0  0.0202850754     0  1.28752992     0
#> [28,]  0.00000000  0.0000000000    0  0.0000000000     0  0.00000000     0
#>       [,13] [,14]        [,15]       [,16] [,17] [,18] [,19]         [,20]
#>  [1,]     0     0 -0.169629982  0.03367932     0     0     0  0.1284517611
#>  [2,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#>  [3,]     0     0 -0.225769657  0.12031784     0     0     0  0.1024203157
#>  [4,]     0     0  2.651463725 -0.30260980     0     0     0 -2.2747441403
#>  [5,]     0     0 -0.842457523 -0.14580900     0     0     0  0.9861434649
#>  [6,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#>  [7,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#>  [8,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#>  [9,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#> [10,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#> [11,]     0     0 -0.007158024  0.02632173     0     0     0 -0.0172614244
#> [12,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#> [13,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#> [14,]     0     0 -0.773015931  0.76415846     0     0     0 -0.0187787761
#> [15,]     0     0 -2.610308950 -0.14014567     0     0     0  2.6804307290
#> [16,]     0     0 -0.029697208  0.01567085     0     0     0  0.0125028211
#> [17,]     0     0  0.012885764 -0.01294444     0     0     0 -0.0003093113
#> [18,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#> [19,]     0     0  0.038606438 -0.09648957     0     0     0  0.0636591730
#> [20,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#> [21,]     0     0 -0.092640090  0.32683917     0     0     0 -0.2080869835
#> [22,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#> [23,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#> [24,]     0     0 -0.269155326  0.19023678     0     0     0  0.0991996707
#> [25,]     0     0 -1.755707075  0.13607092     0     0     0  1.5659427454
#> [26,]     0     0  0.635691238 -0.75653024     0     0     0  0.1439597106
#> [27,]     0     0  1.306682138 -1.32220245     0     0     0 -0.0226295391
#> [28,]     0     0  0.000000000  0.00000000     0     0     0  0.0000000000
#>             [,21]        [,22] [,23] [,24] [,25]         [,26]         [,27]
#>  [1,]  0.03829017  0.165319296     0     0     0 -0.1327886195  1.285879e-01
#>  [2,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#>  [3,]  0.12383535  0.225477239     0     0     0 -0.1012669860  1.003396e-01
#>  [4,] -0.34872748 -2.621591026     0     0     0  2.3046928020 -2.261914e+00
#>  [5,] -0.14588890  0.852313053     0     0     0 -0.9778180486  9.744571e-01
#>  [6,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#>  [7,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#>  [8,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#>  [9,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#> [10,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#> [11,]  0.02569120  0.008710258     0     0     0  0.0192954700 -1.797680e-02
#> [12,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#> [13,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#> [14,]  0.79027556  0.755582492     0     0     0  0.0095201846 -1.956236e-02
#> [15,] -0.10244148  2.585735077     0     0     0 -2.7104637623  2.666214e+00
#> [16,]  0.01672669  0.028712369     0     0     0 -0.0133828623  1.261142e-02
#> [17,] -0.01294409 -0.013291890     0     0     0 -0.0002971552 -1.472753e-05
#> [18,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#> [19,] -0.10093860 -0.033719286     0     0     0 -0.0596002971  6.212525e-02
#> [20,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#> [21,]  0.31769492  0.114060730     0     0     0  0.2356772026 -2.178715e-01
#> [22,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#> [23,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#> [24,]  0.18243920  0.288677168     0     0     0 -0.0755843578  8.865599e-02
#> [25,]  0.16815208  1.732382240     0     0     0 -1.5904827851  1.558932e+00
#> [26,] -0.78001319 -0.620331898     0     0     0 -0.1369322340  1.442983e-01
#> [27,] -1.32185957 -1.348516472     0     0     0 -0.0397046222  7.513698e-03
#> [28,]  0.00000000  0.000000000     0     0     0  0.0000000000  0.000000e+00
#>             [,28]       [,29] [,30] [,31]        [,32] [,33]       [,34] [,35]
#>  [1,]  0.03509039  0.16827670     0     0  0.034061885     0  0.03483700     0
#>  [2,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#>  [3,]  0.12251864  0.22677430     0     0  0.122669876     0  0.12164545     0
#>  [4,] -0.32922069 -2.62814279     0     0 -0.313058630     0 -0.32411541     0
#>  [5,] -0.13460793  0.83027377     0     0 -0.141582684     0 -0.13638722     0
#>  [6,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#>  [7,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#>  [8,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#>  [9,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#> [10,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#> [11,]  0.02648138  0.00801930     0     0  0.026934283     0  0.02629706     0
#> [12,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#> [13,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#> [14,]  0.76890182  0.78474959     0     0  0.773063972     0  0.76554323     0
#> [15,] -0.11207808  2.57524547     0     0 -0.133807862     0 -0.11585038     0
#> [16,]  0.01586360  0.02971636     0     0  0.015803299     0  0.01579750     0
#> [17,] -0.01309443 -0.01316720     0     0 -0.013224161     0 -0.01300224     0
#> [18,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#> [19,] -0.09613104 -0.04041744     0     0 -0.096835676     0 -0.09596821     0
#> [20,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#> [21,]  0.32908542  0.10354012     0     0  0.334860567     0  0.32667944     0
#> [22,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#> [23,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#> [24,]  0.19475512  0.27398347     0     0  0.197237120     0  0.19254146     0
#> [25,]  0.15352919  1.73800012     0     0  0.141622500     0  0.15054944     0
#> [26,] -0.75992429 -0.64892561     0     0 -0.765081095     0 -0.75677571     0
#> [27,] -1.33744991 -1.33557227     0     0 -1.350794376     0 -1.32803528     0
#> [28,]  0.00000000  0.00000000     0     0  0.000000000     0  0.00000000     0
#>  [ reached getOption("max.print") -- omitted 3 rows ]
#> 
#> $a[[2]]
#>                [,1]          [,2]          [,3]         [,4]         [,5]
#>  [1,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000000  0.000000000
#>  [2,]  2.099207e-17 -1.207641e-16  2.166574e-16  0.147799977 -0.150227422
#>  [3,] -1.587079e-18 -3.491342e-16  6.633907e-17  0.169689664 -0.162080581
#>  [4,] -3.145631e-17  4.494977e-16 -1.838799e-15 -1.147047789  1.139718189
#>  [5,]  7.890578e-18 -3.527103e-16  3.959567e-16  0.343204815 -0.336841706
#>  [6,] -1.353494e-17  1.042239e-16 -1.949632e-16 -0.135360946  0.136196792
#>  [7,]  5.167979e-18  3.283977e-16 -5.940305e-16 -0.451611779  0.441491072
#>  [8,]  9.068552e-18 -7.850063e-17  3.602151e-16  0.218883350 -0.218417169
#>  [9,] -2.242180e-18  1.470160e-15  6.322898e-17 -0.521425992  0.493054145
#> [10,] -1.628571e-17  1.773650e-16 -5.716648e-16 -0.365955551  0.364675181
#> [11,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000000  0.000000000
#> [12,]  2.089982e-18 -1.455943e-17 -2.319452e-16 -0.122243183  0.121108925
#> [13,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000000  0.000000000
#> [14,] -4.647919e-17  4.201688e-16 -7.486923e-16 -0.531522094  0.532860863
#> [15,] -3.128772e-17  3.924063e-16  5.717174e-16  0.185866902 -0.183141409
#> [16,]  1.538474e-18 -5.795478e-18  1.056331e-17  0.006781841 -0.007040212
#> [17,]  1.017121e-17  1.207887e-16 -5.906666e-16 -0.374832681  0.367506723
#> [18,] -9.795763e-18 -4.083141e-17  1.303673e-16  0.093810912 -0.089886368
#> [19,]  1.127228e-17 -5.557233e-17  5.792543e-17  0.044068518 -0.045773559
#> [20,]  5.570461e-18 -2.985012e-17  4.971148e-17  0.034154533 -0.034871248
#> [21,] -2.088919e-18  1.899286e-16 -8.852551e-16 -0.552051410  0.545545715
#> [22,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000000  0.000000000
#> [23,]  1.271895e-17 -1.233509e-16 -3.789675e-16 -0.169004122  0.166632463
#> [24,]  1.026426e-17 -8.956616e-17  2.151884e-16  0.143277250 -0.143451491
#> [25,]  6.384979e-18  2.156060e-16 -2.961234e-16 -0.247688043  0.240589780
#> [26,] -5.552223e-17  1.037958e-15  1.633920e-16 -0.262767182  0.257650915
#> [27,] -5.367015e-17  1.076838e-15 -1.808270e-16 -0.466176483  0.458510456
#> [28,]  5.841868e-18 -6.927720e-17  1.936202e-17  0.032390090 -0.032509191
#>                [,6]         [,7] [,8]         [,9] [,10]        [,11] [,12]
#>  [1,]  0.0000000000  0.000000000    0  0.000000000     0  0.000000000     0
#>  [2,]  0.0616201085 -0.141991251    0 -0.147479340     0 -0.078012817     0
#>  [3,]  0.8904447689 -0.171853384    0 -0.172708166     0 -0.875386359     0
#>  [4,]  0.1326181143  1.131983384    0  1.146525785     0 -0.137382618     0
#>  [5,]  0.6407144826 -0.341217155    0 -0.345269351     0 -0.631840424     0
#>  [6,] -0.0752449651  0.131483373    0  0.135297093     0  0.084155110     0
#>  [7,] -0.5312227972  0.451789859    0  0.453794956     0  0.505996316     0
#>  [8,] -0.0683301734 -0.215127694    0 -0.218554775     0  0.065667868     0
#>  [9,] -3.9083271693  0.530048631    0  0.534236668     0  3.859392198     0
#> [10,] -0.0198007979  0.359826860    0  0.365801257     0  0.024270751     0
#> [11,]  0.0000000000  0.000000000    0  0.000000000     0  0.000000000     0
#> [12,]  0.1692346941  0.121470186    0  0.121853921     0 -0.174088519     0
#> [13,]  0.0000000000  0.000000000    0  0.000000000     0  0.000000000     0
#> [14,] -0.3739674736  0.518164609    0  0.531721169     0  0.401316787     0
#> [15,] -1.2047305731 -0.188528892    0 -0.183084043     0  1.233298145     0
#> [16,] -0.0002731032 -0.006365716    0 -0.006740064     0 -0.001106356     0
#> [17,] -0.0144946796  0.375062405    0  0.375456722     0 -0.008911146     0
#> [18,]  0.0850992415 -0.095887734    0 -0.094443735     0 -0.070970326     0
#> [19,]  0.0446315600 -0.041201270    0 -0.043907404     0 -0.054527010     0
#> [20,]  0.0153527059 -0.032640891    0 -0.034062546     0 -0.019882403     0
#> [21,]  0.0568669475  0.547994607    0  0.552211521     0 -0.072844834     0
#> [22,]  0.0000000000  0.000000000    0  0.000000000     0  0.000000000     0
#> [23,]  0.4846401935  0.169532030    0  0.167957578     0 -0.499453964     0
#> [24,]  0.0433218112 -0.140045300    0 -0.143214641     0 -0.048824829     0
#> [25,] -0.4243095721  0.249068688    0  0.249416369     0  0.405621982     0
#> [26,] -2.5149101300  0.255617726    0  0.269344907     0  2.542339618     0
#> [27,] -2.4161374531  0.457923472    0  0.472642563     0  2.435612781     0
#> [28,]  0.1366001694 -0.031204396    0 -0.032676372     0 -0.140428539     0
#>       [,13] [,14]        [,15]         [,16] [,17] [,18] [,19]        [,20]
#>  [1,]     0     0  0.000000000  0.0000000000     0     0     0  0.000000000
#>  [2,]     0     0 -0.202399142  0.0545500535     0     0     0  0.151364042
#>  [3,]     0     0 -1.045171584  0.8762077808     0     0     0  0.169323771
#>  [4,]     0     0  0.973627938  0.1673748110     0     0     0 -1.135082062
#>  [5,]     0     0 -0.964858782  0.6231904322     0     0     0  0.341571139
#>  [6,]     0     0  0.204306820 -0.0691562470     0     0     0 -0.137068110
#>  [7,]     0     0  0.961816084 -0.5131070859     0     0     0 -0.443936259
#>  [8,]     0     0 -0.143032092 -0.0748187233     0     0     0  0.217323821
#>  [9,]     0     0  4.371417616 -3.8512997398     0     0     0 -0.526432663
#> [10,]     0     0  0.371750972 -0.0074018493     0     0     0 -0.364137584
#> [11,]     0     0  0.000000000  0.0000000000     0     0     0  0.000000000
#> [12,]     0     0 -0.049449392  0.1707168502     0     0     0 -0.118873326
#> [13,]     0     0  0.000000000  0.0000000000     0     0     0  0.000000000
#> [14,]     0     0  0.880332620 -0.3499073710     0     0     0 -0.536598593
#> [15,]     0     0  1.011344525 -1.1939419214     0     0     0  0.169412323
#> [16,]     0     0 -0.006185567 -0.0006237357     0     0     0  0.007098576
#> [17,]     0     0  0.376564465 -0.0046298040     0     0     0 -0.365292900
#> [18,]     0     0 -0.175296359  0.0825070958     0     0     0  0.089640828
#> [19,]     0     0 -0.086054525  0.0417248061     0     0     0  0.046652696
#> [20,]     0     0 -0.047832631  0.0136327022     0     0     0  0.035187211
#> [21,]     0     0  0.476216716  0.0723214716     0     0     0 -0.542595267
#> [22,]     0     0  0.000000000  0.0000000000     0     0     0  0.000000000
#> [23,]     0     0 -0.315920244  0.4829915433     0     0     0 -0.160665115
#> [24,]     0     0 -0.180532444  0.0376852438     0     0     0  0.143798729
#> [25,]     0     0  0.659315452 -0.4133176574     0     0     0 -0.242931010
#> [26,]     0     0  2.739632805 -2.4747184140     0     0     0 -0.282709436
#> [27,]     0     0  2.838222881 -2.3713549295     0     0     0 -0.481614603
#> [28,]     0     0 -0.166125875  0.1336051769     0     0     0  0.033955850
#>               [,21]        [,22] [,23] [,24] [,25]        [,26]        [,27]
#>  [1,]  0.000000e+00  0.000000000     0     0     0  0.000000000  0.000000000
#>  [2,]  6.465244e-02  0.199190034     0     0     0 -0.131796007  0.142736388
#>  [3,]  8.831531e-01  1.048863490     0     0     0 -0.170105348  0.168583880
#>  [4,]  1.311471e-01 -0.956444349     0     0     0  1.126506208 -1.124586330
#>  [5,]  6.361687e-01  0.964102085     0     0     0 -0.337184715  0.337716993
#>  [6,] -7.675570e-02 -0.201984782     0     0     0  0.125120399 -0.131520615
#>  [7,] -5.242801e-01 -0.960236665     0     0     0  0.456327019 -0.446020381
#>  [8,] -6.714083e-02  0.139215143     0     0     0 -0.212635314  0.214118054
#>  [9,] -3.879278e+00 -4.389461132     0     0     0  0.515310779 -0.518836195
#> [10,] -2.109143e-02 -0.366108224     0     0     0  0.354421970 -0.358038926
#> [11,]  0.000000e+00  0.000000000     0     0     0  0.000000000  0.000000000
#> [12,]  1.690980e-01  0.051772101     0     0     0  0.124548313 -0.120350429
#> [13,]  0.000000e+00  0.000000000     0     0     0  0.000000000  0.000000000
#> [14,] -3.780271e-01 -0.872290470     0     0     0  0.496397966 -0.517461608
#> [15,] -1.203620e+00 -1.018328492     0     0     0 -0.212813481  0.185355863
#> [16,]  2.306761e-06  0.005968454     0     0     0 -0.005602899  0.006467107
#> [17,] -1.049138e-02 -0.373105513     0     0     0  0.384007644 -0.370342624
#> [18,]  8.208810e-02  0.175576077     0     0     0 -0.101644453  0.093789577
#> [19,]  4.635664e-02  0.084821740     0     0     0 -0.035338918  0.041921170
#> [20,]  1.618953e-02  0.047034856     0     0     0 -0.029892371  0.032888896
#> [21,]  5.881435e-02 -0.469135526     0     0     0  0.552350468 -0.543014593
#> [22,]  0.000000e+00  0.000000000     0     0     0  0.000000000  0.000000000
#> [23,]  4.846903e-01  0.319834921     0     0     0  0.180262215 -0.167336708
#> [24,]  4.435921e-02  0.178204198     0     0     0 -0.135668337  0.139706236
#> [25,] -4.188775e-01 -0.659591601     0     0     0  0.252840544 -0.245301804
#> [26,] -2.507386e+00 -2.747359007     0     0     0  0.216337441 -0.254937802
#> [27,] -2.407907e+00 -2.843083828     0     0     0  0.422026832 -0.455254060
#> [28,]  1.366657e-01  0.166082764     0     0     0 -0.027720641  0.031298560
#>              [,28]        [,29] [,30] [,31]         [,32] [,33]        [,34]
#>  [1,]  0.000000000  0.000000000     0     0  0.0000000000     0  0.000000000
#>  [2,]  0.050267406  0.204432882     0     0  0.0583149773     0  0.032522322
#>  [3,]  0.894245192  1.026210435     0     0  0.8753485492     0  0.897874996
#>  [4,]  0.107891529 -0.939880340     0     0  0.1612884866     0  0.151602732
#>  [5,]  0.645601692  0.945683272     0     0  0.6242595413     0  0.637735187
#>  [6,] -0.069480610 -0.203465546     0     0 -0.0715711070     0 -0.057569607
#>  [7,] -0.552331623 -0.933688710     0     0 -0.5119462175     0 -0.551663148
#>  [8,] -0.065925333  0.138455954     0     0 -0.0730774615     0 -0.076802456
#>  [9,] -3.907570490 -4.301743993     0     0 -3.8491098225     0 -3.920467102
#> [10,] -0.022141670 -0.362929783     0     0 -0.0104201131     0 -0.003925193
#> [11,]  0.000000000  0.000000000     0     0  0.0000000000     0  0.000000000
#> [12,]  0.161207561  0.053490559     0     0  0.1709344534     0  0.162752972
#> [13,]  0.000000000  0.000000000     0     0  0.0000000000     0  0.000000000
#> [14,] -0.356448282 -0.873162137     0     0 -0.3581392474     0 -0.315102730
#> [15,] -1.163863433 -1.017975445     0     0 -1.1986438660     0 -1.151012239
#> [16,] -0.001255709  0.006567053     0     0 -0.0003456895     0 -0.002515937
#> [17,] -0.039279022 -0.355940058     0     0 -0.0029123983     0 -0.040536029
#> [18,]  0.096885619  0.165978217     0     0  0.0807203601     0  0.103075614
#> [19,]  0.036927248  0.088513116     0     0  0.0437274430     0  0.028069203
#> [20,]  0.012124095  0.048601274     0     0  0.0146309004     0  0.007495579
#> [21,]  0.034140587 -0.454221266     0     0  0.0717363032     0  0.045440192
#> [22,]  0.000000000  0.000000000     0     0  0.0000000000     0  0.000000000
#> [23,]  0.463967543  0.322674763     0     0  0.4848776586     0  0.460304961
#> [24,]  0.040803878  0.178327256     0     0  0.0395388000     0  0.030996474
#> [25,] -0.438104461 -0.640702148     0     0 -0.4119509106     0 -0.441173448
#> [26,] -2.466218697 -2.721395456     0     0 -2.4829799895     0 -2.435577901
#> [27,] -2.378434515 -2.811705255     0     0 -2.3794170058     0 -2.345111281
#> [28,]  0.132395227  0.165587850     0     0  0.1345660699     0  0.128322525
#>       [,35]
#>  [1,]     0
#>  [2,]     0
#>  [3,]     0
#>  [4,]     0
#>  [5,]     0
#>  [6,]     0
#>  [7,]     0
#>  [8,]     0
#>  [9,]     0
#> [10,]     0
#> [11,]     0
#> [12,]     0
#> [13,]     0
#> [14,]     0
#> [15,]     0
#> [16,]     0
#> [17,]     0
#> [18,]     0
#> [19,]     0
#> [20,]     0
#> [21,]     0
#> [22,]     0
#> [23,]     0
#> [24,]     0
#> [25,]     0
#> [26,]     0
#> [27,]     0
#> [28,]     0
#>  [ reached getOption("max.print") -- omitted 3 rows ]
#> 
#> 
#> $BIC
#> [1] 9887.381
#>