Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr()
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1),
alt_iter = 1, anneal_iter = 3, em_iter = 3)
#> mixedLSR Start: 1 
#> EM Step...
#> Simulated Annealing Step
#> Full Cycle 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -4050.661
#> 
#> $assign
#>   [1] 2 2 1 2 1 2 2 2 1 2 2 2 2 1 1 1 2 1 1 2 1 1 1 2 2 2 1 1 1 1 1 2 2 1 1 1 1
#>  [38] 2 1 1 2 2 1 2 1 2 2 2 1 1 2 2 2 1 1 2 2 1 1 1 1 1 2 1 2 2 1 2 2 2 1 1 1 2
#>  [75] 2 1 2 1 1 1 2 1 2 2 2 1 2 1 2 2 1 1 2 2 2 1 2 1 2 1
#> 
#> $a
#> $a[[1]]
#>                [,1]         [,2]          [,3]          [,4]         [,5]
#>  [1,] -0.2384644917 -0.089223230  0.2416587418 -0.2322152172  0.138732016
#>  [2,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#>  [3,] -0.0239707045 -0.132365159  0.0256196374 -0.0223661399 -0.107648786
#>  [4,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#>  [5,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#>  [6,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#>  [7,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#>  [8,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#>  [9,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [10,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [11,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [12,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [13,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [14,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [15,] -1.0273640521 -0.572403371  1.0357320984 -1.0054999182  0.429042333
#> [16,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [17,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [18,]  0.2259171167  0.225190669 -0.2294733524  0.2197518964  0.004972521
#> [19,] -0.8317015744  0.471743612  0.8125557212 -0.8353985525  1.304325053
#> [20,] -0.0758171456 -0.214265547  0.0791373712 -0.0720907701 -0.139756316
#> [21,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [22,]  0.6615086998 -0.991159641 -0.6404156753  0.6686493011 -1.642663665
#> [23,]  2.3610619962  0.260572487 -2.3468820552  2.3386339483 -2.074929271
#> [24,] -0.2855028348 -0.383258313  0.2854103529 -0.2819168427 -0.090864649
#> [25,] -0.1999774147  0.352949430  0.1976079455 -0.1985143401  0.538863355
#> [26,]  0.0137772779 -0.233164082 -0.0100837718  0.0164614462 -0.245795989
#> [27,]  0.1084140247 -0.001283897 -0.1074923180  0.1076020344 -0.108617554
#> [28,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [29,]  0.0000000000  0.000000000  0.0000000000  0.0000000000  0.000000000
#> [30,]  0.0005055456 -0.002220613 -0.0004181483  0.0005716002 -0.002821532
#> [31,]  0.2184917398  0.153726821 -0.2607292728  0.1781793335  0.030129396
#>                [,6] [,7]          [,8]         [,9] [,10] [,11] [,12] [,13]
#>  [1,] -0.2324603287    0 -0.2101588637  0.094496364     0     0     0     0
#>  [2,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#>  [3,] -0.0220722100    0 -0.0223976980  0.134003330     0     0     0     0
#>  [4,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#>  [5,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#>  [6,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#>  [7,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#>  [8,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#>  [9,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [10,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [11,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [12,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [13,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [14,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [15,] -1.0109376650    0 -0.9592883649  0.586933217     0     0     0     0
#> [16,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [17,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [18,]  0.2202664302    0  0.2074384443 -0.230139718     0     0     0     0
#> [19,] -0.8526211414    0 -0.8779866966 -0.489495130     0     0     0     0
#> [20,] -0.0714891205    0 -0.0666040794  0.218016601     0     0     0     0
#> [21,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [22,]  0.6834370077    0  0.6866281924  1.009793746     0     0     0     0
#> [23,]  2.3683305725    0  2.3482013094 -0.255115859     0     0     0     0
#> [24,] -0.2860190971    0 -0.3027198690  0.382876961     0     0     0     0
#> [25,] -0.2001162910    0 -0.1749456003 -0.352961385     0     0     0     0
#> [26,]  0.0179528721    0  0.0188824906  0.236748745     0     0     0     0
#> [27,]  0.1090838525    0  0.1085834064  0.001823921     0     0     0     0
#> [28,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [29,]  0.0000000000    0  0.0000000000  0.000000000     0     0     0     0
#> [30,]  0.0006231216    0  0.0008985115  0.002321943     0     0     0     0
#> [31,]  0.1529290729    0 -0.0765416579 -0.210521590     0     0     0     0
#>       [,14]         [,15]         [,16] [,17]        [,18] [,19] [,20] [,21]
#>  [1,]     0 -0.2279997693  0.0854236243     0 -0.223060675     0     0     0
#>  [2,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#>  [3,]     0 -0.0204807119  0.1314732267     0 -0.023075281     0     0     0
#>  [4,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#>  [5,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#>  [6,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#>  [7,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#>  [8,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#>  [9,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [10,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [11,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [12,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [13,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [14,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [15,]     0 -0.9904504396  0.5636898306     0 -0.987125792     0     0     0
#> [16,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [17,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [18,]     0  0.2148000554 -0.2221280140     0  0.215400865     0     0     0
#> [19,]     0 -0.8457069816 -0.4569021312     0 -0.847908675     0     0     0
#> [20,]     0 -0.0683294680  0.2119432634     0 -0.070872539     0     0     0
#> [21,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [22,]     0  0.6835952137  0.9776771760     0  0.668554243     0     0     0
#> [23,]     0  2.3324225983 -0.2699898566     0  2.337750302     0     0     0
#> [24,]     0 -0.2785560639  0.3854026655     0 -0.291874087     0     0     0
#> [25,]     0 -0.2003667384 -0.3538618362     0 -0.186303873     0     0     0
#> [26,]     0  0.0199864301  0.2309050718     0  0.016076960     0     0     0
#> [27,]     0  0.1075659239  0.0006511395     0  0.107707320     0     0     0
#> [28,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [29,]     0  0.0000000000  0.0000000000     0  0.000000000     0     0     0
#> [30,]     0  0.0006403281  0.0021411186     0  0.000691352     0     0     0
#> [31,]     0  0.1457247869 -0.1061303516     0  0.074184660     0     0     0
#>       [,22] [,23] [,24]         [,25]         [,26]        [,27] [,28]
#>  [1,]     0     0     0 -0.2291872601 -0.0792696215  0.096080091     0
#>  [2,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#>  [3,]     0     0     0 -0.0225561528 -0.1311539861  0.133611482     0
#>  [4,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#>  [5,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#>  [6,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#>  [7,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#>  [8,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#>  [9,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [10,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [11,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [12,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [13,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [14,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [15,]     0     0     0 -1.0010385717 -0.5475923049  0.592208696     0
#> [16,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [17,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [18,]     0     0     0  0.2184852934  0.2179598483 -0.230918068     0
#> [19,]     0     0     0 -0.8447826610  0.4509087731 -0.482562216     0
#> [20,]     0     0     0 -0.0715516469 -0.2100291264  0.217719017     0
#> [21,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [22,]     0     0     0  0.6728396497 -0.9754292723  1.001456063     0
#> [23,]     0     0     0  2.3474450787  0.2600577382 -0.269844797     0
#> [24,]     0     0     0 -0.2866752158 -0.3879594649  0.382880730     0
#> [25,]     0     0     0 -0.1947310579  0.3594628415 -0.349765464     0
#> [26,]     0     0     0  0.0167185402 -0.2300749222  0.235718393     0
#> [27,]     0     0     0  0.1080887162 -0.0009674139  0.001100664     0
#> [28,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [29,]     0     0     0  0.0000000000  0.0000000000  0.000000000     0
#> [30,]     0     0     0  0.0006275725 -0.0020683794  0.002314339     0
#> [31,]     0     0     0  0.1350064582  0.0489188329 -0.215330092     0
#>               [,29]         [,30] [,31] [,32]        [,33]        [,34]
#>  [1,]  0.0825479169  0.2277481383     0     0  0.078288312 -0.085977251
#>  [2,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#>  [3,]  0.1342169418  0.0239690297     0     0  0.129878474 -0.133576434
#>  [4,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#>  [5,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#>  [6,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#>  [7,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#>  [8,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#>  [9,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [10,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [11,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [12,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [13,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [14,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [15,]  0.5649535705  1.0069781017     0     0  0.539996500 -0.569099155
#> [16,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [17,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [18,] -0.2240585809 -0.2201925760     0     0 -0.215479762  0.224973897
#> [19,] -0.4595507936  0.8586691308     0     0 -0.451506742  0.468244176
#> [20,]  0.2152273799  0.0730399139     0     0  0.208008769 -0.215207867
#> [21,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [22,]  0.9958517046 -0.6751049719     0     0  0.969868320 -0.996909106
#> [23,] -0.2738316020 -2.3768285065     0     0 -0.247977766  0.264632884
#> [24,]  0.3971464398  0.2975163897     0     0  0.382684676 -0.390563735
#> [25,] -0.3662450892  0.1887107566     0     0 -0.356300472  0.360501461
#> [26,]  0.2353288447 -0.0153570033     0     0  0.228137636 -0.234844506
#> [27,]  0.0006444367 -0.1094392982     0     0  0.001422927 -0.001118607
#> [28,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [29,]  0.0000000000  0.0000000000     0     0  0.000000000  0.000000000
#> [30,]  0.0021233393 -0.0006839632     0     0  0.002061968 -0.002180810
#> [31,] -0.0583340480 -0.0838504737     0     0 -0.055516727  0.108328190
#>               [,35]
#>  [1,]  0.0731795035
#>  [2,]  0.0000000000
#>  [3,]  0.1323605737
#>  [4,]  0.0000000000
#>  [5,]  0.0000000000
#>  [6,]  0.0000000000
#>  [7,]  0.0000000000
#>  [8,]  0.0000000000
#>  [9,]  0.0000000000
#> [10,]  0.0000000000
#> [11,]  0.0000000000
#> [12,]  0.0000000000
#> [13,]  0.0000000000
#> [14,]  0.0000000000
#> [15,]  0.5374176777
#> [16,]  0.0000000000
#> [17,]  0.0000000000
#> [18,] -0.2160261807
#> [19,] -0.4445094933
#> [20,]  0.2102875950
#> [21,]  0.0000000000
#> [22,]  0.9807857318
#> [23,] -0.2606605626
#> [24,]  0.3970932465
#> [25,] -0.3700775649
#> [26,]  0.2315853114
#> [27,]  0.0009055533
#> [28,]  0.0000000000
#> [29,]  0.0000000000
#> [30,]  0.0019945138
#> [31,]  0.0246641724
#> 
#> $a[[2]]
#>               [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>  [2,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>  [3,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>  [4,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>  [5,]  0.334481990  0.471230290 -0.332344292  0.315555111  0.133799616
#>  [6,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>  [7,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>  [8,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>  [9,]  0.248689434 -0.475845231 -0.249236489  0.231936017 -0.727320394
#> [10,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [11,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [12,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [13,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [14,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [15,]  1.199505974 -3.049373007 -1.214708566  1.245436511 -4.214473595
#> [16,]  0.213483092  1.289560430 -0.205636979  0.156840244  1.057013656
#> [17,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [18,]  0.165102307 -0.389487319 -0.165289255  0.149280241 -0.558162610
#> [19,] -0.186256176  0.004796485  0.188167979 -0.204198202  0.181894659
#> [20,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [21,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [22,]  0.896940492  0.916347763 -0.889664542  0.815341544  0.000120487
#> [23,]  0.455291316  0.041026166 -0.456933185  0.464083877 -0.405038288
#> [24,]  0.019126735 -0.001469597 -0.019140558  0.018713975 -0.020500056
#> [25,]  0.423389523  1.275710180 -0.415971374  0.365658380  0.835601373
#> [26,]  0.397547395 -0.066644371 -0.397723657  0.386363079 -0.463148440
#> [27,]  0.005941603  0.003302538 -0.005908331  0.005486649 -0.002733398
#> [28,]  2.598019598  0.381974026 -2.600189217  2.565698931 -2.194364199
#> [29,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [30,]  0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [31,]  0.385275864  0.484420335 -0.351575726 -0.018764940 -0.047382674
#>               [,6] [,7]         [,8]         [,9] [,10] [,11] [,12]
#>  [1,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#>  [2,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#>  [3,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#>  [4,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#>  [5,]  0.340946113    0  0.318792550 -0.475819651     0     0     0
#>  [6,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#>  [7,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#>  [8,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#>  [9,]  0.254938407    0  0.243519682  0.478821421     0     0     0
#> [10,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [11,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [12,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [13,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [14,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [15,]  1.141190918    0  1.218301881  3.102588778     0     0     0
#> [16,]  0.249079317    0  0.181716866 -1.312348983     0     0     0
#> [17,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [18,]  0.172479370    0  0.160943294  0.390891589     0     0     0
#> [19,] -0.169927523    0 -0.188208063 -0.011810529     0     0     0
#> [20,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [21,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [22,]  0.935532601    0  0.848902357 -0.933567181     0     0     0
#> [23,]  0.439771487    0  0.449337000 -0.033297683     0     0     0
#> [24,]  0.019016642    0  0.018675864  0.001621639     0     0     0
#> [25,]  0.455387116    0  0.387662201 -1.295969742     0     0     0
#> [26,]  0.397051138    0  0.387731873  0.069447578     0     0     0
#> [27,]  0.006136369    0  0.005672813 -0.003371027     0     0     0
#> [28,]  2.566860498    0  2.538600728 -0.360230357     0     0     0
#> [29,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [30,]  0.000000000    0  0.000000000  0.000000000     0     0     0
#> [31,]  0.658317003    0  0.255868444 -0.586934498     0     0     0
#>               [,13] [,14]       [,15]        [,16] [,17]       [,18] [,19]
#>  [1,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#>  [2,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#>  [3,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#>  [4,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#>  [5,] -5.352649e-03     0  0.33326909 -0.467248136     0  0.32514464     0
#>  [6,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#>  [7,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#>  [8,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#>  [9,]  6.068046e-03     0  0.25347995  0.481607780     0  0.23781107     0
#> [10,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [11,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [12,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [13,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [14,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [15,]  2.202572e-02     0  1.16141621  3.018891898     0  1.17721629     0
#> [16,] -9.371387e-03     0  0.23044778 -1.267876304     0  0.19974506     0
#> [17,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [18,]  5.490139e-03     0  0.17108220  0.395663161     0  0.15636353     0
#> [19,]  3.460236e-03     0 -0.17233855  0.005900201     0 -0.18745641     0
#> [20,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [21,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [22,] -6.353937e-03     0  0.91135378 -0.890497347     0  0.86247792     0
#> [23,] -4.622541e-03     0  0.43901642 -0.050759392     0  0.44926448     0
#> [24,] -5.729608e-05     0  0.01886167  0.001434961     0  0.01865657     0
#> [25,] -1.046424e-02     0  0.43566460 -1.256031025     0  0.40540032     0
#> [26,] -4.357492e-04     0  0.39356534  0.067218499     0  0.38695266     0
#> [27,] -2.044791e-05     0  0.00600755 -0.003166441     0  0.00572485     0
#> [28,] -1.763289e-02     0  2.54677437 -0.398943758     0  2.54264134     0
#> [29,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [30,]  0.000000e+00     0  0.00000000  0.000000000     0  0.00000000     0
#> [31,]  4.308900e-02     0  0.58062874 -0.299727658     0  0.27436229     0
#>       [,20] [,21] [,22] [,23] [,24]       [,25]         [,26]        [,27]
#>  [1,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#>  [2,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#>  [3,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#>  [4,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#>  [5,]     0     0     0     0     0  0.31219171  0.4675327605 -0.471506076
#>  [6,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#>  [7,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#>  [8,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#>  [9,]     0     0     0     0     0  0.23038932 -0.4810708609  0.481800256
#> [10,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [11,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [12,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [13,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [14,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [15,]     0     0     0     0     0  1.25187242 -3.0391869028  3.078978097
#> [16,]     0     0     0     0     0  0.14873518  1.2762410502 -1.295355507
#> [17,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [18,]     0     0     0     0     0  0.14788860 -0.3944772517  0.394153107
#> [19,]     0     0     0     0     0 -0.20559292 -0.0008974266 -0.005558000
#> [20,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [21,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [22,]     0     0     0     0     0  0.80372698  0.8967300905 -0.913407257
#> [23,]     0     0     0     0     0  0.46387949  0.0448327446 -0.038264769
#> [24,]     0     0     0     0     0  0.01863126 -0.0015441908  0.001637735
#> [25,]     0     0     0     0     0  0.35698919  1.2625798933 -1.279803045
#> [26,]     0     0     0     0     0  0.38443471 -0.0690281493  0.070499563
#> [27,]     0     0     0     0     0  0.00542244  0.0031918493 -0.003265894
#> [28,]     0     0     0     0     0  2.55599604  0.3799556308 -0.364273997
#> [29,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [30,]     0     0     0     0     0  0.00000000  0.0000000000  0.000000000
#> [31,]     0     0     0     0     0 -0.06109227  0.3686357519 -0.470281191
#>       [,28]        [,29]        [,30] [,31] [,32]         [,33]         [,34]
#>  [1,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#>  [2,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#>  [3,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#>  [4,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#>  [5,]     0 -0.469006323 -0.331658428     0     0 -0.4676161913  0.4671993640
#>  [6,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#>  [7,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#>  [8,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#>  [9,]     0  0.482264545 -0.242586245     0     0  0.4814570034 -0.4888834228
#> [10,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [11,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [12,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [13,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [14,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [15,]     0  3.098556628 -1.230652473     0     0  3.0415792670 -3.0805259145
#> [16,]     0 -1.298973905 -0.192717639     0     0 -1.2768753235  1.2838801451
#> [17,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [18,]     0  0.393694487 -0.158475664     0     0  0.3947693504 -0.4004497995
#> [19,]     0 -0.010727368  0.197982992     0     0  0.0007715098  0.0009965765
#> [20,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [21,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [22,]     0 -0.911571581 -0.872908189     0     0 -0.8967932065  0.8919088623
#> [23,]     0 -0.030616189 -0.466690424     0     0 -0.0445865003  0.0397997547
#> [24,]     0  0.001842471 -0.019206361     0     0  0.0015555462 -0.0018042719
#> [25,]     0 -1.280675953 -0.405216649     0     0 -1.2630875726  1.2674719519
#> [26,]     0  0.074335625 -0.397790296     0     0  0.0692814259 -0.0748270011
#> [27,]     0 -0.003244781 -0.005816016     0     0 -0.0031906777  0.0031369502
#> [28,]     0 -0.332333648 -2.622523000     0     0 -0.3787082600  0.3505030676
#> [29,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [30,]     0  0.000000000  0.000000000     0     0  0.0000000000  0.0000000000
#> [31,]     0 -0.522543065 -0.191760710     0     0 -0.3690573030  0.3581087545
#>              [,35]
#>  [1,]  0.000000000
#>  [2,]  0.000000000
#>  [3,]  0.000000000
#>  [4,]  0.000000000
#>  [5,] -0.469368331
#>  [6,]  0.000000000
#>  [7,]  0.000000000
#>  [8,]  0.000000000
#>  [9,]  0.475811043
#> [10,]  0.000000000
#> [11,]  0.000000000
#> [12,]  0.000000000
#> [13,]  0.000000000
#> [14,]  0.000000000
#> [15,]  3.033578734
#> [16,] -1.281373564
#> [17,]  0.000000000
#> [18,]  0.389833270
#> [19,] -0.002118019
#> [20,]  0.000000000
#> [21,]  0.000000000
#> [22,] -0.908653546
#> [23,] -0.043573053
#> [24,]  0.001442756
#> [25,] -1.268195021
#> [26,]  0.066300576
#> [27,] -0.003266813
#> [28,] -0.387046651
#> [29,]  0.000000000
#> [30,]  0.000000000
#> [31,] -0.439983442
#> 
#> 
#> $BIC
#> [1] 10615.74
#>