Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr()
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1),
alt_iter = 1, anneal_iter = 3, em_iter = 3)
#> mixedLSR Start: 1 
#> EM Step...
#> Simulated Annealing Step.
#> Full Cycle 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -3739.122
#> 
#> $assign
#>   [1] 1 1 2 1 1 1 1 1 1 1 1 1 1 1 1 2 2 1 1 1 2 1 2 1 2 1 2 1 2 1 2 2 1 1 2 2 2
#>  [38] 2 1 2 2 2 1 1 2 1 1 1 1 1 1 1 2 1 1 1 1 1 2 1 1 1 1 1 1 2 1 1 1 2 1 1 1 2
#>  [75] 1 1 2 1 1 1 1 1 1 1 1 2 1 2 1 2 1 1 2 1 2 1 2 1 2 1
#> 
#> $a
#> $a[[1]]
#>                [,1]          [,2]          [,3]          [,4] [,5] [,6]
#>  [1,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#>  [2,] -2.106067e-20  5.543753e-19 -1.429149e-19 -0.0002750209    0    0
#>  [3,]  6.039865e-17  8.077314e-16 -5.567022e-16  1.4515171855    0    0
#>  [4,] -5.172803e-17 -4.566740e-16  7.570323e-17  1.1445380007    0    0
#>  [5,]  3.118166e-16  1.358043e-15 -9.193596e-16  0.4900793902    0    0
#>  [6,] -1.459331e-17  5.450107e-16 -2.758765e-16  0.7031541171    0    0
#>  [7,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#>  [8,]  1.685694e-16  2.580038e-16 -4.161160e-16  0.9755064661    0    0
#>  [9,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [10,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [11,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [12,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [13,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [14,] -2.275065e-16  3.248028e-16  2.116417e-17  0.9102278031    0    0
#> [15,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [16,] -2.605738e-17  5.302516e-17  2.169102e-17 -0.0858481610    0    0
#> [17,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [18,] -5.015932e-17  2.643870e-16 -2.645216e-16  1.7059072320    0    0
#> [19,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [20,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [21,]  6.346018e-19 -1.153803e-18 -4.053299e-19  0.0007760150    0    0
#> [22,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [23,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [24,]  0.000000e+00  0.000000e+00  0.000000e+00  0.0000000000    0    0
#> [25,] -1.885070e-19 -1.013148e-17  8.280429e-18 -0.0329839159    0    0
#> [26,]  2.503442e-16 -7.546901e-16  2.685732e-16 -2.1101234716    0    0
#> [27,] -2.264611e-16 -1.879260e-15  1.419184e-15 -3.5715213302    0    0
#> [28,] -5.655127e-17  5.338745e-16 -5.347878e-16  3.0614689459    0    0
#> [29,] -1.687598e-17  4.850122e-16 -2.463999e-16  0.6662664428    0    0
#> [30,]  9.995293e-19  2.899097e-19 -1.605710e-18  0.0026980738    0    0
#> [31,] -1.133953e-15  1.428207e-15  7.532311e-16  0.1552240361    0    0
#>               [,7] [,8]          [,9]         [,10]        [,11] [,12] [,13]
#>  [1,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#>  [2,]  0.001001037    0  0.0007266505 -0.0007489634 -0.001012612     0     0
#>  [3,] -0.379783669    0  1.0684091711 -1.0913738902  0.384175326     0     0
#>  [4,] -1.900251516    0 -0.7256976122  0.7366981883  1.893899456     0     0
#>  [5,]  1.738260688    0  2.1397181721 -2.1559158101 -1.682433056     0     0
#>  [6,] -0.042249427    0  0.6712310268 -0.6925779878  0.032381863     0     0
#>  [7,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#>  [8,] -0.450539876    0  0.4872609134 -0.4787994192  0.486525400     0     0
#>  [9,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [10,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [11,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [12,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [13,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [14,] -0.862683963    0  0.1220107292 -0.1579628057  0.806236889     0     0
#> [15,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [16,]  0.124424491    0  0.0448082792 -0.0494881394 -0.131099284     0     0
#> [17,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [18,] -1.533792255    0  0.2054877142 -0.2204937714  1.519177215     0     0
#> [19,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [20,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [21,] -0.001448870    0 -0.0008406235  0.0009496895  0.001609639     0     0
#> [22,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [23,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [24,]  0.000000000    0  0.0000000000  0.0000000000  0.000000000     0     0
#> [25,]  0.021296677    0 -0.0119819777  0.0123312632 -0.021221605     0     0
#> [26,]  1.598460608    0 -0.6051009375  0.6591309636 -1.531674964     0     0
#> [27,]  0.966545785    0 -2.5749665101  2.6192957372 -0.996662459     0     0
#> [28,] -2.625201594    0  0.4863639316 -0.5118990291  2.605999309     0     0
#> [29,] -0.086287629    0  0.5907564479 -0.6101682201  0.076610761     0     0
#> [30,] -0.001042569    0  0.0014004435 -0.0013051362  0.001270564     0     0
#> [31,]  0.105042204    0  0.5806382337 -0.7526543567 -0.384533696     0     0
#>               [,14] [,15]         [,16]         [,17]         [,18] [,19] [,20]
#>  [1,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#>  [2,]  0.0007317093     0  0.0009885824 -0.0009885026  0.0009781695     0     0
#>  [3,]  1.0859339826     0 -0.3833780929  0.3823354549 -0.3866808986     0     0
#>  [4,] -0.7361332987     0 -1.8960423168  1.8976681113 -1.8791438160     0     0
#>  [5,]  2.1887840083     0  1.7524460970 -1.7590528718  1.7278590132     0     0
#>  [6,]  0.6774093874     0 -0.0516942693  0.0519921126 -0.0548014367     0     0
#>  [7,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#>  [8,]  0.5079437916     0 -0.4332560601  0.4301406593 -0.4336804460     0     0
#>  [9,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [10,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [11,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [12,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [13,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [14,]  0.1043465854     0 -0.8926653239  0.8973555442 -0.8861828485     0     0
#> [15,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [16,]  0.0428766140     0  0.1203136887 -0.1198499057  0.1193869255     0     0
#> [17,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [18,]  0.2050641130     0 -1.5404466654  1.5419232923 -1.5308559670     0     0
#> [19,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [20,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [21,] -0.0007925758     0 -0.0013539146  0.0013420997 -0.0013433358     0     0
#> [22,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [23,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [24,]  0.0000000000     0  0.0000000000  0.0000000000  0.0000000000     0     0
#> [25,] -0.0121546011     0  0.0213761211 -0.0213795030  0.0212951946     0     0
#> [26,] -0.5915872939     0  1.6358219278 -1.6411975457  1.6267362378     0     0
#> [27,] -2.6243622084     0  0.9645125073 -0.9604498480  0.9725813421     0     0
#> [28,]  0.4896983175     0 -2.6342670592  2.6362272327 -2.6187138543     0     0
#> [29,]  0.5958948397     0 -0.0950573349  0.0954137226 -0.0974372278     0     0
#> [30,]  0.0015048248     0 -0.0009256435  0.0009066318 -0.0009314679     0     0
#> [31,]  0.4873782920     0 -0.0505580139  0.0724855243 -0.0478933159     0     0
#>       [,21]         [,22]         [,23] [,24]        [,25]         [,26]
#>  [1,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#>  [2,]     0  0.0007318326  0.0007229077     0 -0.001001790  0.0007244028
#>  [3,]     0  1.0627701524  1.0581872400     0  0.386618322  1.0671773158
#>  [4,]     0 -0.7537646160 -0.7377351450     0  1.885257907 -0.7294015700
#>  [5,]     0  2.1903965740  2.1617508412     0 -1.682729834  2.1531022448
#>  [6,]     0  0.6630819222  0.6603805440     0  0.038532061  0.6676936323
#>  [7,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#>  [8,]     0  0.5005702609  0.4965396461     0  0.476151004  0.4945686249
#>  [9,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [10,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [11,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [12,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [13,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [14,]     0  0.0769388143  0.0871427202     0  0.820997461  0.1064729860
#> [15,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [16,]     0  0.0425369427  0.0422452109     0 -0.128364500  0.0433134199
#> [17,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [18,]     0  0.1751628234  0.1844025092     0  1.518977231  0.1989404600
#> [19,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [20,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [21,]     0 -0.0007650921 -0.0007684922     0  0.001550683 -0.0008023250
#> [22,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [23,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [24,]     0  0.0000000000  0.0000000000     0  0.000000000  0.0000000000
#> [25,]     0 -0.0116101977 -0.0116709919     0 -0.021223959 -0.0119083144
#> [26,]     0 -0.5424992061 -0.5552038703     0 -1.549237141 -0.5850087181
#> [27,]     0 -2.5720047045 -2.5592143532     0 -0.996348909 -2.5767069308
#> [28,]     0  0.4374670125  0.4518750191     0  2.604524336  0.4763765888
#> [29,]     0  0.5823270258  0.5803391893     0  0.082146546  0.5873149405
#> [30,]     0  0.0015145150  0.0014884300     0  0.001203349  0.0014524768
#> [31,]     0  0.4182501211  0.4434346777     0 -0.295301413  0.5112709693
#>               [,27] [,28] [,29] [,30] [,31]         [,32] [,33]        [,34]
#>  [1,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#>  [2,]  0.0007192134     0     0     0     0 -0.0017201719     0 -0.001010345
#>  [3,]  1.0728764944     0     0     0     0 -0.6914500573     0  0.368308738
#>  [4,] -0.7418701989     0     0     0     0  2.6321074393     0  1.895235572
#>  [5,]  2.2112211690     0     0     0     0 -3.9196276661     0 -1.722977835
#>  [6,]  0.6610365670     0     0     0     0 -0.6218214926     0  0.029179575
#>  [7,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#>  [8,]  0.5253870486     0     0     0     0 -0.0622662402     0  0.462357279
#>  [9,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [10,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [11,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [12,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [13,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [14,]  0.0564262937     0     0     0     0  0.7820789104     0  0.826267131
#> [15,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [16,]  0.0380806305     0     0     0     0 -0.1645260841     0 -0.127917953
#> [17,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [18,]  0.1829879008     0     0     0     0  1.3401287251     0  1.515755099
#> [19,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [20,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [21,] -0.0006715179     0     0     0     0  0.0021748331     0  0.001538225
#> [22,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [23,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [24,]  0.0000000000     0     0     0     0  0.0000000000     0  0.000000000
#> [25,] -0.0118210897     0     0     0     0 -0.0093869193     0 -0.021013770
#> [26,] -0.5255412172     0     0     0     0 -1.0428433153     0 -1.547290335
#> [27,] -2.6069492340     0     0     0     0  1.6292613046     0 -0.949240749
#> [28,]  0.4543988174     0     0     0     0  2.1548895651     0  2.595983928
#> [29,]  0.5804801727     0     0     0     0 -0.4974107881     0  0.073995821
#> [30,]  0.0016478617     0     0     0     0 -0.0005212773     0  0.001145491
#> [31,]  0.2750183618     0     0     0     0 -0.4843393280     0 -0.262828662
#>               [,35]
#>  [1,]  0.0000000000
#>  [2,] -0.0017316128
#>  [3,] -0.6927769921
#>  [4,]  2.6448744696
#>  [5,] -3.9265905060
#>  [6,] -0.6265632549
#>  [7,]  0.0000000000
#>  [8,] -0.0538196595
#>  [9,]  0.0000000000
#> [10,]  0.0000000000
#> [11,]  0.0000000000
#> [12,]  0.0000000000
#> [13,]  0.0000000000
#> [14,]  0.7742565171
#> [15,]  0.0000000000
#> [16,] -0.1669014172
#> [17,]  0.0000000000
#> [18,]  1.3450221915
#> [19,]  0.0000000000
#> [20,]  0.0000000000
#> [21,]  0.0022219998
#> [22,]  0.0000000000
#> [23,]  0.0000000000
#> [24,]  0.0000000000
#> [25,] -0.0094443356
#> [26,] -1.0350192296
#> [27,]  1.6278882940
#> [28,]  2.1639685372
#> [29,] -0.5015143544
#> [30,] -0.0004712301
#> [31,] -0.5485281501
#> 
#> $a[[2]]
#>                [,1]          [,2]          [,3]        [,4] [,5] [,6]
#>  [1,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#>  [2,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#>  [3,]  7.839286e-17  1.118381e-16  1.403605e-17  0.67832546    0    0
#>  [4,] -1.962260e-17 -2.101709e-17 -5.852278e-19  0.20319884    0    0
#>  [5,] -8.225855e-17 -1.822720e-16 -4.197286e-17  0.88207736    0    0
#>  [6,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#>  [7,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#>  [8,] -2.401718e-18 -5.200396e-18 -1.174527e-18  0.01508142    0    0
#>  [9,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [10,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [11,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [12,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [13,] -2.027418e-18  1.315617e-17  6.372130e-18  0.04179576    0    0
#> [14,] -4.400991e-17  8.301319e-19  1.881812e-17 -0.91532023    0    0
#> [15,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [16,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [17,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [18,] -4.607910e-17 -2.754530e-17  7.778120e-18 -0.18926575    0    0
#> [19,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [20,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [21,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [22,]  1.458167e-17  5.462441e-17  1.680482e-17 -0.51193899    0    0
#> [23,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [24,]  1.960092e-17  1.894265e-17 -2.762572e-19 -0.89303728    0    0
#> [25,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [26,]  0.000000e+00  0.000000e+00  0.000000e+00  0.00000000    0    0
#> [27,]  1.376082e-16  1.278510e-16 -4.094795e-18  0.35155202    0    0
#> [28,]  1.942319e-17  1.374071e-17 -2.384779e-18  0.09969754    0    0
#> [29,]  3.522232e-17  4.055716e-17  2.238887e-18  0.25111932    0    0
#> [30,] -8.829630e-18 -3.285408e-18  2.326755e-18 -0.10358576    0    0
#> [31,] -6.750220e-17 -2.834620e-16 -9.063233e-17  0.15675809    0    0
#>               [,7]          [,8]        [,9]       [,10]        [,11] [,12]
#>  [1,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#>  [2,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#>  [3,] -0.123047025 -6.678206e-03  0.52842182 -0.51176722  0.148363336     0
#>  [4,]  0.073209391 -3.677002e-03  0.27920305 -0.28240059 -0.077971435     0
#>  [5,]  0.159538244 -1.037067e-02  1.09443503 -1.11951731 -0.215665037     0
#>  [6,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#>  [7,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#>  [8,]  0.003955667 -1.683133e-04  0.02055126 -0.02127154 -0.005516473     0
#>  [9,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [10,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [11,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [12,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [13,]  0.033708542 -1.541922e-03  0.06904217 -0.06746443 -0.028389858     0
#> [14,]  0.125486227  7.157585e-03 -0.80131609  0.79971021 -0.113754339     0
#> [15,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [16,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [17,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [18,]  0.150700628 -1.059799e-03 -0.03997104  0.03498058 -0.151566634     0
#> [19,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [20,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [21,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [22,] -0.022602818  5.262547e-03 -0.55282557  0.55994059  0.042147221     0
#> [23,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [24,] -0.134306456  1.294058e-02 -1.02761622  1.03036071  0.140308868     0
#> [25,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [26,]  0.000000000  0.000000e+00  0.00000000  0.00000000  0.000000000     0
#> [27,] -0.395402885  3.361680e-03 -0.05908464  0.07958557  0.416083782     0
#> [28,] -0.058461125  5.949431e-05  0.04085093 -0.03847741  0.059583461     0
#> [29,] -0.075216999 -1.728774e-03  0.16819958 -0.16193460  0.083036645     0
#> [30,]  0.026399592  5.742330e-04 -0.07816864  0.07744067 -0.025603931     0
#> [31,] -0.129797584  6.749477e-03  0.13018001 -0.16753907  0.034117186     0
#>       [,13]       [,14] [,15]        [,16]        [,17]        [,18] [,19]
#>  [1,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#>  [2,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#>  [3,]     0  0.49315191     0 -0.130785991  0.128377065 -0.141428475     0
#>  [4,]     0  0.28350533     0  0.086628539 -0.078064928  0.073796450     0
#>  [5,]     0  1.16347379     0  0.219926130 -0.160408467  0.155653859     0
#>  [6,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#>  [7,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#>  [8,]     0  0.02252240     0  0.005379423 -0.003947321  0.003977514     0
#>  [9,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [10,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [11,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [12,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [13,]     0  0.06113363     0  0.035230293 -0.037524135  0.034494586     0
#> [14,]     0 -0.81494802     0  0.110888035 -0.138795023  0.146235543     0
#> [15,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [16,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [17,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [18,]     0 -0.04036711     0  0.160538459 -0.162963848  0.161429063     0
#> [19,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [20,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [21,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [22,]     0 -0.57645575     0 -0.045232745  0.020069276 -0.016247097     0
#> [23,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [24,]     0 -1.02942746     0 -0.169320446  0.143405640 -0.126776710     0
#> [25,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [26,]     0  0.00000000     0  0.000000000  0.000000000  0.000000000     0
#> [27,]     0 -0.08219886     0 -0.433035825  0.423693719 -0.421991370     0
#> [28,]     0  0.03982328     0 -0.062054200  0.063075527 -0.063086054     0
#> [29,]     0  0.15771524     0 -0.080052123  0.079931600 -0.083450553     0
#> [30,]     0 -0.07919112     0  0.026102443 -0.028792584  0.029386476     0
#> [31,]     0  0.26056551     0 -0.093531573  0.160164565 -0.134715961     0
#>       [,20] [,21]        [,22]       [,23] [,24]        [,25]       [,26]
#>  [1,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#>  [2,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#>  [3,]     0     0  0.539099764  0.55409876     0  0.150775755  0.55155842
#>  [4,]     0     0  0.279959843  0.28574642     0 -0.080881931  0.28051967
#>  [5,]     0     0  1.038508981  1.07941595     0 -0.242771208  1.06020791
#>  [6,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#>  [7,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#>  [8,]     0     0  0.019021489  0.01988905     0 -0.006192010  0.01948521
#>  [9,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [10,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [11,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [12,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [13,]     0     0  0.078674750  0.07730132     0 -0.026498308  0.07593316
#> [14,]     0     0 -0.766085424 -0.80051470     0 -0.100366157 -0.79315700
#> [15,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [16,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [17,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [18,]     0     0 -0.022839013 -0.03204140     0 -0.148954893 -0.03499024
#> [19,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [20,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [21,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [22,]     0     0 -0.528026565 -0.54922748     0  0.053651301 -0.54120541
#> [23,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [24,]     0     0 -1.027912182 -1.05316339     0  0.149170425 -1.03746035
#> [25,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [26,]     0     0  0.000000000  0.00000000     0  0.000000000  0.00000000
#> [27,]     0     0 -0.081130844 -0.06742461     0  0.417309684 -0.05692371
#> [28,]     0     0  0.034914019  0.03900803     0  0.058632397  0.03987899
#> [29,]     0     0  0.167058078  0.17388073     0  0.083102487  0.17387420
#> [30,]     0     0 -0.073466490 -0.07749681     0 -0.024181909 -0.07711287
#> [31,]     0     0  0.001803507  0.04152547     0 -0.003772613  0.03881251
#>             [,27] [,28] [,29] [,30] [,31]       [,32] [,33]        [,34]
#>  [1,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#>  [2,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#>  [3,]  0.53898528     0     0     0     0 -0.40325546     0  0.155894589
#>  [4,]  0.28153159     0     0     0     0 -0.36273567     0 -0.070656378
#>  [5,]  1.08328928     0     0     0     0 -1.29606084     0 -0.182885887
#>  [6,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#>  [7,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#>  [8,]  0.02015578     0     0     0     0 -0.02546228     0 -0.004827304
#>  [9,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [10,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [11,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [12,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [13,]  0.07292053     0     0     0     0 -0.10467602     0 -0.027336664
#> [14,] -0.79781266     0     0     0     0  0.68713298     0 -0.132229421
#> [15,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [16,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [17,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [18,] -0.03567763     0     0     0     0 -0.11819160     0 -0.151321995
#> [19,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [20,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [21,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [22,] -0.54907456     0     0     0     0  0.59161854     0  0.027147722
#> [23,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [24,] -1.03669973     0     0     0     0  1.19012467     0  0.115824022
#> [25,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [26,]  0.00000000     0     0     0     0  0.00000000     0  0.000000000
#> [27,] -0.06359985     0     0     0     0  0.48134366     0  0.409367077
#> [28,]  0.03970203     0     0     0     0  0.02019585     0  0.059946280
#> [29,]  0.17027595     0     0     0     0 -0.09043984     0  0.085222447
#> [30,] -0.07751188     0     0     0     0  0.05203857     0 -0.027264496
#> [31,]  0.08574136     0     0     0     0 -0.01745178     0  0.048808313
#>             [,35]
#>  [1,]  0.00000000
#>  [2,]  0.00000000
#>  [3,] -0.37479424
#>  [4,] -0.35006758
#>  [5,] -1.25311468
#>  [6,]  0.00000000
#>  [7,]  0.00000000
#>  [8,] -0.02473483
#>  [9,]  0.00000000
#> [10,]  0.00000000
#> [11,]  0.00000000
#> [12,]  0.00000000
#> [13,] -0.10040820
#> [14,]  0.65193390
#> [15,]  0.00000000
#> [16,]  0.00000000
#> [17,]  0.00000000
#> [18,] -0.11955029
#> [19,]  0.00000000
#> [20,]  0.00000000
#> [21,]  0.00000000
#> [22,]  0.56873456
#> [23,]  0.00000000
#> [24,]  1.14220168
#> [25,]  0.00000000
#> [26,]  0.00000000
#> [27,]  0.48022170
#> [28,]  0.02204373
#> [29,] -0.08157824
#> [30,]  0.04860990
#> [31,] -0.02761088
#> 
#> 
#> $BIC
#> [1] 10397.92
#>