Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Mixed Low-Rank and Sparse Multivariate Regression for High-Dimensional Data

Usage

mixed_lsr(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  init_lambda = NULL,
  alt_iter = 5,
  anneal_iter = 1000,
  em_iter = 1000,
  temp = 1000,
  mu = 0.95,
  eps = 1e-06,
  accept_prob = 0.95,
  sim_N = 200,
  verbose = TRUE
)

Arguments

x

A matrix of predictors.

y

A matrix of responses.

k

The number of groups.

nstart

The number of random initializations, the result with the maximum likelihood is returned.

init_assign

A vector of initial assignments, NULL by default.

init_lambda

A vector with the values to initialize the penalization parameter for each group, e.g., c(1,1,1). Set to NULL by default.

alt_iter

The maximum number of times to alternate between the classification expectation maximization algorithm and the simulated annealing algorithm.

anneal_iter

The maximum number of simulated annealing iterations.

em_iter

The maximum number of EM iterations.

temp

The initial simulated annealing temperature, temp > 0.

mu

The simulated annealing decrease temperature fraction. Once the best configuration cannot be improved, reduce the temperature to (mu)T, 0 < mu < 1.

eps

The final simulated annealing temperature, eps > 0.

accept_prob

The simulated annealing probability of accepting a new assignment 0 < accept_prob < 1. When closer to 1, trial assignments will only be small perturbation of the current assignment. When closer to 0, trial assignments are closer to random.

sim_N

The simulated annealing number of iterations for reaching equilibrium.

verbose

A boolean indicating whether to print to screen.

Value

A list containing the likelihood, the partition, the coefficient matrices, and the BIC.

Examples

simulate <- simulate_lsr(50)
mixed_lsr(simulate$x, simulate$y, k = 2, init_lambda = c(1,1), alt_iter = 0)
#> mixedLSR Start: 1 
#> Computing Final Model...
#> Done! 
#> $llik
#> [1] -2790.828
#> 
#> $assign
#>  [1] 2 2 2 1 1 2 2 1 2 2 2 2 2 1 1 1 2 1 2 2 1 2 1 1 2 2 1 1 2 1 2 2 2 1 2 2 2 2
#> [39] 1 2 1 2 1 1 2 2 2 1 2 2
#> 
#> $a
#> $a[[1]]
#>               [,1]          [,2]        [,3]          [,4]          [,5]
#>  [1,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#>  [2,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#>  [3,]  0.443270459  3.844313e-17  0.61126368  1.501799e-32  6.983362e-33
#>  [4,]  0.314266921  5.953817e-17 -0.36683246  1.189578e-31  2.575886e-34
#>  [5,] -0.374333947 -7.556238e-17  1.12153804 -7.070155e-32 -9.182835e-33
#>  [6,]  0.706629750  1.235739e-16 -0.77761703  2.012985e-31  5.565714e-33
#>  [7,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#>  [8,]  0.034146976  4.666649e-18 -0.03343722  1.061629e-33  9.317191e-34
#>  [9,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [10,] -0.390701183 -6.546697e-17  0.18460013 -1.282411e-31 -5.659428e-34
#> [11,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [12,]  0.003084198  3.813772e-19  0.03600593  5.743148e-33 -5.478850e-34
#> [13,]  0.217478874  3.793655e-17  0.34173181  1.496091e-31 -7.978969e-33
#> [14,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [15,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [16,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [17,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [18,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [19,]  0.346043230  8.392844e-17 -1.97530403  2.294957e-32  1.633126e-32
#> [20,] -1.720547320 -3.307966e-16  2.00924245 -6.855889e-31  1.288767e-33
#> [21,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [22,] -0.388445843 -5.868859e-17  0.02946354 -1.053311e-31 -1.569982e-33
#> [23,] -0.075358589 -1.116772e-17  0.10087390 -4.415495e-33 -2.022842e-33
#> [24,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [25,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [26,] -0.080237504 -4.670922e-18 -1.38122479 -1.795855e-31  1.876259e-32
#> [27,]  2.374091972  3.678778e-16 -1.96900825  4.372301e-31  3.444497e-32
#> [28,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [29,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [30,]  0.000000000  0.000000e+00  0.00000000  0.000000e+00  0.000000e+00
#> [31,] -0.184045404 -8.735124e-17 -0.04658415 -4.832682e-31  3.368311e-32
#>                [,6]         [,7]         [,8]          [,9]        [,10]
#>  [1,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#>  [2,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#>  [3,]  1.382013e-32  0.969604776 -1.099031923 -6.131995e-33 -0.438220943
#>  [4,]  3.974489e-33 -0.045802270  0.081229504  2.288882e-33 -0.323773454
#>  [5,] -1.966390e-32  0.681116178 -0.735959998  6.980997e-33  0.390352105
#>  [6,]  1.655472e-32 -0.059463570  0.096485621 -8.655546e-34 -0.724858085
#>  [7,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#>  [8,]  1.816313e-33  0.000956381 -0.004588313 -8.381408e-34 -0.034638139
#>  [9,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [10,] -4.836967e-33 -0.191712608  0.180770465 -2.201225e-33  0.398759547
#> [11,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [12,] -8.817361e-34  0.035797052 -0.035905641  6.281108e-34 -0.002924507
#> [13,] -1.091050e-32  0.513379105 -0.496617547  1.054862e-32 -0.220192798
#> [14,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [15,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [16,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [17,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [18,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [19,]  3.196355e-32 -1.488855565  1.589161051 -1.459880e-32 -0.369468394
#> [20,] -1.759224e-32  0.251744540 -0.465940246 -1.575341e-32  1.773966968
#> [21,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [22,] -6.090396e-33 -0.331787256  0.346170073 -7.870856e-34  0.393885751
#> [23,] -4.005132e-33  0.022697203 -0.018050329  1.774843e-33  0.076822142
#> [24,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [25,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [26,]  3.069560e-32 -1.338181798  1.358573600 -2.114691e-32  0.072402300
#> [27,]  7.879348e-32  0.390613529 -0.457822642 -2.253173e-32 -2.418772309
#> [28,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [29,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [30,]  0.000000e+00  0.000000000  0.000000000  0.000000e+00  0.000000000
#> [31,]  5.039820e-32 -0.210700745 -0.032328486 -4.138066e-32  0.203173557
#>               [,11]       [,12]        [,13]         [,14]        [,15]
#>  [1,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#>  [2,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#>  [3,] -6.559542e-34 -1.08978658 -1.104752449  2.287407e-33 -0.439967019
#>  [4,]  2.244779e-34  0.08056541  0.074271455 -3.320954e-33 -0.309903558
#>  [5,]  7.555397e-34 -0.74829221 -0.733139333 -1.542410e-33  0.389178605
#>  [6,] -1.251612e-34  0.10246888  0.083916177 -3.622104e-33 -0.704891751
#>  [7,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#>  [8,] -8.949587e-35 -0.00330599 -0.004786937  3.322797e-34 -0.035151731
#>  [9,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [10,] -2.136181e-34  0.18116295  0.189266626  3.467870e-33  0.383688367
#> [11,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [12,]  6.599981e-35 -0.03672663 -0.036312989 -3.785215e-34 -0.001819551
#> [13,]  1.099722e-33 -0.50914569 -0.506457537 -7.409817e-33 -0.196288051
#> [14,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [15,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [16,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [17,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [18,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [19,] -1.559575e-33  1.61207607  1.591380951  5.699375e-33 -0.379936369
#> [20,] -1.566760e-33 -0.45826370 -0.426181807  2.021957e-32  1.692107340
#> [21,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [22,] -6.665692e-35  0.34501524  0.353994459  2.417998e-33  0.382037712
#> [23,]  1.898705e-34 -0.02082382 -0.017591151 -6.607060e-34  0.077719483
#> [24,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [25,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [26,] -2.224300e-33  1.38657552  1.371947467  1.247110e-32  0.036441704
#> [27,] -2.472748e-33 -0.41399358 -0.490335794  8.380526e-34 -2.390095098
#> [28,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [29,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [30,]  0.000000e+00  0.00000000  0.000000000  0.000000e+00  0.000000000
#> [31,] -4.330989e-33  0.01864891 -0.007390150  2.701516e-32  0.122421180
#>              [,16]         [,17]        [,18] [,19] [,20] [,21]       [,22]
#>  [1,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#>  [2,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#>  [3,]  0.380352893 -3.637197e-34 -0.455192873     0     0     0  0.41817878
#>  [4,]  0.350812520 -4.662427e-35 -0.315197824     0     0     0  0.32700646
#>  [5,] -0.389914600  4.925671e-34  0.354538754     0     0     0 -0.39429157
#>  [6,]  0.748439733 -3.429834e-34 -0.703880616     0     0     0  0.72567922
#>  [7,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#>  [8,]  0.030924609 -4.826321e-35 -0.033336428     0     0     0  0.03385289
#>  [9,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [10,] -0.418776625  6.510215e-35  0.395921307     0     0     0 -0.39841124
#> [11,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [12,]  0.005338631  2.659470e-35 -0.004344126     0     0     0  0.00299358
#> [13,]  0.270372704  3.688020e-34 -0.236797797     0     0     0  0.22285385
#> [14,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [15,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [16,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [17,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [18,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [19,]  0.354806006 -8.471797e-34 -0.303941488     0     0     0  0.37820929
#> [20,] -1.941066596  1.259369e-34  1.728658600     0     0     0 -1.79478638
#> [21,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [22,] -0.399951299  1.103652e-34  0.394967762     0     0     0 -0.39006185
#> [23,] -0.070045797  1.053755e-34  0.073174396     0     0     0 -0.07553510
#> [24,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [25,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [26,] -0.147562050 -9.155418e-34  0.125180669     0     0     0 -0.07269859
#> [27,]  2.358869325 -1.895875e-33 -2.357469381     0     0     0  2.39372807
#> [28,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [29,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [30,]  0.000000000  0.000000e+00  0.000000000     0     0     0  0.00000000
#> [31,] -0.446091325 -1.598487e-33  0.226743245     0     0     0 -0.24006515
#>              [,23]       [,24] [,25] [,26] [,27]        [,28] [,29]
#>  [1,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#>  [2,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#>  [3,] -0.994862798 -1.37785108     0     0     0 -1.085277517     0
#>  [4,]  0.028795473 -0.30159634     0     0     0  0.065596142     0
#>  [5,] -0.678973271 -0.32305237     0     0     0 -0.731775153     0
#>  [6,]  0.027507017 -0.67976539     0     0     0  0.076039760     0
#>  [7,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#>  [8,] -0.001646285 -0.03116237     0     0     0 -0.003654371     0
#>  [9,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [10,]  0.214972676  0.61494142     0     0     0  0.195370921     0
#> [11,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [12,] -0.037061593 -0.04273235     0     0     0 -0.036997772     0
#> [13,] -0.541493036 -0.80582963     0     0     0 -0.521216307     0
#> [14,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [15,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [16,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [17,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [18,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [19,]  1.505888104  1.19891748     0     0     0  1.592951810     0
#> [20,] -0.155012726  1.67226541     0     0     0 -0.372767476     0
#> [21,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [22,]  0.354947002  0.74056609     0     0     0  0.355508003     0
#> [23,] -0.021373963  0.04490051     0     0     0 -0.019604241     0
#> [24,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [25,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [26,]  1.381153985  1.54454958     0     0     0  1.392842349     0
#> [27,] -0.482597848 -2.72981005     0     0     0 -0.472526715     0
#> [28,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [29,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [30,]  0.000000000  0.00000000     0     0     0  0.000000000     0
#> [31,]  0.268537954  0.68416584     0     0     0  0.067721890     0
#>              [,30]       [,31]        [,32] [,33] [,34] [,35]
#>  [1,]  0.000000000  0.00000000  0.000000000     0     0     0
#>  [2,]  0.000000000  0.00000000  0.000000000     0     0     0
#>  [3,] -0.477412793  0.61298272  1.103320399     0     0     0
#>  [4,] -0.299709718 -0.35258007 -0.062791917     0     0     0
#>  [5,]  0.361584104  1.12264025  0.728555364     0     0     0
#>  [6,] -0.689976900 -0.75701784 -0.066067572     0     0     0
#>  [7,]  0.000000000  0.00000000  0.000000000     0     0     0
#>  [8,] -0.035412605 -0.03395689  0.004605176     0     0     0
#>  [9,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [10,]  0.381809966  0.16824987 -0.201365409     0     0     0
#> [11,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [12,] -0.002567726  0.03727539  0.037008143     0     0     0
#> [13,] -0.202137519  0.36830976  0.522977259     0     0     0
#> [14,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [15,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [16,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [17,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [18,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [19,] -0.327880501 -1.99129909 -1.592967397     0     0     0
#> [20,]  1.632917058  1.92509116  0.359186538     0     0     0
#> [21,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [22,]  0.387372913  0.01595213 -0.363458642     0     0     0
#> [23,]  0.077207287  0.10186415  0.017761939     0     0     0
#> [24,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [25,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [26,]  0.067494025 -1.42314556 -1.394009947     0     0     0
#> [27,] -2.384375764 -1.93742509  0.521886260     0     0     0
#> [28,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [29,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [30,]  0.000000000  0.00000000  0.000000000     0     0     0
#> [31,]  0.084609935 -0.13086311 -0.049515053     0     0     0
#> 
#> $a[[2]]
#>                [,1]          [,2]         [,3] [,4] [,5] [,6]         [,7]
#>  [1,] -0.1707981759 -1.239488e-16  0.114005378    0    0    0 -0.062105222
#>  [2,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#>  [3,] -0.1548075055 -1.104317e-16  1.130530248    0    0    0  0.829021846
#>  [4,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#>  [5,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#>  [6,]  0.3563319540  2.398351e-16 -0.528135303    0    0    0 -0.265476500
#>  [7,] -0.0959634112 -7.032173e-17  0.027342461    0    0    0 -0.071404775
#>  [8,]  0.0001614907 -2.301532e-19 -0.006378015    0    0    0 -0.008014719
#>  [9,]  0.2697932991  1.870347e-16 -0.353271723    0    0    0 -0.118222996
#> [10,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [11,] -0.0706517986 -5.803505e-17 -0.133962517    0    0    0 -0.232890713
#> [12,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [13,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [14,]  0.0435661793  1.665840e-17 -0.651552244    0    0    0 -0.630378648
#> [15,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [16,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [17,]  0.1447101350  1.020928e-16 -0.380233928    0    0    0 -0.210894300
#> [18,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [19,] -0.2883102375 -2.328794e-16 -0.073141206    0    0    0 -0.517883459
#> [20,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [21,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [22,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [23,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [24,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [25,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [26,] -0.0238386233 -6.964408e-17 -2.008225082    0    0    0 -2.139491790
#> [27,]  0.4328411267  3.148971e-16 -0.262909972    0    0    0  0.185642150
#> [28,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [29,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [30,]  0.0000000000  0.000000e+00  0.000000000    0    0    0  0.000000000
#> [31,] -0.3837259499 -4.042867e-16  1.455165111    0    0    0 -0.124375465
#>               [,8] [,9]         [,10] [,11]        [,12]        [,13] [,14]
#>  [1,]  0.056375113    0  1.667360e-01     0  0.054502125  0.060341362     0
#>  [2,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#>  [3,] -0.910261301    0  2.274365e-01     0 -0.906620025 -1.063710052     0
#>  [4,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#>  [5,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#>  [6,]  0.224104471    0 -3.357743e-01     0  0.222093745  0.162650743     0
#>  [7,]  0.068387792    0  9.208502e-02     0  0.066996390  0.072948964     0
#>  [8,]  0.007105625    0 -3.402028e-06     0  0.006993549  0.006262696     0
#>  [9,]  0.105459935    0 -2.606657e-01     0  0.105477024  0.080508487     0
#> [10,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [11,]  0.217537488    0  6.738538e-02     0  0.214265876  0.212454794     0
#> [12,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [13,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [14,]  0.621247861    0 -6.328435e-02     0  0.615150007  0.641499783     0
#> [15,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [16,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [17,]  0.226908106    0 -1.579150e-01     0  0.226440846  0.254081627     0
#> [18,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [19,]  0.432758983    0  3.044996e-01     0  0.422611247  0.356032058     0
#> [20,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [21,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [22,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [23,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [24,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [25,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [26,]  2.089029384    0 -3.737215e-02     0  2.066488675  2.138882598     0
#> [27,] -0.170029511    0 -4.219714e-01     0 -0.164969990 -0.180123302     0
#> [28,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [29,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [30,]  0.000000000    0  0.000000e+00     0  0.000000000  0.000000000     0
#> [31,] -0.503198845    0  7.002108e-01     0 -0.532171709 -1.363407056     0
#>               [,15]         [,16] [,17]        [,18] [,19] [,20] [,21]
#>  [1,]  1.788847e-01 -0.1811884358     0  0.180254502     0     0     0
#>  [2,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#>  [3,]  2.042111e-01 -0.1595827649     0  0.278400321     0     0     0
#>  [4,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#>  [5,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#>  [6,] -3.570049e-01  0.3950294976     0 -0.304708054     0     0     0
#>  [7,]  9.994747e-02 -0.1013507790     0  0.100657331     0     0     0
#>  [8,]  9.163249e-05  0.0004808034     0  0.001041312     0     0     0
#>  [9,] -2.766576e-01  0.2939099319     0 -0.256043654     0     0     0
#> [10,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [11,]  7.648614e-02 -0.0693015630     0  0.090917938     0     0     0
#> [12,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [13,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [14,] -4.984819e-02  0.0572905099     0 -0.037006263     0     0     0
#> [15,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [16,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [17,] -1.594855e-01  0.1546425310     0 -0.170824205     0     0     0
#> [18,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [19,]  3.267899e-01 -0.2837436873     0  0.406565894     0     0     0
#> [20,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [21,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [22,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [23,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [24,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [25,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [26,]  1.696037e-02  0.0144504590     0  0.075459050     0     0     0
#> [27,] -4.534016e-01  0.4585507867     0 -0.458080946     0     0     0
#> [28,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [29,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [30,]  0.000000e+00  0.0000000000     0  0.000000000     0     0     0
#> [31,]  6.484531e-01 -0.2727809423     0  1.269310045     0     0     0
#>               [,22]        [,23]        [,24] [,25] [,26] [,27]        [,28]
#>  [1,] -1.835908e-01  0.050179586  0.229085110     0     0     0  0.060075958
#>  [2,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#>  [3,] -1.772790e-01 -0.970532886 -0.824033418     0     0     0 -1.054793149
#>  [4,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#>  [5,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#>  [6,]  3.677899e-01  0.225569807 -0.146839772     0     0     0  0.167817845
#>  [7,] -1.033777e-01  0.065955624  0.166913096     0     0     0  0.072714670
#>  [8,] -9.733612e-05  0.006949296  0.006809591     0     0     0  0.006339752
#>  [9,]  2.833604e-01  0.112005647 -0.169615361     0     0     0  0.082717520
#> [10,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [11,] -8.048275e-02  0.215120810  0.290621225     0     0     0  0.213270572
#> [12,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [13,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [14,]  3.955415e-02  0.635567357  0.596702838     0     0     0  0.641473271
#> [15,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [16,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [17,]  1.561448e-01  0.243710982  0.095983163     0     0     0  0.252724922
#> [18,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [19,] -3.303628e-01  0.402226993  0.703709217     0     0     0  0.362371260
#> [20,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [21,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [22,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [23,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [24,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [25,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [26,] -5.300445e-02  2.124116535  2.170370175     0     0     0  2.140004030
#> [27,]  4.657232e-01 -0.154627561 -0.608266247     0     0     0 -0.179509215
#> [28,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [29,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [30,]  0.000000e+00  0.000000000  0.000000000     0     0     0  0.000000000
#> [31,] -5.538376e-01 -0.770056191 -0.411613197     0     0     0 -1.302691821
#>       [,29]         [,30]        [,31]        [,32]         [,33] [,34] [,35]
#>  [1,]     0  0.1767414712  0.121304540 -0.063143995 -0.0005338233     0     0
#>  [2,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#>  [3,]     0  0.1827392156  1.173494470  0.816891439 -0.0191019341     0     0
#>  [4,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#>  [5,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#>  [6,]     0 -0.3695650035 -0.565929999 -0.240651946 -0.0316172329     0     0
#>  [7,]     0  0.0986830934  0.029731091 -0.070933089 -0.0009118324     0     0
#>  [8,]     0 -0.0002026845 -0.006833353 -0.007586183 -0.0005859598     0     0
#>  [9,]     0 -0.2804189303 -0.376621547 -0.105670491 -0.0136059116     0     0
#> [10,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [11,]     0  0.0710884026 -0.141689208 -0.224470535 -0.0103397817     0     0
#> [12,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [13,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [14,]     0 -0.0545493489 -0.684754157 -0.608565236 -0.0173613889     0     0
#> [15,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [16,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [17,]     0 -0.1552427073 -0.398072082 -0.204360618  0.0005966875     0     0
#> [18,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [19,]     0  0.2995533496 -0.085847783 -0.491959934 -0.0443863831     0     0
#> [20,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [21,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [22,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [23,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [24,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [25,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [26,]     0 -0.0049905885 -2.111617897 -2.065600616 -0.0647005859     0     0
#> [27,]     0 -0.4475552770 -0.279991779  0.187205352  0.0024132578     0     0
#> [28,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [29,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [30,]     0  0.0000000000  0.000000000  0.000000000  0.0000000000     0     0
#> [31,]     0  0.4623212057  1.427567947 -0.002725755 -0.2900024304     0     0
#> 
#> 
#> $BIC
#> [1] 8014.935
#>