Simulate Heterogeneous, Low-Rank, and Sparse Data

Simulate Heterogeneous, Low-Rank, and Sparse Data

Usage

simulate_lsr(
  N = 100,
  k = 2,
  p = 30,
  m = 35,
  b = 1,
  d = 20,
  h = 0.2,
  case = "independent"
)

Arguments

N

The sample size, default = 100.

k

The number of groups, default = 2.

p

The number of predictor features, default = 30.

m

The number of response features, default = 35.

b

The signal-to-noise ratio, default = 1.

d

The singular value, default = 20.

h

The lower bound for the singular matrix simulation, default = 0.2.

case

The covariance case, "independent" or "dependent", default = "independent".

Value

A list of simulation values, including x matrix, y matrix, coefficients and true clustring assignments.

Examples

simulate_lsr()
#> $x
#>               [,1]         [,2]        [,3]        [,4]         [,5]
#>   [1,]  0.54738795  0.276420018  0.25566692  0.87466494 -1.440630794
#>   [2,] -0.61449491 -1.037935501  1.55391276  0.34446015  1.424380655
#>   [3,] -0.07661918  0.454425332  0.36754039  1.03067604 -0.751737987
#>   [4,]  0.58731164 -0.845432495  0.21559888  0.84562944  0.736475452
#>   [5,]  0.48368543  0.356204558 -1.25620710  0.32954607 -1.729515155
#>   [6,]  0.88221608 -0.688900349 -0.55039719 -0.94687710 -0.961383254
#>   [7,] -0.40808101 -0.101754651 -2.54109577 -2.12444793  1.139207251
#>   [8,] -0.88327311  0.700899555  0.22202469  1.41338705 -1.151648841
#>   [9,] -0.18087093 -0.782587442  0.53675078 -0.02067991 -0.619860375
#>  [10,] -0.72094091  1.226776054  1.83461519  1.89296492 -0.476549506
#>  [11,]  1.27541661  1.765614345  0.96853903  1.49849644 -0.131786865
#>  [12,]  0.51976270 -1.610791179 -1.57078174 -0.03797831  0.125877430
#>  [13,] -1.29340334  0.024583072 -0.68311267 -0.73843839 -1.013435190
#>  [14,] -0.56844096 -0.190851820  1.48249567 -0.95263917  1.679446085
#>  [15,]  0.75504199  1.378348210  0.68611040 -0.53421937  0.383111100
#>  [16,] -1.24372650 -0.655448717  0.88375951 -1.99937556  1.466660019
#>  [17,] -0.21085394  0.398019417 -0.90120708 -0.44816807  0.539677700
#>  [18,]  0.07434740  0.912428014  0.27936915 -1.83670356 -0.300010679
#>  [19,] -0.31417809  0.627535099 -1.49657420 -0.42399725 -0.207461770
#>  [20,]  0.17144451  0.369144843 -0.19762782 -1.02569035  0.292997623
#>  [21,]  0.44000533 -0.554189306  1.79529283  0.84687469  0.181086436
#>  [22,]  1.69578305  1.138854642  0.42803315  1.03674768  0.519905563
#>  [23,] -0.70893637 -0.089382345 -2.21962668  0.02685537  0.754859990
#>  [24,]  2.00072509  0.663363588 -0.08506655  1.04124207 -0.316585641
#>  [25,]  0.16352930  1.306449549  0.25537912 -0.66743179  0.008304234
#>  [26,]  0.88094293 -0.462889264  1.97307229 -1.27817629 -0.799981844
#>  [27,] -0.87035535 -0.903524717  1.53436255  0.15089836  0.373671751
#>  [28,]  0.19815723  0.041066929 -0.17511513 -2.19148849 -2.168073850
#>  [29,] -0.90112031 -0.002224776  0.40384176 -0.72884467  0.715149971
#>  [30,]  0.07415803  0.311444655  2.48691044  0.07008647 -0.377123962
#>  [31,] -0.55818504  0.377087056  0.46711299 -0.54741241  1.982278730
#>  [32,]  0.21114274  1.782257727  0.92966750  3.60809803  0.748683396
#>  [33,] -0.07067894  1.688150369  0.39104294 -0.66252221 -0.033098935
#>               [,6]        [,7]         [,8]         [,9]        [,10]
#>   [1,]  0.37851087 -0.57219102  0.488897402  1.710988311  2.650354816
#>   [2,]  0.44566688  0.35579185  0.343154984 -0.550376537  1.808531875
#>   [3,]  0.26154000  0.44156179  1.745694473  0.141044820 -2.563719238
#>   [4,] -1.68890598 -0.11548520  0.156531736 -0.476414229 -0.593869928
#>   [5,]  0.33218792  1.25414755  0.040915329 -0.025677217 -0.725982534
#>   [6,] -0.07241143 -0.49093971  0.254556348  0.249593101 -0.670357504
#>   [7,]  0.77855806  1.14109685 -2.083676365 -0.694502550 -0.876389241
#>   [8,]  0.35048101 -0.88817736 -0.549030883  0.252709934 -0.472581858
#>   [9,] -0.35131386  0.67575601 -0.055783734  1.021585654 -0.072135258
#>  [10,]  0.67401984 -1.09310698 -0.564908039  1.713750540  0.758101208
#>  [11,] -1.03667772 -0.51680910  0.369838475 -0.204627421  1.070246861
#>  [12,]  0.85566863 -1.74628539  1.461044406  1.352897619  0.532693195
#>  [13,]  0.34664529 -0.78077312  0.053043206 -1.608069688  0.658389378
#>  [14,] -0.97175360  0.33049514 -1.584186101  0.086278498 -0.244155640
#>  [15,]  1.03964715 -0.08694785  1.590001495  0.061548123  1.224148341
#>  [16,] -0.08724902  1.25788900  0.220615368 -0.278237702  0.525661796
#>  [17,] -0.24337344 -0.90727228  0.397181619  0.749671341  0.582102809
#>  [18,]  0.83756382 -0.07475662  1.000524247 -0.511816152 -0.081225411
#>  [19,] -0.63175827  0.86061175  1.338087557 -1.750096050 -0.024878009
#>  [20,] -1.52994097 -1.03131556  1.237665863  0.995860504  0.525253409
#>  [21,]  0.13937778 -1.35455681  0.461526123  1.586855627 -0.652860854
#>  [22,] -1.21095409 -2.67724034  0.144645963 -0.702959917  2.281982268
#>  [23,] -1.36084115 -1.10828420  0.106933960 -0.847797585 -0.517353936
#>  [24,] -0.31441953 -0.97349306 -0.805488839 -0.426117517 -0.296860606
#>  [25,] -0.26456888 -0.26195464 -0.609155650 -0.192463983  0.879329897
#>  [26,] -0.22662791  0.47590518 -0.082702936 -0.213561551  0.488562541
#>  [27,] -0.81016435  0.25307614  0.310101495  0.615480262 -1.389581191
#>  [28,]  0.03973555  0.55462392 -0.723787167  0.416522978  0.905044008
#>  [29,]  0.85854371 -0.04972340 -0.247088121 -1.481890624  0.884553851
#>  [30,]  0.50023078 -0.49306880  0.123158018 -0.106236106 -0.822152112
#>  [31,]  2.44276048  2.52006628 -0.097227237  1.475538939 -1.983656298
#>  [32,]  0.65477136 -1.16031917  0.449158525  0.001298862 -1.234475289
#>  [33,] -0.54265565 -2.00959170  0.038074265  0.282783381  1.081489246
#>              [,11]         [,12]         [,13]         [,14]        [,15]
#>   [1,] -0.15168558 -0.2876166685  0.5368739236 -0.0741223963 -0.216744949
#>   [2,]  0.02552932 -1.6228753230  0.9119057688 -1.5463128939 -0.558609498
#>   [3,]  0.71542267  0.2428439428  0.7572440086  0.6428964853  1.197290449
#>   [4,]  2.50730611 -0.5283399123 -2.4117276691 -0.4012823706  0.835723486
#>   [5,] -0.83556272  0.8315481719 -1.4975329901 -0.4102525908 -1.922557723
#>   [6,]  0.41029907  0.5681883171  0.4528848396 -0.7450334015  0.472743338
#>   [7,]  0.10879180  0.6681667534 -0.5462602224 -2.2820576734 -1.353332529
#>   [8,]  0.04155468  0.0001097528 -0.1031282401 -0.2993468058  0.095506159
#>   [9,] -0.88179461  0.6028692606  0.0004746908 -0.0179072863 -1.714724761
#>  [10,] -1.08149171  1.1785586196  0.8162885018  0.1237787759  0.573082355
#>  [11,]  0.73681759 -0.9886286681 -1.0127117614  0.7451836585 -0.509473438
#>  [12,] -0.80519230  0.8891305257  0.0870832900  1.7341376741  0.051095471
#>  [13,] -0.69712731 -0.5849486876 -1.6776131352  1.1914368497 -0.603244972
#>  [14,]  2.20535948 -0.5872224955 -0.5691139868 -1.5104941110  0.832733533
#>  [15,]  0.87928277  0.8450729127 -0.9317664464  1.0079361654  0.210997940
#>  [16,] -0.66693638 -0.6594826126 -0.9884894042  0.3591449490 -2.033512683
#>  [17,]  0.65238837 -0.6503101427  1.0926777219 -0.5544215241  0.139249362
#>  [18,]  1.24871656 -1.3079836904  0.3082665955  1.6161429304  0.966398612
#>  [19,]  0.60995676 -0.3270526529 -0.4340210273 -0.3210471110 -1.485801464
#>  [20,]  1.53078921  0.5245239484 -4.1637229079 -0.8218483300 -0.989930354
#>  [21,]  0.75254536 -1.2314733062  0.2727240334  0.5335423251  1.434410592
#>  [22,] -2.51788877 -0.4552644666 -0.7448480294 -0.6127083520  1.402274718
#>  [23,] -0.07891372  0.3393176999  1.3786897942 -0.2341629175 -0.475566386
#>  [24,] -0.09674329 -0.4870604370 -1.6028041586 -1.2898177851 -0.834452324
#>  [25,]  0.58011826 -0.8501280769  0.2982426363  0.7829005890 -0.562654987
#>  [26,] -0.38307162 -0.5347043271 -1.4523416393 -2.2409548398  1.166459078
#>  [27,] -0.35551878 -1.2335067881  0.7366662655  0.6281065154  1.725239006
#>  [28,] -1.21722728  1.3848158814  0.2517359338 -0.4405988571  0.735081036
#>  [29,]  0.04171296 -1.5532817759 -0.5927646952  1.0199200242 -2.521824173
#>  [30,] -0.62183634  0.5206533044 -1.5308381518 -0.9278444348 -0.590997654
#>  [31,]  1.08526500  0.8893455570  1.5458965893 -2.1200792327  1.129654458
#>  [32,] -0.78189714  0.1538875161 -0.1337951574  2.2933971604 -0.244529177
#>  [33,]  0.60498823  0.9688936585  2.1681947253 -0.5394139993  1.307317730
#>                [,16]        [,17]       [,18]        [,19]       [,20]
#>   [1,] -0.1179688818 -0.024890797 -0.21356450 -2.880409907  0.02411560
#>   [2,]  1.0369074411 -0.119318296  0.30962108 -0.269173588  0.61822397
#>   [3,]  0.3300279484  0.899067750  1.20931039 -1.321985493  0.07233446
#>   [4,] -0.0135314113  0.343497766 -0.83847561 -0.238619553 -0.45585947
#>   [5,] -0.1474051291 -1.301742732  0.56628974  1.323446657  0.52046628
#>   [6,]  1.5234376135  0.641250874 -1.73573079  0.195701668 -0.96416140
#>   [7,] -1.8572914211  0.666410566  0.34667986  1.008791277  1.83267443
#>   [8,] -0.5805803963  1.126794098 -1.20461460  1.394508822  0.09102377
#>   [9,]  0.5442270038 -0.312976261  0.31559446 -1.222129044 -1.67996299
#>  [10,] -0.4501432891 -1.019387655 -0.68206369 -0.716950857 -0.12204875
#>  [11,]  0.9200440697  1.335642026  1.91112129  0.272381221  0.18134371
#>  [12,] -0.1974152049 -0.617665683  0.52272868  0.855255135  0.04794499
#>  [13,]  0.5960900757 -1.170815274  0.77817564  2.014180437  0.41954979
#>  [14,]  0.1283146351 -0.178409673 -0.35911750  0.176104312  1.05393673
#>  [15,] -0.1520199554  2.747445253 -0.08194879  0.504833713  0.48069269
#>  [16,] -0.8683524842 -0.604701586 -1.14201162  0.536960005  1.82368260
#>  [17,]  0.8456958843  1.045011080 -0.15906137 -0.107915971  1.88942615
#>  [18,]  0.6319101060 -0.064835033  0.52503770 -0.120106921  0.86562330
#>  [19,] -0.9438327627  1.094075594 -0.76978656  0.362643399 -0.59606592
#>  [20,] -0.3193311437  1.389779281  0.50993779  0.139211530 -2.34546435
#>  [21,] -0.8642798041  1.470133774  0.77301708 -0.009934425  0.31378011
#>  [22,] -0.4227302563 -0.059064332 -1.02736734  0.240719868 -0.44867275
#>  [23,] -0.2280388102  0.467909129  0.31909353 -0.355467192  0.72588262
#>  [24,]  1.2789106018  0.131073119  1.00063206  0.619557693  0.27764551
#>  [25,] -1.1908926147  0.594156071 -0.81019490  0.786343850  2.74657471
#>  [26,] -0.2204505007 -1.453127017 -0.19147823 -0.988655209  1.13687347
#>  [27,] -0.9650955634 -0.475680822  1.30804077  0.165508669 -0.27815898
#>  [28,]  0.9603833553 -2.079472237  0.64862018 -1.374672043  0.24576007
#>  [29,] -0.5708550036  0.369199863 -0.84746538  0.819914810  0.20218077
#>  [30,]  1.7000992845  1.783951467  2.17295842 -2.121197909  1.11201312
#>  [31,]  0.3325106386 -1.614735735  0.41508819  0.140385261 -1.59199932
#>  [32,]  0.0560065024 -0.034046009 -0.28939833  1.056497922  0.46463258
#>  [33,]  0.4869331210 -0.725237745  1.28611690 -1.320427644  0.32653112
#>               [,21]       [,22]        [,23]        [,24]       [,25]
#>   [1,] -0.194805529 -0.08359650  0.899120654 -0.573356827  1.04726807
#>   [2,]  0.650387644 -0.74213107  0.477200356 -0.042931444  1.62251318
#>   [3,]  0.492746068  0.64931132 -2.267834820  0.373265830 -1.78145516
#>   [4,]  0.035255625  1.79078334  0.784633862 -1.865942097 -0.05858237
#>   [5,] -1.508986935 -0.29822364  0.331016695 -0.463581695 -0.23229147
#>   [6,]  0.947746782 -1.35978340 -1.087518817  0.323744140 -1.58579414
#>   [7,]  0.538798111 -1.48583867  0.891752284  0.624834956 -1.15110819
#>   [8,] -1.247785608 -0.24684618 -0.088383718  1.363864817 -0.20008413
#>   [9,] -0.307147602  2.16844986 -1.535098829 -0.356264380  0.59298842
#>  [10,]  1.696686728  0.26052048  0.628051254  0.397422618 -0.31130976
#>  [11,]  0.325147985 -0.05441325 -1.503108890  1.599681917  0.84707329
#>  [12,]  0.280379327  0.30062965 -0.073177300 -0.833343139 -0.28182161
#>  [13,]  1.788072794  0.27978255  1.142746102  0.832397277  1.13459759
#>  [14,] -1.071197260  0.44379034  0.211050272 -0.748799034  0.21950061
#>  [15,] -0.731394307 -0.04456162  0.237391237  0.926672285 -1.73311031
#>  [16,] -0.683481785 -0.83484167 -2.045398266  0.493647419 -0.14749913
#>  [17,] -1.403117996 -0.04729694  0.401524946 -1.986908443  1.00752998
#>  [18,]  1.805566915 -1.72578162  2.530207092 -0.971082407  0.11789022
#>  [19,]  3.066129437 -1.35727116 -0.514904945  0.518988808  2.22056637
#>  [20,] -0.075060694  0.53171481 -1.376449355  1.107874482  1.51380765
#>  [21,] -1.275747916  1.04719059 -0.008728628  0.794447877  1.83817730
#>  [22,]  1.938314506  0.29898332 -0.293860990  0.750864375 -0.23387334
#>  [23,]  0.670512327 -1.55744085 -0.273321154  1.566373637 -0.45430367
#>  [24,] -0.733889480 -0.50348035 -0.569386269  0.201621693 -0.23593277
#>  [25,]  0.717277980  0.12694278  1.072382915 -0.953599528 -0.48033953
#>  [26,]  0.864195697  0.73887517 -1.235958385 -0.918614010 -0.15080965
#>  [27,] -1.256846589  0.93913243  1.179330433 -0.176626893 -0.82575552
#>  [28,]  1.913662543  1.23460666 -0.101222731 -0.429470903  0.80210557
#>  [29,] -2.920506093  0.38213615  1.520858518  0.765615417 -1.56279154
#>  [30,]  0.533287847  0.51516314  0.381797937  0.297896477 -0.33541095
#>  [31,] -1.072535324  0.08431047  1.189745763 -2.181650740  1.46593435
#>  [32,]  0.244005470  0.97639529  0.834840558  0.088373946  0.75781058
#>  [33,] -0.311209859 -0.32764497  1.850481484  0.552569837 -0.45538225
#>              [,26]       [,27]        [,28]       [,29]        [,30]
#>   [1,]  0.70447896 -0.24493808  0.823219111 -1.95782240 -1.063858409
#>   [2,] -1.05082874 -0.10618477 -0.600258110  0.79343466 -2.056833950
#>   [3,] -0.85137058 -2.03349192 -1.254497529  1.16888031  0.872887142
#>   [4,] -0.58582325  2.27556262 -0.096542097 -0.12091656  0.309518158
#>   [5,]  0.19730360  1.27992902 -0.990008367 -0.70679792 -0.900999843
#>   [6,]  1.50160485 -1.01723433  0.197516654  1.46379850 -0.413364326
#>   [7,] -1.57392986  0.41195829 -1.754047735 -0.75992365 -0.140781170
#>   [8,]  0.19924062 -1.41374191 -0.112964503 -1.25132599 -0.347970981
#>   [9,] -0.85117623  1.89264825 -0.220318410  1.15399897 -0.923691494
#>  [10,] -0.41041212  0.59781619 -1.650257019  0.88464230 -0.947582920
#>  [11,]  0.33001618 -1.09262274  0.673111740  0.96750309  1.244866485
#>  [12,]  0.53026140 -0.53459976  0.479400631  0.48104607 -0.401446131
#>  [13,]  1.95082523  0.13208638  0.421752455 -0.42320666  1.122883379
#>  [14,] -0.74305794  1.79324870 -0.920535536  0.15607731  1.988424123
#>  [15,]  0.02573082  0.76263201 -0.037052581 -0.82215296  0.634999103
#>  [16,]  0.89003987 -1.04114447 -1.447830608 -1.52354009 -1.689308813
#>  [17,] -0.86137519  0.85190958  0.361831609  0.51894761 -0.008174066
#>  [18,]  0.74375055  1.82482127 -0.324760565  0.33124332 -0.559811282
#>  [19,] -0.53294303  0.09631463 -1.086100730 -0.50097692  0.404140456
#>  [20,] -0.53862557  1.25679254 -0.502961226 -1.45291742  0.985758857
#>  [21,]  0.94434812 -0.56568184 -0.232245396 -0.28685716 -0.186849381
#>  [22,]  1.14068886 -0.11727483 -0.197525421  0.79777980 -1.040493075
#>  [23,] -1.44207965 -1.43175158  0.540776001  0.24042482  0.039713020
#>  [24,] -0.53502838  0.54605970 -1.004283412  0.35024354 -1.193516529
#>  [25,]  0.06898262 -1.27415813 -0.217827564  0.86557853 -1.987959027
#>  [26,] -0.83430001 -0.26842636 -1.601626475 -0.93158468 -0.407807569
#>  [27,]  0.16261282  2.88606102  0.913227133  0.37539102  0.170450954
#>  [28,]  0.32055018 -0.26952085 -0.579579915 -1.17567908 -0.185594314
#>  [29,] -1.69133575 -0.08977804 -0.442225600  0.29176769  0.251096938
#>  [30,] -2.08010771 -1.20854347  0.434042222  0.29772063 -0.588520838
#>  [31,]  1.38778287  0.91292510 -2.464046277  0.44500691  0.533500073
#>  [32,] -1.69585607  0.37001434 -1.761199931 -1.32353572  2.166985580
#>  [33,] -0.71209647  0.33460535 -2.292689218  1.23761944 -0.614730013
#>  [ reached getOption("max.print") -- omitted 67 rows ]
#> 
#> $y
#>                [,1]          [,2]          [,3]         [,4]          [,5]
#>   [1,]  0.035327823  -9.419331772 -5.304745e-04   9.23128464 -0.0044124315
#>   [2,]  0.156843199   1.283833445 -3.069873e-01  -1.63431549  0.1620750314
#>   [3,] -0.012871448   0.515323551 -1.120914e-01  -0.06453677  0.1360785958
#>   [4,]  0.208424934  -0.533186651  2.079602e-01   0.37183461  0.0082170801
#>   [5,] -0.310766048 -11.332535233  1.146556e+01  11.62519105  0.0961843799
#>   [6,] -0.001789709   6.545888271 -1.283613e-02  -6.95833881  0.0510027076
#>   [7,] -0.058773954   3.937225412  1.329414e-01  -4.19616157 -0.0355803486
#>   [8,] -0.425490799   5.814022623  1.164230e-01  -6.07186156  0.0584483119
#>   [9,] -0.089166023  -9.784097287  9.527283e+00   9.32625532 -0.0237649920
#>  [10,] -0.197455821  -4.505583691  4.863372e+00   4.79371274  0.2931431735
#>  [11,] -0.164011037   1.102694111 -7.305477e-01  -1.48312076  0.0404040933
#>  [12,]  0.385016024  -4.442882208 -1.208360e-01   4.32225016  0.5066583977
#>  [13,] -0.077766948  -0.881790770 -3.427182e-02   1.07000894 -0.3436649620
#>  [14,] -0.326310143   5.637440158 -5.222070e+00  -5.77000013  0.3687842135
#>  [15,]  0.031504402   0.277108395  2.042477e-01  -0.28911310 -0.0797062178
#>  [16,]  0.222734663   0.286401009 -7.112494e-01  -0.05747980 -0.2878674379
#>  [17,] -0.038317260  -5.341471791 -2.661181e-01   5.00451663  0.1010393744
#>  [18,]  0.100434172  -2.507553073 -1.785161e-01   2.82896238  0.1396279570
#>  [19,]  0.053855865  -8.213537556  8.217281e+00   8.76177538 -0.1428450702
#>  [20,] -0.052358770 -11.878090131  1.268165e+01  12.29116855 -0.3074680468
#>  [21,]  0.050396942  -7.858743032 -3.590169e-01   7.87247262 -0.3838217188
#>  [22,] -0.136646487  -3.572762623 -1.227020e-01   3.52255927 -0.0344005683
#>  [23,] -0.095109734  -7.157601617  4.734450e-01   7.36754832 -0.2233551281
#>  [24,]  0.104514657  -0.006066004  1.090369e-01  -0.33383664 -0.0375451485
#>  [25,]  0.034258450   3.842750289  1.183338e-01  -4.01660466  0.3144173611
#>  [26,] -0.027989436   6.078037403 -6.684465e+00  -6.67838513 -0.2793231398
#>  [27,] -0.067353533  -2.758596185 -4.961738e-01   3.13669601  0.2107068868
#>  [28,] -0.107325841  -4.961558130  9.008998e-02   4.77378483 -0.1472926203
#>                 [,6]        [,7]          [,8]          [,9]         [,10]
#>   [1,] -0.1191430533  0.24151466 -0.1925569701   9.060580820   9.461402614
#>   [2,]  0.1491893155 -0.16411680  0.0597651921  -1.481794600  -1.728322925
#>   [3,]  0.2867733871  0.31410413  0.2408229715  -0.121862991  -0.002239617
#>   [4,]  0.0361114456  0.06860127 -0.0223287814   0.026329590  -0.421529495
#>   [5,]  0.1363253812  0.11564690 -0.1681050426   0.260682406 -11.358436689
#>   [6,] -0.1075204537 -0.07562742 -0.0463404420  -6.625297508  -6.539965415
#>   [7,] -0.2128660647 -0.12895147 -0.2073011444  -4.224357719  -4.063547907
#>   [8,] -0.0510275325  0.31858731 -0.0381453130  -5.885492842  -6.247457158
#>   [9,]  0.0341755966  0.16418496 -0.1127130756  -0.276180272  -9.551409166
#>  [10,]  0.2353291550 -0.11719003 -0.2530063700   0.176310040  -4.287442077
#>  [11,]  0.1775982362 -0.05900670 -0.2879048601   0.058207450   0.409358415
#>  [12,] -0.0761272213 -0.15192015 -0.0924077805   4.966669924   4.866827652
#>  [13,] -0.0139121106 -0.25774466  0.3203127370   0.782727396   1.178723898
#>  [14,] -0.3723278149 -0.39588480 -0.1149323182   0.437075143   5.771636361
#>  [15,] -0.0350911425 -0.35894584 -0.4417422734   0.170851765  -0.576861295
#>  [16,] -0.0158186517 -0.05462114 -0.1349833684  -0.274145031   0.288887555
#>  [17,]  0.0312981682  0.26071539 -0.0170805865   5.155738070   5.601321937
#>  [18,]  0.0893615534  0.33215341 -0.1932935078   2.940888026   2.736073525
#>  [19,]  0.2168838718  0.19616271  0.1952357483  -0.080222634  -7.990027053
#>  [20,]  0.4470086362  0.05316002 -0.1994371877  -0.233756018 -11.974039986
#>  [21,] -0.0335064987 -0.10841431  0.0726599726   7.612975259   7.504105430
#>  [22,] -0.2627159196 -0.10256921 -0.0623985056   3.396407767   3.359557175
#>  [23,] -0.0268014523  0.32662381 -0.0171494074   7.066154226   7.545047598
#>  [24,] -0.0097444748  0.12802457 -0.0789888311   0.204091254   0.341741188
#>  [25,] -0.3275144611  0.20390299 -0.2786648597  -4.181171376  -4.118968702
#>  [26,]  0.2296379564 -0.21145385  0.1930108687   0.257493863   6.585508644
#>  [27,] -0.0888610565 -0.03374397  0.1363451225   2.710791680   3.052858652
#>  [28,]  0.2093751729  0.07010210  0.1266835910   5.109316240   4.398299689
#>                [,11]         [,12]        [,13]        [,14]        [,15]
#>   [1,]  -9.634095968   9.236389354  0.248735694  0.256750339 -0.097395264
#>   [2,]   1.562139813  -1.385107802 -0.032388921 -0.095091117 -0.057039324
#>   [3,]   0.178534863   0.298908559  0.228522577  0.424186535 -0.188786931
#>   [4,]  -0.031911670   0.042103519 -0.229483851  0.231233714 -0.032606727
#>   [5,]   0.083659753  -0.587087164 -0.235180385 -0.229448918 -0.273678644
#>   [6,]   6.687298819  -6.516210829  0.223789236  0.004223829 -0.207417087
#>   [7,]   4.087075957  -4.087554223 -0.588247266  0.118041126  0.171008325
#>   [8,]   5.954321620  -5.934027475 -0.275095586  0.074891337 -0.114429202
#>   [9,]  -0.309253669  -0.361504284  0.129556317 -0.246377201 -0.097414255
#>  [10,]   0.027900340  -0.193430104 -0.396281234 -0.128926477 -0.062664219
#>  [11,]   0.184018223  -0.243157766  0.013951344 -0.227104755 -0.240508509
#>  [12,]  -4.800302416   4.932450370 -0.067345103 -0.003925100  0.160280019
#>  [13,]  -1.047655376   1.199900991 -0.242150475  0.097555186 -0.025544142
#>  [14,]   0.236718587  -0.099497110 -0.065894294 -0.069965663 -0.313397271
#>  [15,]   0.019956690  -0.264784324  0.123468270  0.158553555  0.487027948
#>  [16,]   0.309209491   0.085462952  0.129186899 -0.235636875 -0.223835046
#>  [17,]  -5.223730882   5.068816393 -0.269927236  0.492291895  0.069107454
#>  [18,]  -3.027520205   2.656720267 -0.152326918  0.101479674 -0.556743403
#>  [19,]  -0.283223080  -0.233709403  0.047690408 -0.250508007  0.315010421
#>  [20,]  -0.081670593  -0.309760847 -0.368135181  0.549378745  0.210016767
#>  [21,]  -7.851854099   7.767979585  0.048581527  0.179870257  0.260395218
#>  [22,]  -3.297249670   3.312974650 -0.155080593 -0.096538448 -0.334085927
#>  [23,]  -7.265519491   7.466489961  0.059688568  0.319868927 -0.132130388
#>  [24,]  -0.266763667   0.148439418  0.410833149 -0.225778233 -0.025436054
#>  [25,]   4.317620525  -4.022623777 -0.107714854  0.063909548 -0.141909855
#>  [26,]  -0.144411583  -0.147478174  0.107414870 -0.053175682 -0.640672074
#>  [27,]  -2.871335642   3.020397598  0.015973171  0.113554341 -0.330132801
#>  [28,]  -4.879088449   4.692344350  0.253321849 -0.018683316 -0.108186094
#>                [,16]        [,17]        [,18]        [,19]         [,20]
#>   [1,] -4.152388e-02  0.005093689   0.41383599 -0.305520263 -9.517527e+00
#>   [2,]  2.283375e-01  0.057006907  -0.10582631  0.584521793  2.049893e+00
#>   [3,]  2.907314e-01  0.204310926  -0.34388974  0.358296325  1.422945e-01
#>   [4,]  2.691835e-01  0.296815642   0.70597275  0.097273533 -1.259258e-03
#>   [5,]  1.144367e+01 -0.041404819  11.21615871  0.010068430  1.175391e-01
#>   [6,] -3.994297e-01  0.213949305  -0.07348717 -0.213720917  6.600063e+00
#>   [7,] -1.795940e-01  0.101422375   0.11681186 -0.100917034  4.247486e+00
#>   [8,]  5.677077e-02 -0.041877067   0.25186676  0.114920103  6.255524e+00
#>   [9,]  1.006955e+01  0.165564654   9.39670150 -0.172056880 -1.529723e-01
#>  [10,]  4.301388e+00  0.181228021   3.93123225  0.364208966  1.766075e-01
#>  [11,] -4.618847e-01 -0.211710198  -1.10489477 -0.028300131 -1.916417e-01
#>  [12,]  4.242608e-01 -0.024682927   0.02625675 -0.420527889 -4.661404e+00
#>  [13,]  1.028826e-01  0.483250320   0.05313261  0.193006564 -1.491397e+00
#>  [14,] -6.102661e+00  0.197566916  -5.57875756 -0.045465513  2.990198e-02
#>  [15,]  2.951576e-01 -0.033628655  -0.26886464 -0.264356230  6.696915e-01
#>  [16,] -2.789698e-01 -0.267646727  -0.36964739  0.022020263  1.953935e-01
#>  [17,]  1.037623e-01  0.100659564  -0.01964264 -0.238910090 -5.282720e+00
#>  [18,]  1.774540e-01 -0.065148958   0.11693857  0.066633956 -3.066822e+00
#>  [19,]  7.943624e+00 -0.062530293   8.07967192 -0.240627750  2.841138e-01
#>  [20,]  1.216421e+01  0.314369602  11.96121737 -0.001696908 -8.085125e-02
#>  [21,] -1.855450e-01 -0.242387966   0.17998877  0.044090029 -7.539348e+00
#>  [22,] -3.161327e-01 -0.102190415   0.13554659  0.559012170 -2.940122e+00
#>  [23,]  4.629596e-02  0.355038402  -0.01645260  0.021309966 -7.327824e+00
#>  [24,]  2.505763e-01 -0.267157522  -0.31699444  0.043964020 -7.439220e-02
#>  [25,] -1.955395e-02  0.179787204   0.12973709  0.187845440  3.885310e+00
#>  [26,] -6.414904e+00  0.177289871  -6.88944045  0.102341115 -2.817965e-02
#>  [27,] -3.708308e-01 -0.081554743   0.00274078 -0.024899764 -3.005621e+00
#>  [28,] -2.410993e-01 -0.037454030   0.09150710  0.053682810 -4.955394e+00
#>                [,21]        [,22]        [,23]        [,24]         [,25]
#>   [1,]  0.0490289655  0.288759501 -0.119384389   9.67938796 -0.2445369691
#>   [2,] -0.1451703561  0.174464536  0.428802463  -1.49460470 -0.1403236842
#>   [3,] -0.3360242302 -0.329054631 -0.114885949   0.01463026 -0.1207199328
#>   [4,]  0.3310330436  0.112640346  0.029465107   0.60524195 -0.2323324792
#>   [5,]  0.4284803095 -0.307200622  0.239918915  11.51131544  0.4119595099
#>   [6,]  0.0040614854  0.033910832  0.182547417  -6.65170517  0.0322978227
#>   [7,] -0.0556700125 -0.217976599  0.254246471  -4.21631481  0.1688391134
#>   [8,]  0.0566279595 -0.615895308  0.147036304  -6.01199926  0.0479749287
#>   [9,]  0.4473390526 -0.071048451  0.084082967   9.72572792  0.0251822858
#>  [10,] -0.3044501460 -0.132452369 -0.102571558   4.65461032 -0.0823105972
#>  [11,] -0.2685675063  0.100643048  0.065614962  -1.22508723  0.1980947326
#>  [12,] -0.1396367891  0.259860557  0.153453192   4.63219083  0.2214593611
#>  [13,]  0.1434743500 -0.103640861  0.395068870   0.69588812  0.5744460159
#>  [14,]  0.3486991906 -0.240189736 -0.033770848  -5.70947603 -0.1318889296
#>  [15,]  0.0074544348 -0.076441687 -0.015517990  -0.07474830  0.2184514763
#>  [16,] -0.2785462555 -0.026754906 -0.051610147  -0.25367010  0.0729839308
#>  [17,]  0.1407554878 -0.071361649  0.164601651   5.46412916  0.0822780992
#>  [18,]  0.0994774946 -0.001251129 -0.223189977   2.70604477  0.1389126119
#>  [19,] -0.3219067377 -0.328326700  0.175195356   8.43008730 -0.0427316955
#>  [20,] -0.0434674115  0.224970835 -0.077255550  11.80607704  0.1714264722
#>  [21,] -0.1800908621  0.062467540 -0.133107026   7.77238415  0.1459574999
#>  [22,]  0.4022383350  0.059397913  0.232590909   3.45142704 -0.1057678104
#>  [23,] -0.2286755198 -0.216153876 -0.204513690   7.37823385 -0.1141876741
#>  [24,] -0.1485123220 -0.020642720  0.008249325  -0.15563054  0.1374082095
#>  [25,] -0.3442223426  0.113271831  0.236219071  -3.95697449  0.0033670719
#>  [26,]  0.0559421348  0.578332296 -0.013971890  -6.39226615 -0.1404780460
#>  [27,] -0.0343021761  0.440799308 -0.266132553   3.23520248  0.0574626515
#>  [28,]  0.3438200173 -0.474419894 -0.126208929   5.09025761 -0.0555518036
#>                [,26]         [,27]        [,28]         [,29]         [,30]
#>   [1,]   0.403392342  -0.066316359  0.148336661   9.556471115 -0.0760051471
#>   [2,]  -0.046424827   0.177905533 -0.180084723  -1.252207733 -0.2559436882
#>   [3,]   0.006984231  -0.341122161  0.112251768   0.516627305 -0.1346411601
#>   [4,]  -0.499423594  -0.668852857  0.097908585   0.121231234 -0.3202104597
#>   [5,] -11.503658783 -11.352703986 -0.052883502  -0.359831145  0.1619576715
#>   [6,]  -0.217249997  -0.152905110 -0.136676547  -6.756198219 -0.1824097109
#>   [7,]  -0.056948478  -0.149026185 -0.294905401  -4.212137989 -0.1951740297
#>   [8,]  -0.172476047  -0.145613925  0.170714798  -6.251372719 -0.0001673083
#>   [9,]  -9.299504491  -9.287029278  0.325250886  -0.392107928 -0.4395297719
#>  [10,]  -4.469787347  -4.809473152  0.326294297  -0.262377249 -0.0133331130
#>  [11,]   0.890857532   1.471432619 -0.124353361  -0.209675955  0.1041274292
#>  [12,]  -0.112495416   0.160300219  0.217399278   4.962422493 -0.3341788324
#>  [13,]  -0.146119554  -0.164106952 -0.192289571   0.890952850  0.2481643959
#>  [14,]   6.125088525   5.839897718 -0.363226553   0.140485886  0.0499701065
#>  [15,]   0.474949717   0.329952287  0.148175588  -0.463271213 -0.1038090590
#>  [16,]   0.427103447   0.279893578 -0.177665988  -0.110176996 -0.5257657049
#>  [17,]   0.053161412   0.067253580 -0.155394207   5.516088955 -0.0599441181
#>  [18,]   0.115621080   0.158822715 -0.256994389   2.746214309  0.0321583278
#>  [19,]  -8.083921453  -8.618968003 -0.295505563  -0.298239216  0.2942549213
#>  [20,] -11.980838727 -12.379910108  0.330882026  -0.116417928  0.2129975972
#>  [21,]  -0.324591816  -0.135782296  0.031082689   7.723075825 -0.0195280274
#>  [22,]   0.263151210  -0.032005209  0.020444114   3.086372202 -0.0292763919
#>  [23,]  -0.128424610  -0.149839767 -0.126805625   7.431517683  0.4934524243
#>  [24,]   0.162016681   0.116575661 -0.284873374   0.264253942  0.0750842882
#>  [25,]  -0.342009168  -0.206180046  0.107447447  -4.199059282 -0.1719866621
#>  [26,]   6.480836172   6.432329790  0.011598038  -0.025893210 -0.0757294465
#>  [27,]  -0.033528387  -0.215306495 -0.016553337   3.101179776  0.2816847841
#>  [28,]  -0.212652180  -0.113404677 -0.089191043   4.453115106 -0.1651487104
#>               [,31]        [,32]         [,33]         [,34]        [,35]
#>   [1,] -0.063563049  -9.26725767  -0.390998863 -6.828741e-05  0.157927065
#>   [2,] -0.313663894   1.27742808  -0.044065447 -4.129165e-01  0.061052831
#>   [3,] -0.282016467   0.34416368  -0.162507124 -3.988214e-02  0.056810332
#>   [4,] -0.036523103   0.52578156  -0.538851039  2.171444e-01  0.130909188
#>   [5,]  0.230605069   0.04235453 -11.292857790  2.677860e-02  0.130482611
#>   [6,]  0.523629440   6.67391529   0.209384474 -2.680911e-01  0.226542009
#>   [7,] -0.268169372   3.98045733   0.175096021  2.945521e-01 -0.034307912
#>   [8,]  0.078693504   5.55059335   0.261856349 -1.994484e-01  0.019863806
#>   [9,] -0.213407914   0.08496591  -9.334153767  9.636482e-03 -0.208501514
#>  [10,]  0.063583537   0.28478306  -4.845265384 -1.925803e-01 -0.122141365
#>  [11,]  0.421018002   0.42959015   1.328474344  1.124729e-01  0.019888708
#>  [12,] -0.012234252  -4.79005821   0.202537148  1.804605e-01 -0.011423257
#>  [13,] -0.297136242  -0.82511678  -0.140558076  7.309319e-02  0.477884855
#>  [14,]  0.196307102   0.33822191   5.931149853 -2.634829e-01  0.138433805
#>  [15,] -0.010135662   0.50238278  -0.298916714  5.596422e-02  0.066322473
#>  [16,]  0.028282708   0.15500749   0.425154394  1.143731e-02  0.417152934
#>  [17,]  0.120496019  -5.17952102  -0.044492620  1.195495e-02 -0.024327470
#>  [18,] -0.152035650  -3.28017593  -0.253095683 -1.191878e-01  0.148426143
#>  [19,] -0.492108601  -0.22437605  -8.114867723  1.693154e-01  0.042667162
#>  [20,]  0.309545544  -0.17852889 -12.320141591 -2.909783e-01  0.170107411
#>  [21,] -0.114848203  -7.40171497   0.360997365  1.503127e-01  0.084443508
#>  [22,] -0.248362235  -3.32610338  -0.050788500 -3.694846e-01  0.009462688
#>  [23,]  0.001212405  -7.06804365   0.246867867 -1.775471e-01 -0.168295964
#>  [24,]  0.246125012  -0.42098930   0.507053635 -8.228283e-02  0.093832347
#>  [25,] -0.195516937   3.97668309  -0.183315088  3.827617e-02  0.309348519
#>  [26,] -0.179900738   0.36494260   6.540238187 -1.167187e-01  0.048643083
#>  [27,] -0.025276614  -2.93000824  -0.187349730 -8.407198e-02  0.159596713
#>  [28,]  0.302567307  -4.84653557   0.108019969 -3.338914e-02  0.172690134
#>  [ reached getOption("max.print") -- omitted 72 rows ]
#> 
#> $a
#> $a[[1]]
#>       [,1]      [,2]      [,3]      [,4] [,5] [,6] [,7] [,8] [,9]     [,10]
#>  [1,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#>  [2,]    0 -1.620495  1.620495  1.620495    0    0    0    0    0 -1.620495
#>  [3,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#>  [4,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#>  [5,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#>  [6,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#>  [7,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#>  [8,]    0 -3.397702  3.397702  3.397702    0    0    0    0    0 -3.397702
#>  [9,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [10,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [11,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [12,]    0 -3.761768  3.761768  3.761768    0    0    0    0    0 -3.761768
#> [13,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [14,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [15,]    0  2.436885 -2.436885 -2.436885    0    0    0    0    0  2.436885
#> [16,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [17,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [18,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [19,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [20,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [21,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [22,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [23,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [24,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [25,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [26,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#> [27,]    0 -2.395894  2.395894  2.395894    0    0    0    0    0 -2.395894
#> [28,]    0  0.000000  0.000000  0.000000    0    0    0    0    0  0.000000
#>       [,11] [,12] [,13] [,14] [,15]     [,16] [,17]     [,18] [,19] [,20] [,21]
#>  [1,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#>  [2,]     0     0     0     0     0  1.620495     0  1.620495     0     0     0
#>  [3,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#>  [4,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#>  [5,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#>  [6,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#>  [7,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#>  [8,]     0     0     0     0     0  3.397702     0  3.397702     0     0     0
#>  [9,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [10,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [11,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [12,]     0     0     0     0     0  3.761768     0  3.761768     0     0     0
#> [13,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [14,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [15,]     0     0     0     0     0 -2.436885     0 -2.436885     0     0     0
#> [16,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [17,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [18,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [19,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [20,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [21,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [22,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [23,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [24,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [25,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [26,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#> [27,]     0     0     0     0     0  2.395894     0  2.395894     0     0     0
#> [28,]     0     0     0     0     0  0.000000     0  0.000000     0     0     0
#>       [,22] [,23]     [,24] [,25]     [,26]     [,27] [,28] [,29] [,30] [,31]
#>  [1,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#>  [2,]     0     0  1.620495     0 -1.620495 -1.620495     0     0     0     0
#>  [3,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#>  [4,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#>  [5,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#>  [6,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#>  [7,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#>  [8,]     0     0  3.397702     0 -3.397702 -3.397702     0     0     0     0
#>  [9,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [10,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [11,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [12,]     0     0  3.761768     0 -3.761768 -3.761768     0     0     0     0
#> [13,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [14,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [15,]     0     0 -2.436885     0  2.436885  2.436885     0     0     0     0
#> [16,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [17,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [18,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [19,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [20,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [21,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [22,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [23,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [24,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [25,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [26,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#> [27,]     0     0  2.395894     0 -2.395894 -2.395894     0     0     0     0
#> [28,]     0     0  0.000000     0  0.000000  0.000000     0     0     0     0
#>       [,32]     [,33] [,34] [,35]
#>  [1,]     0  0.000000     0     0
#>  [2,]     0 -1.620495     0     0
#>  [3,]     0  0.000000     0     0
#>  [4,]     0  0.000000     0     0
#>  [5,]     0  0.000000     0     0
#>  [6,]     0  0.000000     0     0
#>  [7,]     0  0.000000     0     0
#>  [8,]     0 -3.397702     0     0
#>  [9,]     0  0.000000     0     0
#> [10,]     0  0.000000     0     0
#> [11,]     0  0.000000     0     0
#> [12,]     0 -3.761768     0     0
#> [13,]     0  0.000000     0     0
#> [14,]     0  0.000000     0     0
#> [15,]     0  2.436885     0     0
#> [16,]     0  0.000000     0     0
#> [17,]     0  0.000000     0     0
#> [18,]     0  0.000000     0     0
#> [19,]     0  0.000000     0     0
#> [20,]     0  0.000000     0     0
#> [21,]     0  0.000000     0     0
#> [22,]     0  0.000000     0     0
#> [23,]     0  0.000000     0     0
#> [24,]     0  0.000000     0     0
#> [25,]     0  0.000000     0     0
#> [26,]     0  0.000000     0     0
#> [27,]     0 -2.395894     0     0
#> [28,]     0  0.000000     0     0
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> $a[[2]]
#>       [,1]      [,2] [,3]      [,4] [,5] [,6] [,7] [,8]      [,9]     [,10]
#>  [1,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#>  [2,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#>  [3,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#>  [4,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#>  [5,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#>  [6,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#>  [7,]    0  3.624407    0 -3.624407    0    0    0    0 -3.624407 -3.624407
#>  [8,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#>  [9,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [10,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [11,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [12,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [13,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [14,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [15,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [16,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [17,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [18,]    0 -3.501061    0  3.501061    0    0    0    0  3.501061  3.501061
#> [19,]    0  3.338333    0 -3.338333    0    0    0    0 -3.338333 -3.338333
#> [20,]    0 -1.215743    0  1.215743    0    0    0    0  1.215743  1.215743
#> [21,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [22,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [23,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [24,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [25,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [26,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [27,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#> [28,]    0  0.000000    0  0.000000    0    0    0    0  0.000000  0.000000
#>           [,11]     [,12] [,13] [,14] [,15] [,16] [,17] [,18] [,19]     [,20]
#>  [1,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#>  [2,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#>  [3,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#>  [4,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#>  [5,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#>  [6,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#>  [7,]  3.624407 -3.624407     0     0     0     0     0     0     0  3.624407
#>  [8,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#>  [9,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [10,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [11,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [12,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [13,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [14,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [15,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [16,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [17,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [18,] -3.501061  3.501061     0     0     0     0     0     0     0 -3.501061
#> [19,]  3.338333 -3.338333     0     0     0     0     0     0     0  3.338333
#> [20,] -1.215743  1.215743     0     0     0     0     0     0     0 -1.215743
#> [21,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [22,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [23,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [24,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [25,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [26,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [27,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#> [28,]  0.000000  0.000000     0     0     0     0     0     0     0  0.000000
#>       [,21] [,22] [,23]     [,24] [,25] [,26] [,27] [,28]     [,29] [,30] [,31]
#>  [1,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#>  [2,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#>  [3,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#>  [4,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#>  [5,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#>  [6,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#>  [7,]     0     0     0 -3.624407     0     0     0     0 -3.624407     0     0
#>  [8,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#>  [9,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [10,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [11,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [12,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [13,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [14,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [15,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [16,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [17,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [18,]     0     0     0  3.501061     0     0     0     0  3.501061     0     0
#> [19,]     0     0     0 -3.338333     0     0     0     0 -3.338333     0     0
#> [20,]     0     0     0  1.215743     0     0     0     0  1.215743     0     0
#> [21,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [22,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [23,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [24,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [25,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [26,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [27,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#> [28,]     0     0     0  0.000000     0     0     0     0  0.000000     0     0
#>           [,32] [,33] [,34] [,35]
#>  [1,]  0.000000     0     0     0
#>  [2,]  0.000000     0     0     0
#>  [3,]  0.000000     0     0     0
#>  [4,]  0.000000     0     0     0
#>  [5,]  0.000000     0     0     0
#>  [6,]  0.000000     0     0     0
#>  [7,]  3.624407     0     0     0
#>  [8,]  0.000000     0     0     0
#>  [9,]  0.000000     0     0     0
#> [10,]  0.000000     0     0     0
#> [11,]  0.000000     0     0     0
#> [12,]  0.000000     0     0     0
#> [13,]  0.000000     0     0     0
#> [14,]  0.000000     0     0     0
#> [15,]  0.000000     0     0     0
#> [16,]  0.000000     0     0     0
#> [17,]  0.000000     0     0     0
#> [18,] -3.501061     0     0     0
#> [19,]  3.338333     0     0     0
#> [20,] -1.215743     0     0     0
#> [21,]  0.000000     0     0     0
#> [22,]  0.000000     0     0     0
#> [23,]  0.000000     0     0     0
#> [24,]  0.000000     0     0     0
#> [25,]  0.000000     0     0     0
#> [26,]  0.000000     0     0     0
#> [27,]  0.000000     0     0     0
#> [28,]  0.000000     0     0     0
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> 
#> $true
#>   [1] 2 2 1 1 1 2 2 2 1 1 1 2 2 1 2 1 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 1 1 2 2
#>  [38] 2 2 2 2 1 2 1 1 1 2 1 2 2 1 2 2 1 1 1 1 2 2 1 2 2 2 2 2 2 1 1 1 1 1 2 2 1
#>  [75] 1 1 2 1 2 1 2 1 2 2 1 1 2 2 1 2 2 1 2 1 2 1 2 2 1 1
#>