Simulate Heterogeneous, Low-Rank, and Sparse Data

Simulate Heterogeneous, Low-Rank, and Sparse Data

Usage

simulate_lsr(
  N = 100,
  k = 2,
  p = 30,
  m = 35,
  b = 1,
  d = 20,
  h = 0.2,
  case = "independent"
)

Arguments

N

The sample size, default = 100.

k

The number of groups, default = 2.

p

The number of predictor features, default = 30.

m

The number of response features, default = 35.

b

The signal-to-noise ratio, default = 1.

d

The singular value, default = 20.

h

The lower bound for the singular matrix simulation, default = 0.2.

case

The covariance case, "independent" or "dependent", default = "independent".

Value

A list of simulation values, including x matrix, y matrix, coefficients and true clustring assignments.

Examples

simulate_lsr()
#> $x
#>               [,1]        [,2]        [,3]         [,4]         [,5]
#>   [1,]  1.03684888 -0.16848852  0.73386647 -1.345756210 -0.005396678
#>   [2,]  0.35787232 -0.34378850 -0.93546188  1.165418365 -0.506856641
#>   [3,]  0.65810698 -0.28963655 -0.52386120 -0.675981304 -1.606791488
#>   [4,] -0.04725715 -1.45787965  1.30326155  0.743071425 -0.493604590
#>   [5,] -1.25568997 -0.66830896 -1.31901522  0.569836089  0.261958696
#>   [6,]  0.49869493  0.23249105 -0.11229256  0.297725868  0.903889107
#>   [7,] -1.34039729 -1.46585010  0.65716155  0.343789093  0.782790835
#>   [8,]  0.76138037 -0.08079267 -0.42003774  0.247319805 -1.915905174
#>   [9,] -0.15591679 -1.00981613 -0.16937918 -0.091906617 -1.052916656
#>  [10,] -0.32256423  0.43100254 -0.28889420  0.036800842  0.699632298
#>  [11,]  0.26662899  0.58127990  0.08029803 -0.448561938 -1.071385546
#>  [12,]  1.40287757 -1.21505050 -1.05273341 -0.587328885 -0.414716376
#>  [13,]  0.24031633 -0.90806643  0.14115092 -0.188491074 -0.847735880
#>  [14,]  1.54354695 -0.79115729  1.40262359 -0.511515074 -0.336936555
#>  [15,]  1.41075752  1.96492469 -1.26992415  1.712766604 -0.167461809
#>  [16,] -1.20953979  0.13435488  2.09513270 -0.099435760 -1.203477281
#>  [17,]  0.11950124  0.73488446 -0.81358468 -0.260753617  0.703715955
#>  [18,]  0.55900741 -0.42433759  0.18369735 -0.385158886 -0.993328767
#>  [19,]  0.26555500 -0.79382011  0.44227056 -1.515072435  1.151815434
#>  [20,] -1.34398081  0.48704399 -1.82248397 -0.408517053  2.674992278
#>  [21,]  0.56209725 -0.16534322  0.72475946 -0.430043576  1.370843284
#>  [22,] -1.71595217 -1.83333897  1.16478460  1.136169779 -1.186050643
#>  [23,] -2.19575905  0.06961239 -0.41443566 -0.171166214  1.973164735
#>  [24,]  0.67063290 -0.11027135 -0.06257946  1.092989317 -0.360420806
#>  [25,]  0.40495799 -1.65244352  0.22245954 -0.882456921 -2.265252351
#>  [26,] -0.32639903  0.98200778  0.69248037 -0.679613827 -0.457356644
#>  [27,] -0.27080124 -0.30004631 -1.01780310  0.419532638  0.343570390
#>  [28,]  0.20008868 -0.21400245  0.75733259  0.707797122 -0.451921834
#>  [29,] -0.41686886  0.32604684 -0.51108808  0.118368906 -1.628292024
#>  [30,]  0.38946963 -1.35588526  0.20262793 -0.076189985  0.705801027
#>  [31,] -0.96399105  0.37967590 -0.92327361  0.042990124 -0.279901211
#>  [32,] -0.53092174 -0.54023138 -0.80609115 -0.934179039 -0.865429410
#>  [33,] -0.87188195  1.68340660 -0.05598607  0.692743865 -0.462078958
#>                 [,6]        [,7]        [,8]        [,9]         [,10]
#>   [1,]  0.8161150524 -1.32526232 -0.05925486 -1.80725158 -0.2228576390
#>   [2,]  1.3026275545 -0.28027878  0.97976721  0.89735742  1.0319883251
#>   [3,]  1.2760637264 -1.24165996 -1.04906503  0.27407036 -0.6884241962
#>   [4,] -0.6195036087 -0.96407828 -0.14917210  1.09353376  1.5209294794
#>   [5,]  0.0806752917 -1.56879368  1.74303849  0.26137321 -1.6591075006
#>   [6,]  0.8446438932 -1.73819935  0.93679639  0.83594414  0.6714169621
#>   [7,] -0.2864957430 -0.01206839 -0.28667960  0.58867406 -0.2368361198
#>   [8,]  0.8417883147  2.26487301  1.59873164 -0.83875277 -0.8122083782
#>   [9,]  0.9989344144  0.09280005  0.43683146 -1.02304854  1.3253238379
#>  [10,] -0.3048303416 -2.04407352  1.29843929 -0.65569446 -1.1025781313
#>  [11,] -1.8259185290  1.05099865  0.02206940 -0.24490812 -2.0787901850
#>  [12,] -0.2591174721  0.68930074  0.18132038 -0.21337623  1.7751081724
#>  [13,]  0.1104126375  3.03847396 -0.44115730  0.61610397 -1.1576542871
#>  [14,] -0.1064447818 -0.32610279 -0.39375651 -1.33205021 -0.4700200638
#>  [15,] -0.6470664482  1.63726452  0.98696525  1.88193932  1.9827864244
#>  [16,] -1.5987933630 -0.09194249  0.17413299  0.12573537 -1.8143280820
#>  [17,] -1.4658079045 -1.17000793 -0.52985484  0.88711229 -0.0890597288
#>  [18,]  0.7168410618  1.57231646 -1.10752963  0.66583343 -0.4301966855
#>  [19,]  0.5905470776 -0.12291065  0.60499621  0.95517750  0.6840979700
#>  [20,]  0.6198502674 -1.54733766  1.37941593  0.75331740 -1.2763109836
#>  [21,]  0.6751841647 -0.22186839  1.48459450  0.56693092 -0.5838791874
#>  [22,] -0.3098327559  0.32780860 -1.31152860 -0.60564983 -0.2297702130
#>  [23,] -0.0523118007  0.55015539 -0.43277697 -0.14163153 -0.5854982431
#>  [24,] -1.1000027959 -1.90519424  1.30529695 -1.13238326  0.9177610009
#>  [25,] -0.2732891435  0.77601859 -1.73991153  1.22520284  0.7363941207
#>  [26,] -0.7866280516 -1.15407869 -0.19443077 -0.98643170 -1.9065629266
#>  [27,]  1.3619355464 -0.86914195  2.36278973  0.30035523  0.6964353234
#>  [28,] -1.5914318032 -0.53009403 -1.70237613  0.22153448  2.2556581436
#>  [29,]  0.1437574173  0.01680229  1.97085638 -0.70690595  0.7849626292
#>  [30,]  1.0047638366 -0.23386969 -0.19568383  0.31612525  0.4643771708
#>  [31,]  1.7827927289  1.26288799  1.53197814  0.93345850  0.2357362054
#>  [32,] -0.0004544437 -0.46782192  0.25395661 -0.14693828 -0.7185639536
#>  [33,]  0.2959891870  0.34831657  1.32481041  0.40445178  0.9132169655
#>              [,11]        [,12]        [,13]       [,14]       [,15]
#>   [1,] -1.58836342  1.128464977  0.845313566  0.40629924 -0.02813158
#>   [2,] -0.17777315  0.152386713  0.238458741 -1.72535940 -1.39696050
#>   [3,]  1.36477324  0.161861956 -0.018743711 -0.93617354 -0.71059636
#>   [4,] -0.37582368  0.604538341  0.333636085 -0.38781243  0.01995929
#>   [5,]  1.66113954  0.988035141  0.491464851  0.99830856  1.05477040
#>   [6,]  0.86848974  1.121230430 -2.639508634  1.06154550  0.11385419
#>   [7,] -1.96464296 -0.453096917 -0.907023605 -0.78974907  1.12048779
#>   [8,] -0.09851888  0.793800305  0.877485782 -0.29159342 -0.10200583
#>   [9,] -0.88452812  0.966877592  0.913093112  0.27722339 -0.90868566
#>  [10,]  1.40988195 -0.519993510  0.705718087  0.71049247 -0.08516300
#>  [11,] -1.46194905  1.109375201 -1.266981111 -0.24547788  1.21803954
#>  [12,] -0.24969636 -0.868249663  0.220088978 -0.15641842  0.40534735
#>  [13,]  0.77316206  0.244991744  0.949064607 -0.16165744 -0.84186759
#>  [14,] -0.27473502 -1.460411153 -1.503165952 -1.34973387  0.18444210
#>  [15,]  0.58291282 -1.272102961  1.218690405 -0.79498652 -0.39713504
#>  [16,] -1.70053186 -0.766739590  1.023850897 -0.87353624  0.66181126
#>  [17,]  2.44637825  0.081501079 -2.029670889  1.30762801  0.29706147
#>  [18,] -0.27152031  0.509824954 -0.069291838  1.32543559  0.29310291
#>  [19,] -0.05174752  1.226525335  0.878991650  1.16669578  2.38699085
#>  [20,] -0.75121668 -0.023071525 -1.239997531 -1.65920043 -0.58078356
#>  [21,] -0.74788441  1.405622690  0.943181874  0.75378954 -0.96851297
#>  [22,] -1.39281187  2.215159371 -0.443018222 -1.50726695 -1.61824277
#>  [23,]  2.00434098  0.914391882 -0.602679879  0.50626202 -0.98448264
#>  [24,] -1.29037149 -1.609545112  1.715301983 -0.63663665  1.35271368
#>  [25,] -0.23644396  0.511667377  0.227164521  0.06214976 -1.06571518
#>  [26,]  1.39465600  1.059165048 -1.581522243 -0.30382356  2.56055091
#>  [27,]  1.34492956  3.168106693  1.727775938 -0.02255606 -1.14796607
#>  [28,] -0.57047377 -0.716630917  1.433430798  0.53303168  0.19851744
#>  [29,]  0.27731419  1.192004438  0.407756640 -0.10621309 -1.26864991
#>  [30,] -1.34541226 -1.288178902  0.791427104  0.61072811  0.24224753
#>  [31,] -2.55218589  0.353575690  0.456958026  0.96593565  0.62373133
#>  [32,] -0.65311881  0.163796007 -1.360546139 -1.52446907 -1.30797829
#>  [33,]  0.55789056  0.924534586 -0.023968889  0.08354743 -1.63253314
#>              [,16]        [,17]        [,18]        [,19]        [,20]
#>   [1,]  0.31263862 -1.872622328  0.857483777  0.291638153 -0.956376184
#>   [2,]  1.49268491 -0.324328310  0.553888565 -1.111348371 -0.931393627
#>   [3,]  0.60497069  0.954829436 -0.152763927  0.322478316 -0.845551899
#>   [4,] -1.55981972  0.083874797 -1.083644578  0.699417130  0.095473349
#>   [5,]  0.30611824 -0.230761690  0.639191600 -0.266194422  0.864237583
#>   [6,] -0.46208381  0.745909948  1.751696408  0.649054906  0.329521438
#>   [7,] -0.43507132  0.539998006 -1.237616514  1.553999617  1.333739515
#>   [8,]  1.32520240 -0.470610862  0.372755664  1.828167605 -1.444986715
#>   [9,] -0.80580104 -0.819817499 -1.348073788 -0.596426486  0.869011279
#>  [10,]  0.73715758  1.696014563  2.343878547  0.471250172  1.974871788
#>  [11,]  0.12855384  1.002696858 -0.941592782  0.404014989  1.384554769
#>  [12,]  0.95601336 -0.177871242 -2.311560425 -0.548078110 -0.903982847
#>  [13,]  0.81858211 -1.030634274  1.600342345 -0.527880727 -1.328570871
#>  [14,] -0.80891514  0.234621377  1.934399066  0.869909787  0.294307922
#>  [15,] -0.44013137  1.249333076 -2.003866958  0.151324438 -0.769611970
#>  [16,] -2.09849519  0.176627362  0.624958116 -0.863819656 -0.619329721
#>  [17,] -2.24181569  0.855818450 -0.582174651  0.048889813 -0.103493163
#>  [18,] -0.58978899 -0.980583011 -0.593039467 -1.370843219 -0.516841653
#>  [19,]  0.78921832 -0.115935765  0.727789740 -0.691478733 -0.151933606
#>  [20,] -1.20995061 -0.896483410 -1.496211431  0.378437438 -1.309275443
#>  [21,] -0.43068362 -2.525596191  0.045338521 -0.797398126  0.691020466
#>  [22,] -0.76879469 -1.193367923  0.060539983 -0.566135078 -0.497080694
#>  [23,] -1.02101702  1.422943171  1.978868069 -1.030950042  0.567359978
#>  [24,]  0.09320311  0.056107282  0.810776101 -1.211003520  0.009278435
#>  [25,]  0.65932546  1.482971843  0.959207437  0.985568210 -0.405689660
#>  [26,]  0.77947668  1.024266143 -0.123490622 -0.299272770  0.853071380
#>  [27,] -0.21825388 -0.424017433  2.232039027  1.767301116  1.210663358
#>  [28,] -0.33167566 -0.108196998 -0.917704898 -0.429887627 -1.218545706
#>  [29,] -3.19649457  0.186452049 -0.637239729  0.489294751  0.206291810
#>  [30,]  0.54425255 -0.124421962 -1.132419335  0.320281236 -0.071031389
#>  [31,]  0.86386704  1.044478078 -0.091001301 -0.122603281 -0.689992647
#>  [32,]  1.71731415  0.958192069 -0.234650902 -0.233581261  0.128280115
#>  [33,] -0.05272091  1.934166603  1.465600213  1.617180445  0.100859279
#>              [,21]       [,22]       [,23]        [,24]       [,25]
#>   [1,]  0.09331746  1.67402917 -2.33120387 -0.975785353  1.85267225
#>   [2,]  2.47298063  0.64434017  0.68852936  2.240927475 -0.02354623
#>   [3,]  1.50059086  1.21991147 -0.82984885  0.303146079 -0.17576964
#>   [4,] -0.55467074 -1.08759004  1.48688724 -0.341418801 -1.22014014
#>   [5,]  0.40903029 -0.12633065  0.68181315 -0.550029365 -0.08215729
#>   [6,] -0.94354574  0.36619438  1.51390338 -0.009924312  0.13357513
#>   [7,] -0.39396858  1.39268058  0.17459269  1.071514154  0.07966715
#>   [8,] -0.23671050  0.40888923 -0.50238796 -0.656085110 -0.88574745
#>   [9,] -1.29542099  0.26580364 -1.22480176 -1.460975826 -0.07156310
#>  [10,]  1.64773302  0.38224281 -1.16405369  0.885241707  0.91995070
#>  [11,] -0.79318904  0.08612277 -0.69381127  0.163485573  1.27754998
#>  [12,] -0.43332613  0.61170490 -0.40135277 -1.106790859  0.18090118
#>  [13,]  0.60195144 -0.24890982  1.85195609  0.275620716 -1.46375389
#>  [14,]  1.08327513  0.04220991  2.04931748  1.096529080  1.32963295
#>  [15,] -1.57842105  1.32792420 -1.59871323  0.463774211 -0.06398180
#>  [16,] -1.71406309 -0.73803195  0.52020573  0.479339343  0.27556164
#>  [17,] -1.67437204 -0.86820330 -0.51982766 -0.252451433 -0.48863325
#>  [18,]  0.46071777  0.82623078 -1.86639937  0.418004780  0.45100065
#>  [19,]  0.32910339 -1.09061176 -0.47992408 -0.942944940  0.31350550
#>  [20,]  0.06699068 -0.91156983  0.66292147 -1.076904182  0.30502789
#>  [21,] -1.84290984 -0.69047532  1.34983748 -1.750426776 -0.44139471
#>  [22,]  2.74694733  0.88529412 -1.50357738  0.203374881  0.26629744
#>  [23,] -0.25319392 -0.61590005 -0.91805706 -1.344556695 -1.11381085
#>  [24,] -1.08207956  0.90034876  0.77787583  0.958207433 -0.31044616
#>  [25,]  0.94379819 -0.15002519  0.05116942 -0.274151981 -0.64989573
#>  [26,] -1.01821405 -1.38002290  0.57109257 -0.051022678 -0.51932825
#>  [27,]  0.98766741 -1.07664639  0.53203277 -2.025895802 -1.51368448
#>  [28,] -1.26561910  0.44795462  1.71966596 -0.423892551 -1.08040679
#>  [29,] -1.33096766 -0.06424171 -0.44637247 -0.717457190  0.55182688
#>  [30,] -0.31399937 -0.05538218  2.50642892 -1.224599617  0.31341041
#>  [31,] -2.86067310  0.10964893  0.46462655 -1.413143055 -0.20491756
#>  [32,] -1.48556534 -0.63605072 -1.34897593 -1.945298830 -2.38106917
#>  [33,]  0.49419991  0.04119821  1.04487804  0.487540302  0.62273041
#>               [,26]       [,27]       [,28]       [,29]       [,30]
#>   [1,]  0.283043423 -0.55066164 -0.63603706  1.08580216 -1.85178015
#>   [2,]  2.626035080 -2.17914986  0.47512653  0.97261418 -1.51206425
#>   [3,] -0.564796889 -1.22431717  0.10192162 -0.11178747  1.76401203
#>   [4,]  0.397916668 -0.94658057  1.56126692 -0.55501206  1.09760013
#>   [5,] -0.369988162  0.60224821 -1.13798604  1.61565136  1.07560822
#>   [6,]  1.196224864  1.46558938  0.80305901 -1.46132431  1.31002167
#>   [7,] -2.436699208  0.50064341 -0.45969538 -0.47892725 -1.11642229
#>   [8,]  0.785356363 -1.24896024 -0.22626227 -0.30614301 -0.17913445
#>   [9,] -1.808282778  1.32493451 -2.11465035  1.57482206  1.23016877
#>  [10,]  1.899646280  1.06550089 -1.22730388  0.79031498  0.39992936
#>  [11,] -0.018583622  1.17672084 -0.48507150 -0.64547523 -0.80974033
#>  [12,] -0.367425974  1.43191136 -1.10620101 -0.58373037  1.51779051
#>  [13,] -2.334474684  0.13425189  0.72521291 -0.51886810 -1.06014593
#>  [14,] -1.177540501  1.20344408 -0.68191729 -0.07464020  0.03406570
#>  [15,] -0.516203654  0.15533712  0.21988755 -0.25454989 -0.98911988
#>  [16,] -0.413151035  0.62894115  0.76645981 -0.43264183 -0.24351745
#>  [17,]  0.759820053  0.58626318 -0.84541663 -1.57468317 -1.14616090
#>  [18,]  0.175728058 -0.81101638 -1.72987938 -0.72161077 -0.37685840
#>  [19,] -0.238978170  1.09763233  0.74117494  0.23267037 -1.88465869
#>  [20,] -0.688917097 -0.10306125 -1.72784757 -0.95895265  1.33558804
#>  [21,] -0.064668324  0.21670918 -2.01231657  1.50070594  1.98381344
#>  [22,] -0.945502883 -0.13447020  1.07170137  0.08376431 -0.44197458
#>  [23,]  2.288234085 -0.88326618  1.06509744 -1.05721631 -0.98818587
#>  [24,]  0.384683396  0.41713678  0.12960542  0.26022877  1.07544084
#>  [25,] -0.865027891  2.66836696 -0.31017460 -0.76318929 -0.30054098
#>  [26,] -0.999788159  0.33883486 -1.84675638 -0.97084453 -1.04739206
#>  [27,]  0.101418262 -0.01719711 -0.80517662  1.02193136 -0.48944201
#>  [28,]  2.495593218  0.83058417 -0.14198347 -0.46940024  0.27548921
#>  [29,] -0.701086769 -0.12790089 -0.53444207 -0.99898420 -0.92981833
#>  [30,]  0.802579037 -0.24157536 -0.31575716 -0.08176980  0.11289056
#>  [31,] -0.651958216  0.55573406 -0.55869013  0.44249941  0.39241987
#>  [32,]  0.823895335 -0.51727856 -0.30254126 -0.35291691 -1.79962809
#>  [33,]  0.420126699  0.10609222 -0.43914543  0.51513114 -1.38510194
#>  [ reached getOption("max.print") -- omitted 67 rows ]
#> 
#> $y
#>                [,1]         [,2]          [,3]          [,4]          [,5]
#>   [1,] -0.025323249 -0.205001538  -0.850805387  -0.586680484 -0.2782659052
#>   [2,]  0.105385642  0.349898238   9.506453258   9.805887328  0.2378522152
#>   [3,] -0.011666493 -0.063523999   0.082547933  -0.288225138 -0.0138495645
#>   [4,] -0.038479828 -0.485668389   1.785291682   1.738143605  0.0780248643
#>   [5,] -0.030815816  0.428829577 -10.758072436 -10.147559713  0.0719078601
#>   [6,]  0.283150126  0.453835431  -0.222500658   0.202092651 -0.1896204641
#>   [7,] -0.065814468  0.045150233  -2.294526471  -2.333218286 -0.2136806191
#>   [8,]  0.400339680  0.319231977  -7.858573121  -7.743589838  0.0103504524
#>   [9,]  0.025725955  0.203236898   5.324623671   5.516414103  0.3012747892
#>  [10,]  0.315761162 -0.032288313  -1.460465802  -1.245862620  0.0774182029
#>  [11,] -0.230899496 -0.211771997 -20.487149523 -20.562850558 -0.1411520587
#>  [12,] -0.201699685 -0.178394898   4.617785831   4.747700997  0.0290794633
#>  [13,]  0.027371758 -0.632467103  -0.114974939   0.308758285 -0.2356486975
#>  [14,] -0.108725101  0.412478812  -1.354528836  -1.202126290  0.2037257534
#>  [15,] -0.061147490 -0.300908447  -0.057667634  -0.532789151  0.0647639875
#>  [16,] -0.059429963  0.062642307  -0.136065687   0.202352361  0.1875595659
#>  [17,]  0.228551737 -0.087120675  -2.702819131  -2.432297767  0.0481487959
#>  [18,]  0.050982363 -0.166659065   0.218714524  -0.157341265 -0.1039932193
#>  [19,] -0.100763191  0.343940732  -3.036419923  -3.051653786 -0.0752996495
#>  [20,]  0.012161212 -0.150853216   6.501842987   5.954172567  0.2191974350
#>  [21,]  0.063100859 -0.101206417   0.328729149   0.215309855 -0.1842809910
#>  [22,]  0.070954124 -0.081726120   0.435476580  -0.125703791 -0.1092242487
#>  [23,]  0.290129846 -0.200131282  -0.032128233   0.006167084  0.3820865474
#>  [24,] -0.103095771 -0.333539282  -2.335689768  -2.325523910 -0.2645516018
#>  [25,] -0.209157437 -0.112474035  -1.553179527  -1.705973070 -0.1796356646
#>  [26,]  0.272649552 -0.032466895 -20.381821033 -20.439018502  0.3008794719
#>  [27,]  0.044959220 -0.039138432   0.006758919   0.032075202  0.2087375352
#>  [28,] -0.146150490 -0.509727408   0.035181944   0.154055712 -0.1965376433
#>               [,6]          [,7]          [,8]          [,9]        [,10]
#>   [1,] -0.04073532   0.396305215   0.068290839  -0.071976457  -0.99017249
#>   [2,]  0.58000991   0.110912332  -0.234211140   0.014279427   9.56026566
#>   [3,]  0.09567867  10.213729359  -9.802840252   9.581247611   0.04322659
#>   [4,]  0.15147127  -0.019245722  -0.096428711  -0.094911927   1.64389668
#>   [5,]  0.15364594  -0.202310217   0.302468369   0.129391793 -10.22392939
#>   [6,] -0.18940946 -12.055001464  11.629084103 -11.632792381  -0.19932294
#>   [7,] -0.04989824   0.131716194   0.250349838   0.164121792  -2.13982253
#>   [8,]  0.13206275  -0.076963863   0.081589276   0.110329727  -7.72693351
#>   [9,] -0.24705173   0.067862920   0.090934605   0.080053645   5.41158289
#>  [10,] -0.16806931   0.190194695  -0.289376490  -0.179432789  -1.41119017
#>  [11,] -0.22460123   0.030326437   0.113593277  -0.003143018 -20.25809623
#>  [12,]  0.22748532  -0.138542595   0.290980104  -0.006429887   4.70706003
#>  [13,] -0.07791188  -7.112476005   7.307086958  -7.043077785  -0.01200800
#>  [14,] -0.01598243  -0.101449511   0.081909194   0.359026795  -1.17946026
#>  [15,] -0.54338484   6.005081117  -5.718493767   5.972316136  -0.12279926
#>  [16,]  0.26719993  -7.143113613   7.200967418  -7.334539907   0.16763958
#>  [17,] -0.23634252  -0.376254917  -0.115316732   0.178959728  -2.43686951
#>  [18,]  0.68460539  17.080309336 -17.435340976  17.345254440  -0.02054307
#>  [19,]  0.07176025  -0.224157310   0.008977502  -0.198667918  -3.09124076
#>  [20,] -0.05791030  -0.379496565   0.168304564  -0.112781185   6.66017764
#>  [21,] -0.43933954  -0.044872981  -0.406880019   0.314608734   0.17355364
#>  [22,] -0.01330381   3.549977945  -3.253545377   3.290120932   0.15321100
#>  [23,]  0.40511488   0.942436993  -0.888305102   1.448883249   0.31083825
#>  [24,] -0.04054486   0.091594320  -0.388857519  -0.149831337  -2.13342608
#>  [25,] -0.09285993   0.041323948  -0.022959590   0.195959077  -1.46686253
#>  [26,]  0.34490228   0.101659787  -0.203099540   0.067198046 -20.37478194
#>  [27,] -0.02459880  -0.587658083  -0.141839190  -0.118553191   0.09422854
#>  [28,] -0.07436164  -7.921525654   8.290146756  -8.457218058   0.04370275
#>                [,11]         [,12]        [,13]         [,14]         [,15]
#>   [1,]   0.300839139   0.242136934  0.030921493   0.238345129 -0.2260545203
#>   [2,]  -0.038690354   0.285993985 -0.236310896  -0.080418837 -0.0816923854
#>   [3,]   9.352704154   9.798841886  0.092670371 -10.156106821 -0.0476240115
#>   [4,]   0.393205531   0.347964659  0.230585461  -0.174303858 -0.1249117818
#>   [5,]   0.169603677   0.144556061 -0.270534294   0.008893191  0.0005062193
#>   [6,] -11.526143320 -11.552917891 -0.069441880  11.574582596 -0.3209820976
#>   [7,]   0.202674122  -0.078260812 -0.033904126  -0.231060245  0.1291873759
#>   [8,]   0.227333280   0.021366453  0.163006256  -0.075201528  0.1980649688
#>   [9,]  -0.024272225   0.325167921  0.276559304   0.072731985 -0.1113912170
#>  [10,]   0.424469400   0.204925474  0.084878945  -0.363639854 -0.4247322169
#>  [11,]   0.127665109  -0.122792293 -0.225022309   0.073239224  0.1286435294
#>  [12,]   0.041779667  -0.195220973  0.150950609   0.332836300 -0.0514760760
#>  [13,]  -7.003247235  -6.611046550 -0.078516303   7.120709433  0.2039047842
#>  [14,]  -0.209430370   0.131108040 -0.165389446   0.011151505 -0.2445055412
#>  [15,]   6.350322666   6.291308684  0.404012265  -6.537709557  0.0870742330
#>  [16,]  -7.076587534  -7.262298953  0.198284104   7.154573907  0.5309844477
#>  [17,]   0.187592924  -0.366014561  0.134679063   0.084956162 -0.0021781341
#>  [18,]  17.338161458  17.477940243  0.008773041 -17.609194490 -0.0327927408
#>  [19,]  -0.114240968   0.146066551  0.315599826  -0.088227953  0.0958198604
#>  [20,]   0.075316447  -0.147910677 -0.047575031  -0.092859021 -0.3885630692
#>  [21,]  -0.533391881  -0.439180046  0.009159252   0.350509613  0.1192442871
#>  [22,]   3.484387048   3.179516271  0.027828901  -3.789650669  0.1495088911
#>  [23,]   0.685061769   0.770908830  0.245278544  -0.310159253 -0.1562694263
#>  [24,]  -0.258129503   0.166777422 -0.305152990  -0.319981918 -0.0703675120
#>  [25,]  -0.022109187   0.204062962 -0.186819359  -0.181104087 -0.0147848695
#>  [26,]   0.059157124  -0.129432420  0.203697588   0.556700151  0.0050107727
#>  [27,]  -0.341402023   0.274598491 -0.190749021   0.322803927 -0.3316651511
#>  [28,]  -8.438833400  -8.224605444  0.027120573   8.774724272 -0.0254617674
#>               [,16]        [,17]        [,18]        [,19]         [,20]
#>   [1,]  0.057732632 -0.089974242 -0.529717214  -0.67497536 -2.843887e-02
#>   [2,]  0.130722275  0.033284119 -0.112252635   9.95718610  3.446455e-02
#>   [3,]  0.037888908  0.022744495  0.321907827  10.40519528 -6.189180e-01
#>   [4,] -0.139447368  0.322036965 -0.240747272   1.85563052 -1.479880e-01
#>   [5,]  0.048123554  0.197646606 -0.425479548 -10.35266005  1.572988e-01
#>   [6,] -0.144724466 -0.231776737  0.200479801 -11.63044567 -4.175070e-01
#>   [7,]  0.124552570  0.146367361 -0.066470100  -2.18926244 -1.422646e-01
#>   [8,] -0.226158759  0.379728324  0.087563902  -8.06976275  1.258313e-01
#>   [9,] -0.120400078  0.082540867 -0.335457938   5.41478687 -2.426557e-01
#>  [10,]  0.086218423 -0.030910756  0.095486284  -1.37255669 -3.145065e-01
#>  [11,]  0.160429400  0.060749062  0.243632788 -19.93594296  1.049021e-01
#>  [12,]  0.392298641  0.096738008  0.177521828   4.25544099  1.253953e-01
#>  [13,] -0.429116571  0.055384051 -0.224877196  -6.74090422 -1.995889e-01
#>  [14,]  0.001481908  0.580139043  0.268152444  -1.23381501  1.237822e-01
#>  [15,]  0.352549647 -0.020206008 -0.002475077   5.80965787 -4.714244e-01
#>  [16,] -0.328057537  0.141733608 -0.196484826  -7.28813166 -6.664424e-05
#>  [17,]  0.171932680 -0.207641459 -0.029910929  -2.49218851  1.024999e-02
#>  [18,] -0.174263509 -0.368205169  0.202785865  17.14711178  2.787492e-01
#>  [19,] -0.234058794 -0.325523066 -0.126050704  -2.93839813 -1.008803e-01
#>  [20,]  0.069463960 -0.186971134  0.037491936   6.36092090  1.240503e-01
#>  [21,] -0.014150080 -0.291421167 -0.050915265  -0.38739497  2.244409e-01
#>  [22,]  0.108525122 -0.478844456  0.196097406   3.50133016  1.008473e-01
#>  [23,] -0.234929918  0.282345811  0.301370370   0.69852622  2.472112e-01
#>  [24,] -0.012946701  0.001101686 -0.063895033  -2.15535695 -6.321282e-02
#>  [25,]  0.261366994  0.135296182  0.202496193  -1.64480780  1.007623e-01
#>  [26,]  0.256083783 -0.099024423  0.371009949 -20.20625334  1.625126e-01
#>  [27,] -0.313487957  0.215097046 -0.221959007  -0.30089728 -3.142826e-01
#>  [28,] -0.773882083  0.362276127 -0.169907615  -8.31919426  7.345486e-02
#>                [,21]         [,22]         [,23]        [,24]         [,25]
#>   [1,]   0.599677761  0.0337721581   0.968229379 -0.189387411   0.823916175
#>   [2,]  -9.636350796 -0.0471019827  -9.271206014  0.290841303  -9.798805206
#>   [3,]   0.166448219  0.2871388698  -0.159801534 -0.235578391   0.121604085
#>   [4,]  -1.595987621  0.1907390093  -1.575410446 -0.425479777  -1.750314324
#>   [5,]   9.874552975 -0.1627591162  10.446192452 -0.013837352  10.144957467
#>   [6,]  -0.036573037 -0.0587991550  -0.595137321  0.066809900  -0.588504585
#>   [7,]   2.354590772 -0.2082910399   2.577978696  0.050170564   2.399979439
#>   [8,]   7.722150267 -0.0101228344   7.396785676 -0.035807796   7.239738313
#>   [9,]  -5.744533477 -0.0897526322  -5.701492327  0.090533132  -4.982350435
#>  [10,]   1.419376984  0.1570243622   1.379119198  0.055191805   1.148779417
#>  [11,]  20.543958719  0.1871270456  20.370633475 -0.178040973  20.458114119
#>  [12,]  -4.330310257  0.1572937401  -4.768167603  0.174140495  -4.489119311
#>  [13,]  -0.200903193 -0.2034517597  -0.170232429  0.695064819   0.253076292
#>  [14,]   1.361397085 -0.2627784241   0.790804213  0.261191582   1.132310548
#>  [15,]  -0.258320510 -0.4831954520   0.281453877  0.200275115   0.211966334
#>  [16,]   0.231603408  0.3468141363  -0.237069793  0.240546467   0.134043641
#>  [17,]   2.741775079  0.0492315800   2.551824992  0.216474299   2.574534378
#>  [18,]   0.127013667  0.0695058729   0.177316490 -0.190208331   0.018144717
#>  [19,]   3.126954981  0.4170918565   3.058958067 -0.182370156   3.043052134
#>  [20,]  -6.699772980 -0.3505395849  -6.836152588  0.244033100  -6.263784980
#>  [21,]   0.091929101 -0.0047059743  -0.114493190  0.308024765  -0.195755916
#>  [22,]   0.299653967 -0.1514524345   0.463155469  0.207177941   0.025493304
#>  [23,]   0.100341575 -0.0091299885  -0.096534997 -0.675580542   0.345547484
#>  [24,]   2.057240947  0.5608598426   2.089538015 -0.039844730   2.177504793
#>  [25,]   1.404378519  0.0983334330   1.697760125 -0.185530452   1.799020117
#>  [26,]  20.361937963 -0.1018398899  20.164852761 -0.066115534  20.066940558
#>  [27,]  -0.223031459  0.4332454360  -0.027064202  0.353274838  -0.137218658
#>  [28,]  -0.254821659  0.1848112918   0.026190128 -0.180077565   0.208392867
#>               [,26]        [,27]        [,28]         [,29]         [,30]
#>   [1,]   0.96083126 -0.160055888   0.46228219   0.705923799  1.043177e-01
#>   [2,]  -9.25346373  0.077160334  -9.28986910  -9.889032719  1.523209e-01
#>   [3,]  -0.04320818  0.058922228  -0.08702655  -0.131463885  9.624668e+00
#>   [4,]  -1.81996520 -0.556775698  -1.51874915  -1.406581134  2.430245e-01
#>   [5,]   9.90855315  0.038829451   9.89709011  10.632251962  1.548252e-02
#>   [6,]  -0.03750152  0.123435054   0.54161180   0.017921947 -1.207179e+01
#>   [7,]   2.44515308 -0.337934027   2.10588074   2.309036481  1.931454e-01
#>   [8,]   7.65448285 -0.315667236   7.34531627   7.529247513 -9.386446e-02
#>   [9,]  -5.39946295 -0.038484588  -5.68979517  -5.127614666 -8.329359e-02
#>  [10,]   1.25927891  0.118506176   1.44600194   1.679561731  3.451156e-02
#>  [11,]  20.51160140 -0.065416909  20.16380460  20.147607806  1.157681e-02
#>  [12,]  -4.42964288  0.219997602  -4.71174931  -4.627902190  1.246318e-01
#>  [13,]   0.07045868 -0.069197565  -0.24640477  -0.037622557 -6.963895e+00
#>  [14,]   1.02716947 -0.050652189   1.00898660   1.143150316 -2.288810e-01
#>  [15,]  -0.20603486  0.095536237   0.20159821   0.233408353  6.066058e+00
#>  [16,]   0.19093904  0.204614214   0.13794843   0.135680046 -7.086459e+00
#>  [17,]   2.19088443  0.225172957   2.11867927   2.490922361 -5.452593e-04
#>  [18,]   0.02372241  0.433359197  -0.30824403  -0.070735042  1.719329e+01
#>  [19,]   3.04463143 -0.175226297   3.14733050   3.058782810 -1.687669e-02
#>  [20,]  -6.60753902 -0.228135153  -6.40490447  -6.677748743  5.302722e-02
#>  [21,]  -0.73481668 -0.289527173   0.30290243   0.295536365 -2.342093e-01
#>  [22,]   0.14644506  0.719208600   0.10759815   0.068999597  3.691558e+00
#>  [23,]  -0.04256783  0.346413506   0.30112183   0.493004889  1.238733e+00
#>  [24,]   2.27436042  0.401284762   1.99837511   2.134699217  1.312737e-02
#>  [25,]   1.66684065  0.014719037   1.54305540   2.136588364 -1.228404e-01
#>  [26,]  20.21012655 -0.009962774  20.24040756  20.137816926  6.158084e-02
#>  [27,]   0.19453841 -0.147518359   0.49546460  -0.130530322  1.259248e-01
#>  [28,]  -0.21981620  0.463733764   0.05485318   0.151135767 -8.327631e+00
#>                [,31]         [,32]        [,33]        [,34]         [,35]
#>   [1,]  -0.361254838 -0.0765381943 -0.015952640  0.001687558 -3.001870e-02
#>   [2,]   0.079126770 -0.4123791614  0.037565116 -0.327671763  8.177528e-02
#>   [3,]  10.575352403 -0.0089473873  0.350271677  0.331037875 -9.598137e+00
#>   [4,]  -0.203355266 -0.1500563748  0.009072717  0.503823275  2.989038e-01
#>   [5,]   0.248336610 -0.1393546395  0.031883977 -0.126009375  3.530174e-02
#>   [6,] -11.738937648 -0.4921077284  0.056309207  0.241717186  1.174364e+01
#>   [7,]  -0.001690481  0.0714480886 -0.228596908  0.099448135  1.579346e-01
#>   [8,]   0.011130800 -0.3086398029 -0.273964583 -0.522120618  2.183940e-01
#>   [9,]   0.048058499 -0.4016432117  0.068543487 -0.063240198  2.307574e-02
#>  [10,]   0.439134627 -0.0398539556 -0.203861567 -0.263682470  2.020930e-02
#>  [11,]   0.009037789 -0.3062757421 -0.261966849 -0.182299834 -1.165615e-01
#>  [12,]   0.301808669  0.2613873932  0.081302751  0.078948254 -1.258691e-04
#>  [13,]  -7.290791386 -0.1656891833  0.010117395  0.372386725  7.070590e+00
#>  [14,]   0.031885961 -0.3104693984 -0.191726893  0.058856737  3.990899e-02
#>  [15,]   6.198231726 -0.2721494018 -0.144380688  0.465353403 -6.369444e+00
#>  [16,]  -6.854150417  0.2180921702  0.069903828 -0.021768689  6.805415e+00
#>  [17,]  -0.103855737  0.1775125350 -0.111733933 -0.116096751  1.597214e-01
#>  [18,]  16.649296026 -0.1046028002 -0.266879112  0.344970964 -1.789685e+01
#>  [19,]   0.111820394  0.0005019003 -0.051736311 -0.061014861  2.604665e-01
#>  [20,]   0.261431725 -0.2038859129 -0.019025703  0.169828118  2.242117e-01
#>  [21,]  -0.277536591  0.1085982455  0.286302135 -0.041595513 -3.047225e-01
#>  [22,]   4.004485676  0.1211523335 -0.257625336 -0.259664021 -3.819694e+00
#>  [23,]   1.218668889 -0.1121518188  0.108905929  0.320268917 -9.956949e-01
#>  [24,]  -0.168062716 -0.0505883074 -0.216516519 -0.015659183  3.566919e-01
#>  [25,]  -0.117958729  0.0910350328 -0.391980689 -0.274477434  2.483352e-01
#>  [26,]   0.147546388  0.0472776492  0.138759765 -0.366383626  1.304360e-01
#>  [27,]   0.021909008  0.0163672930 -0.009667434 -0.035492872  1.152432e-01
#>  [28,]  -8.313887229  0.2531779611  0.123199662  0.275906583  8.038739e+00
#>  [ reached getOption("max.print") -- omitted 72 rows ]
#> 
#> $a
#> $a[[1]]
#>       [,1] [,2]      [,3]      [,4] [,5] [,6] [,7] [,8] [,9]     [,10] [,11]
#>  [1,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#>  [2,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#>  [3,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#>  [4,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#>  [5,]    0    0  2.899222  2.899222    0    0    0    0    0  2.899222     0
#>  [6,]    0    0  2.092087  2.092087    0    0    0    0    0  2.092087     0
#>  [7,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#>  [8,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#>  [9,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [10,]    0    0  3.583310  3.583310    0    0    0    0    0  3.583310     0
#> [11,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [12,]    0    0 -1.611913 -1.611913    0    0    0    0    0 -1.611913     0
#> [13,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [14,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [15,]    0    0 -3.432100 -3.432100    0    0    0    0    0 -3.432100     0
#> [16,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [17,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [18,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [19,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [20,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [21,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [22,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [23,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [24,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [25,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [26,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [27,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#> [28,]    0    0  0.000000  0.000000    0    0    0    0    0  0.000000     0
#>       [,12] [,13] [,14] [,15] [,16] [,17] [,18]     [,19] [,20]     [,21] [,22]
#>  [1,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#>  [2,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#>  [3,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#>  [4,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#>  [5,]     0     0     0     0     0     0     0  2.899222     0 -2.899222     0
#>  [6,]     0     0     0     0     0     0     0  2.092087     0 -2.092087     0
#>  [7,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#>  [8,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#>  [9,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [10,]     0     0     0     0     0     0     0  3.583310     0 -3.583310     0
#> [11,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [12,]     0     0     0     0     0     0     0 -1.611913     0  1.611913     0
#> [13,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [14,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [15,]     0     0     0     0     0     0     0 -3.432100     0  3.432100     0
#> [16,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [17,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [18,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [19,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [20,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [21,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [22,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [23,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [24,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [25,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [26,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [27,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#> [28,]     0     0     0     0     0     0     0  0.000000     0  0.000000     0
#>           [,23] [,24]     [,25]     [,26] [,27]     [,28]     [,29] [,30] [,31]
#>  [1,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#>  [2,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#>  [3,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#>  [4,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#>  [5,] -2.899222     0 -2.899222 -2.899222     0 -2.899222 -2.899222     0     0
#>  [6,] -2.092087     0 -2.092087 -2.092087     0 -2.092087 -2.092087     0     0
#>  [7,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#>  [8,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#>  [9,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [10,] -3.583310     0 -3.583310 -3.583310     0 -3.583310 -3.583310     0     0
#> [11,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [12,]  1.611913     0  1.611913  1.611913     0  1.611913  1.611913     0     0
#> [13,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [14,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [15,]  3.432100     0  3.432100  3.432100     0  3.432100  3.432100     0     0
#> [16,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [17,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [18,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [19,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [20,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [21,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [22,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [23,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [24,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [25,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [26,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [27,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#> [28,]  0.000000     0  0.000000  0.000000     0  0.000000  0.000000     0     0
#>       [,32] [,33] [,34] [,35]
#>  [1,]     0     0     0     0
#>  [2,]     0     0     0     0
#>  [3,]     0     0     0     0
#>  [4,]     0     0     0     0
#>  [5,]     0     0     0     0
#>  [6,]     0     0     0     0
#>  [7,]     0     0     0     0
#>  [8,]     0     0     0     0
#>  [9,]     0     0     0     0
#> [10,]     0     0     0     0
#> [11,]     0     0     0     0
#> [12,]     0     0     0     0
#> [13,]     0     0     0     0
#> [14,]     0     0     0     0
#> [15,]     0     0     0     0
#> [16,]     0     0     0     0
#> [17,]     0     0     0     0
#> [18,]     0     0     0     0
#> [19,]     0     0     0     0
#> [20,]     0     0     0     0
#> [21,]     0     0     0     0
#> [22,]     0     0     0     0
#> [23,]     0     0     0     0
#> [24,]     0     0     0     0
#> [25,]     0     0     0     0
#> [26,]     0     0     0     0
#> [27,]     0     0     0     0
#> [28,]     0     0     0     0
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> $a[[2]]
#>       [,1] [,2] [,3] [,4] [,5] [,6]      [,7]      [,8]      [,9] [,10]
#>  [1,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#>  [2,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#>  [3,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#>  [4,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#>  [5,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#>  [6,]    0    0    0    0    0    0  1.577557 -1.577557  1.577557     0
#>  [7,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#>  [8,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#>  [9,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [10,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [11,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [12,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [13,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [14,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [15,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [16,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [17,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [18,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [19,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [20,]    0    0    0    0    0    0 -2.307096  2.307096 -2.307096     0
#> [21,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [22,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [23,]    0    0    0    0    0    0 -4.031503  4.031503 -4.031503     0
#> [24,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [25,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [26,]    0    0    0    0    0    0  0.000000  0.000000  0.000000     0
#> [27,]    0    0    0    0    0    0 -2.415380  2.415380 -2.415380     0
#> [28,]    0    0    0    0    0    0 -3.178293  3.178293 -3.178293     0
#>           [,11]     [,12] [,13]     [,14] [,15] [,16] [,17] [,18]     [,19]
#>  [1,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#>  [2,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#>  [3,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#>  [4,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#>  [5,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#>  [6,]  1.577557  1.577557     0 -1.577557     0     0     0     0  1.577557
#>  [7,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#>  [8,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#>  [9,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [10,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [11,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [12,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [13,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [14,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [15,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [16,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [17,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [18,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [19,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [20,] -2.307096 -2.307096     0  2.307096     0     0     0     0 -2.307096
#> [21,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [22,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [23,] -4.031503 -4.031503     0  4.031503     0     0     0     0 -4.031503
#> [24,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [25,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [26,]  0.000000  0.000000     0  0.000000     0     0     0     0  0.000000
#> [27,] -2.415380 -2.415380     0  2.415380     0     0     0     0 -2.415380
#> [28,] -3.178293 -3.178293     0  3.178293     0     0     0     0 -3.178293
#>       [,20] [,21] [,22] [,23] [,24] [,25] [,26] [,27] [,28] [,29]     [,30]
#>  [1,]     0     0     0     0     0     0     0     0     0     0  0.000000
#>  [2,]     0     0     0     0     0     0     0     0     0     0  0.000000
#>  [3,]     0     0     0     0     0     0     0     0     0     0  0.000000
#>  [4,]     0     0     0     0     0     0     0     0     0     0  0.000000
#>  [5,]     0     0     0     0     0     0     0     0     0     0  0.000000
#>  [6,]     0     0     0     0     0     0     0     0     0     0  1.577557
#>  [7,]     0     0     0     0     0     0     0     0     0     0  0.000000
#>  [8,]     0     0     0     0     0     0     0     0     0     0  0.000000
#>  [9,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [10,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [11,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [12,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [13,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [14,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [15,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [16,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [17,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [18,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [19,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [20,]     0     0     0     0     0     0     0     0     0     0 -2.307096
#> [21,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [22,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [23,]     0     0     0     0     0     0     0     0     0     0 -4.031503
#> [24,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [25,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [26,]     0     0     0     0     0     0     0     0     0     0  0.000000
#> [27,]     0     0     0     0     0     0     0     0     0     0 -2.415380
#> [28,]     0     0     0     0     0     0     0     0     0     0 -3.178293
#>           [,31] [,32] [,33] [,34]     [,35]
#>  [1,]  0.000000     0     0     0  0.000000
#>  [2,]  0.000000     0     0     0  0.000000
#>  [3,]  0.000000     0     0     0  0.000000
#>  [4,]  0.000000     0     0     0  0.000000
#>  [5,]  0.000000     0     0     0  0.000000
#>  [6,]  1.577557     0     0     0 -1.577557
#>  [7,]  0.000000     0     0     0  0.000000
#>  [8,]  0.000000     0     0     0  0.000000
#>  [9,]  0.000000     0     0     0  0.000000
#> [10,]  0.000000     0     0     0  0.000000
#> [11,]  0.000000     0     0     0  0.000000
#> [12,]  0.000000     0     0     0  0.000000
#> [13,]  0.000000     0     0     0  0.000000
#> [14,]  0.000000     0     0     0  0.000000
#> [15,]  0.000000     0     0     0  0.000000
#> [16,]  0.000000     0     0     0  0.000000
#> [17,]  0.000000     0     0     0  0.000000
#> [18,]  0.000000     0     0     0  0.000000
#> [19,]  0.000000     0     0     0  0.000000
#> [20,] -2.307096     0     0     0  2.307096
#> [21,]  0.000000     0     0     0  0.000000
#> [22,]  0.000000     0     0     0  0.000000
#> [23,] -4.031503     0     0     0  4.031503
#> [24,]  0.000000     0     0     0  0.000000
#> [25,]  0.000000     0     0     0  0.000000
#> [26,]  0.000000     0     0     0  0.000000
#> [27,] -2.415380     0     0     0  2.415380
#> [28,] -3.178293     0     0     0  3.178293
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> 
#> $true
#>   [1] 1 1 2 1 1 2 1 1 1 1 1 1 2 1 2 2 1 2 1 1 2 2 2 1 1 1 2 2 1 1 2 1 2 1 1 1 2
#>  [38] 1 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 2 2 2 1 1 2 1 2 2 1 1 2 2 1 2 2 1 1 2 2
#>  [75] 2 1 2 1 2 1 1 2 2 2 1 1 2 1 2 2 2 1 2 2 1 2 1 1 1 1
#>